% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_sub_dt_agg_gen.r
\name{agg_gen}
\alias{agg_gen}
\title{Organise sequencial processing of phenomena aggregations.}
\usage{
agg_gen(...)
}
\arguments{
\item{agg_offset}{A vector of two strings that can be coerced to a
\code{lubridate} period. These can be used to offset the date-time from
which aggregations are totalled. For example, for rainfall totals estimated
from 8 am to 8pm the \code{agg_offset} should be set to c(8 hours, 8 hours).}
}
\description{
An alias for list in pipe processing.
}
\author{
Paul J. Gordijn
}
