% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_sub_dt_agg_param_eval.r
\name{agg_param_eval}
\alias{agg_param_eval}
\title{Ascribe aggregation functions using raw data phenomena tables.}
\usage{
agg_param_eval(phens_summary = NULL, ignore_nas = FALSE, ...)
}
\arguments{
\item{phens_summary}{A summary of the phens table of the particular
phenomena that will be aggregated by \code{ipayipi::dt_agg()}.}

\item{ignore_nas}{If \code{NA} values are ignores any periods or time
interval of the data with any NA values will result in an \code{NA}
aggregation value. Important if missing values have not been
interpolated.}
}
\description{
Ascribe aggregation functions using raw data phenomena tables.
}
\author{
Paul J. Gordijn
}
