% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_sub_dt_agg_params.r
\name{agg_params}
\alias{agg_params}
\title{Aggregate organisation for phenomena}
\usage{
agg_params(aggs = NULL)
}
\description{
Function to organise a list or a data.table of
aggregation parameters.
}
