% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_sub_agg_list.r
\name{aggs}
\alias{aggs}
\title{Organise sequencial processing of phenomena aggregations.}
\usage{
aggs(agg_offset = c("0 secs", "0 secs"), ...)
}
\arguments{
\item{agg_offset}{A vector of two strings that can be coerced to a
\code{lubridate} period. These can be used to offset the date-time from
which aggregations are totalled. For example, for rainfall totals estimated
from 8 am to 8pm the \code{agg_offset} should be set to c(8 hours, 8 hours).}
}
\description{
An alias for list in pipe processing.
}
\author{
Paul J. Gordijn
}
