% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_sub_append_phen_overlap_data.r
\name{append_phen_overlap_data}
\alias{append_phen_overlap_data}
\title{Appends data tables by phenomena}
\usage{
append_phen_overlap_data(
  nd_min = NULL,
  nd_max = NULL,
  sfo = NULL,
  ndt = NULL,
  phen_dt = NULL,
  phens = NULL,
  sf_phen_ds = NULL,
  new_phen_ds = NULL,
  ri = NULL,
  tn = tn,
  overwrite_sf = FALSE,
  ...
)
}
\arguments{
\item{ndt}{Data from the new data file.}

\item{phen_dt}{Phenomena table of from both the station and new data files
combined. These are combined and phids standarised in
ipayipi::station_append().}

\item{sf_phen_ds}{Phenomena summary from the station file.}

\item{new_phen_ds}{Phenomena summary for the new data.}

\item{ri}{The record interval associated with the data sets. As a
standardised string.}

\item{tn}{The name of the phenomena data tables.}
}
\value{
List of useful items for internal use. Overap data.
}
\description{
This function is for internal use. The function is designed to
append phenomena data and associated phenomena standards efficiently, but
retaining metadata records for each phenomena. Moreover, the function
evaluates missing data and compares new and older data records so avoid
loosing data. This part of the function only deals with overlap data.
}
\details{
This function is an internal function called by others in the
pipeline.
}
\author{
Paul J Gordijn
}
\keyword{append}
\keyword{data,}
\keyword{join}
\keyword{overwrite}
\keyword{phenomena}
\keyword{tables,}
