% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_sub_append_phen_overlap_data2.r
\name{append_phen_overlap_data2}
\alias{append_phen_overlap_data2}
\title{Appends data tables by phenomena}
\usage{
append_phen_overlap_data2(
  sfo = NULL,
  ndt = NULL,
  phen_dt = NULL,
  phens = NULL,
  phen_id = FALSE,
  sf_phen_ds = NULL,
  new_phen_ds = NULL,
  ri = NULL,
  tn = NULL,
  overwrite_sf = FALSE,
  cores = getOption("mc.cores", 2L),
  ...
)
}
\arguments{
\item{sfo}{Overlapping phenomena data table from the 'x' (left) table or station file.}

\item{ndt}{Data from the new data file (or 'y' or right table).}

\item{phen_dt}{Phenomena table of both the station and new data files combined. These are combined and phids standarised in \code{ipayipi::station_append()}.}

\item{phens}{Ordered names of phenomena to include in the joined table.}

\item{phen_id}{Set this to TRUE for generating updated station file phenomena data summaries. When set to FALSE processing to produce this data summary will not be executed.}

\item{sf_phen_ds}{Phenomena summary from the station file.}

\item{new_phen_ds}{Phenomena summary for the new data.}

\item{ri}{The record interval associated with the data sets. As a standardised string. Used to generate a time-series sequence using base::seq().}

\item{tn}{The name of the phenomena data tables. Argument only required for processing phenomena summary data.}

\item{overwrite_sf}{Logical. If TRUE the station file or 'x' table values, if not NA, will be overwritten by values from the 'y' or new data table.}

\item{cores}{Integer representing the number of cores available for processing data.}
}
\value{
List of useful items for internal use. Overap data.
}
\description{
Custom full join of overlapping time series data. The function is designed to append phenomena data and associated phenomena standards efficiently, and can retain temporal metadata records for each phenomena if a phenomena summary data table is provided. Missing data is evaluated to avoid loosing data.
}
\details{
This function is an internal function called by \code{ipayipi::append_phen_data()}.
}
\author{
Paul J Gordijn
}
\keyword{append}
\keyword{data,}
\keyword{join}
\keyword{overwrite}
\keyword{phenomena}
\keyword{tables,}
