% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_e_append_station.R
\name{append_station}
\alias{append_station}
\title{Append station data using chunck reads}
\usage{
append_station(
  pipe_house = NULL,
  station_file = NULL,
  new_data = NULL,
  overwrite_sf = TRUE,
  phen_id = TRUE,
  verbose = FALSE,
  xtra_v = FALSE,
  chunk_v = FALSE,
  sfc = NULL,
  ...
)
}
\arguments{
\item{pipe_house}{List of pipeline directories. \strong{See} \code{\link[=ipip_house]{ipip_house()}} \strong{for details}.}

\item{station_file}{The name main data set to which data will be appended.}

\item{new_data}{The data that will be appended to the main 'hobo station'.}

\item{overwrite_sf}{If \code{TRUE} then original data (station data) is disgarded in favour of new data. If TRUE both data sets will be evaluated
and where there are NA values, a replacement, if available, will be used to replace the NA value. Defaults to FALSE.}

\item{phen_id}{Logical. If TRUE then the station 'phens' table will be used to update the 'phens_data_summary'. If FALSE this will be ignored.}

\item{verbose}{Logical. Print some details and progress of function progress?}

\item{xtra_v}{Logical. Should some 'x'tra messaging be done? Use to help diagnose problems, and for guidance.}

\item{chunk_v}{Set to TRUE to view messaging regarding the progress of data chunking.}

\item{sfc}{List of file paths to the temporary station-file directory. Generated using \code{\link[=sf_open_con]{sf_open_con()}}.}
}
\description{
Appends two standardised 'ipayipi' station files.
}
\author{
Paul J. Gordijn
}
\keyword{append}
\keyword{data}
\keyword{data;}
\keyword{logger}
\keyword{pipeline;}
