% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/e_append_station_batch.R
\name{append_station_batch}
\alias{append_station_batch}
\title{Append logger data files to 'station file'}
\usage{
append_station_batch(
  pipe_house = NULL,
  overwrite_sf = FALSE,
  wanted = NULL,
  unwanted = NULL,
  phen_id = TRUE,
  prompt = FALSE,
  verbose = FALSE,
  xtra_v = FALSE,
  chunk_v = FALSE,
  ...
)
}
\arguments{
\item{pipe_house}{List of pipeline directories. \strong{See} \code{\link[=ipip_house]{ipip_house()}} \strong{for details}.}

\item{overwrite_sf}{If \code{TRUE} then original data (station data) is disgarded in favour of new data. If TRUE both data sets will be evaluated
and where there are NA values, a replacement, if available, will be used to replace the NA value. Defaults to FALSE.}

\item{wanted}{Regex string of files to select for listing. Seperate search tags by using the bar character '|'.}

\item{unwanted}{Regex string of files to filter out the listing. Seperate search tags by using the bar character '|'.
in the import.}

\item{phen_id}{Logical. If TRUE then the station 'phens' table will be used to update the 'phens_data_summary'. If FALSE this will be ignored.}

\item{prompt}{Set to TRUE for interactive mode. Note this will not work if embedded in a parallel processing instance.}

\item{verbose}{Logical. Print some details and progress of function progress?}

\item{xtra_v}{Logical. Should some 'x'tra messaging be done? Use to help diagnose problems, and for guidance.}

\item{chunk_v}{Set to TRUE to view messaging regarding the progress of data chunking.}
}
\description{
Appends phenomena data and metadata into continuous or discontinuous data streams in the form of station files.
}
\details{
This function searches for all standardised files in the \code{nomvet_room} and existing 'station files' in the \code{ipip_room}. Files from the \code{nomvet_room} that have not been appended to station files will be appended to appropriate stations. Appending is done via the \code{ipayipi::append_station()}, \code{ipayipi::append_phen_data()}, and \code{ipayipi::append_tables()} functions.

If the data is continuous the function ensures continuous date-time values between missing data chuncks. Missing values are fillled as NA. Event-based or discontinuous data are not treated as continuous data streams.

New phenomena are seemlessly included in the appending process. Moreover, each phenomena is evaluated in turn allowing for a choise of overwring old data with new data or \emph{vice versa}.
}
\author{
Paul J. Gordijn
}
\keyword{append}
\keyword{data}
\keyword{data;}
\keyword{logger}
\keyword{pipeline;}
\keyword{time-series}
