% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_sub_append_tables.r
\name{append_tables}
\alias{append_tables}
\title{Appends a list of tables}
\usage{
append_tables(original_tbl = NULL, new_tbl = NULL, overwrite_old = FALSE, ...)
}
\arguments{
\item{overwrite_old}{NOTE: THIS FUNCTIONALITY HAS NOT YET BEEN INCLUDED.
Logical. If \code{TRUE} then original data is disgarded
in favour of new data. If there are NA values, in either old or new data,
these are not retained if there are values in the old or new data that can
replace these. Defaults to FALSE.}

\item{original_tbl_list}{List of original tables to append. Both original
and new table lists must be named.}

\item{new_tbl_list}{Named list of new tables to append to the originals.}
}
\value{
List of joined tables.
}
\description{
Appends tables based on date-time stamps, and unique rows.
}
\details{
This function is an internal function called by others in the
pipeline.
}
\author{
Paul J Gordijn
}
\keyword{append}
\keyword{data,}
\keyword{join}
\keyword{overwrite}
\keyword{tables,}
