% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_dt_calc_param_eval.R
\name{calc_param_eval}
\alias{calc_param_eval}
\title{Organise sequencial processing of phenomena aggregations.}
\usage{
calc_param_eval(
  f_params = NULL,
  full_eval = FALSE,
  pipe_house = NULL,
  ppsij = NULL,
  sfc = NULL,
  ...
)
}
\arguments{
\item{f_params}{If \code{full_eval} is FALSE: a list of \code{ipayipi::chainer()} operations.}
}
\description{
An alias for list in ipayipi::dt_calc. Used for sequence of
\code{ipayipi::chainer()} calculations.
\itemize{
\item := indicates new phenomena for which units, measure and var_types are necessary for further processing.
}
}
\details{
This function analyses the text of provided data.table chain syntax. In order to populate phenomena (variable) information new (and old) phenomena must be described in line with iPayipi variable standards. This includes describing the variable type, measure, and units of a variable. This function will auto populate the 'measure', 'variable type', and 'units' for flags and checks (and phenomena begining keyed by regex "^flag|^check"). All date-time phenomena keyed by regex 'dttm|date|time' will be automatically populated as well.

NB! New phenomena are extracted from data.table syntax using the ':=' operator as a que. Using the operator to describe more than one variable at a time, in a single chain link is not supported at this time.
}
\author{
Paul J. Gordijn
}
