% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_sub_dt_calc_chainer.r
\name{chainer}
\alias{chainer}
\title{Organise sequencial processing of phenomena aggregations.}
\usage{
chainer(
  dt_syn_bc = NULL,
  dt_syn_ac = NULL,
  dt_syn_exc = NULL,
  temp_var = TRUE,
  measure = NULL,
  units = NULL,
  var_type = NULL,
  notes = NULL,
  fork_table = NULL,
  ...
)
}
\arguments{
\item{dt_syn_bc}{String to be placed before the first comma in data.table
syntax.}

\item{dt_syn_ac}{String to be placed after the first comma in data.table
syntax.}

\item{dt_syn_exc}{String to be placed after the third comma in data.table
syntax.}

\item{temp_var}{designates the calculated variable as temporary. Temporary
variables don't need to have their measure, units, and variable type
described. \code{temp_vars} will be removed at the end of a chain sequence.}

\item{measure}{One of the following to describe the type of measure:
\itemize{
\item "smp" for a sample (instantaneous),
\item "tot" for total,
\item "avg" for mean,
\item "min" for a minimum,
\item "max" for a maximum,
\item "sd" for standard deviation, and
\item "cumm" for a cummulative measure.
}}

\item{units}{Standard shorthand for the units of the phenomena being
calculated.}

\item{var_type}{The type of variable, a string description using one of the
following accepted options:
\itemize{
\item "num" for numeric,
\item "int" for integer,
\item "chr" for string or character,
\item "fac" for factor, and
\item "posix" for POSIX date-time.
}}

\item{notes}{Description of the calculation step. Character string.}

\item{fork_table}{Character string which if supplied will fork the reverse
of the logic provided by \code{dt_syn_bc}.}
}
\description{
An alias for list in ipayipi::dt_calc.
}
\author{
Paul J. Gordijn
}
