% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_sub_change_interval.r
\name{change_intervals}
\alias{change_intervals}
\title{Finds sequential intervals of change}
\usage{
change_intervals(int_dta = NULL, ...)
}
\arguments{
\item{int_dta}{A vector of integer values (e.g., an itentifier in a table).}
}
\value{
Vector with an index indicating the incremental change in a value.
}
\description{
Determines when an interval changes in a vector based on a
the change in integer values in a vector.
}
\details{
This function is an internal function called by others in the
pipeline.
}
\author{
Paul J Gordijn
}
\keyword{function}
\keyword{internal}
\keyword{size;}
\keyword{window}
