% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_sub_sf_dta_chunkr_inter_chk.r
\name{chunkr_inter_chk}
\alias{chunkr_inter_chk}
\title{Generates chunking index table templates}
\usage{
chunkr_inter_chk(
  dta_room = NULL,
  indx = FALSE,
  dta_sets = NULL,
  i_zeros = NULL,
  verbose = FALSE,
  xtra_v = FALSE,
  ...
)
}
\arguments{
\item{dta_room}{Directory where data chunks are to be stored.}

\item{indx}{The indx list from \code{ipayipi::sf_dta_chunkr()}.}

\item{dta_sets}{List of data sets to be written as chunks per the write_tbl.}

\item{i_zeros}{The number of leading zeros used to name files.}

\item{verbose}{Logical. Whether or not to report messages and progress.}

\item{cores}{Number of cores/cpus used to write data chunks.}
}
\value{
Updated chunk index table. This must be used to update a chunked folder's chunk index table.
}
\description{
Internal function. Performs checks on intra chunk integrity. Writes missing date-time values to chunks where missing.
}
\author{
Paul J Gordijn
}
\keyword{chunking}
\keyword{function.}
\keyword{index;}
\keyword{internal}
