% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_sub_sf_dta_chunkr_it.r
\name{chunkr_sub_it}
\alias{chunkr_sub_it}
\title{Generates chunking (i)ndex (t)able templates}
\usage{
chunkr_sub_it(
  dta_indx = FALSE,
  chunk_i = NULL,
  mn = NULL,
  mx = NULL,
  buff_period = "10 years",
  ri = NULL,
  rit = NULL,
  ...
)
}
\arguments{
\item{dta_indx}{Logical. If TRUE generates an index table for a \emph{data set}. If FALSE an index table for a data \emph{chunking folder} is prepared. A _chunking folder contains a fuller chunk index for future data sets---chunking indices are generated by subtraction and addition of the buffer period (\code{buff_period}) to the max and min dates, respectively. An index table for a data set only contains chunk intervals for the min and max times provided with no 'buffer period'.}

\item{chunk_i}{Standardised chunking interval.}

\item{mn}{The minimum date-time of the data for which to generate an index table.}

\item{mx}{The maximum date-time of the data for which to generate an index table.}

\item{buff_period}{This is parsed to \code{lubridate::as.duration()} and subtracted or added to the minimum and maximum date-time's supplied, respectively, when generating the index table.}

\item{ri}{Not implemented.}

\item{rit}{Not yet implemented.}
}
\value{
A chunk index table.
}
\description{
Internal function.
}
\details{
The buffer period around max and min dates of the data prevents frequent rechunking and indexing of data when new data is added that is beyond the limits of the dataset.
}
\author{
Paul J Gordijn
}
\keyword{chunking}
\keyword{function.}
\keyword{index;}
\keyword{internal}
