% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_sub_sf_dta_chunkr_sub_wr.r
\name{chunkr_sub_wr}
\alias{chunkr_sub_wr}
\title{Writes data to chunk files}
\usage{
chunkr_sub_wr(
  dta_room = NULL,
  write_tbl = FALSE,
  dta_sets = NULL,
  i_zeros = 5,
  ri = NULL,
  rit = NULL,
  overwrite = TRUE,
  verbose = FALSE,
  xtra_v = FALSE,
  cores = getOption("mc.cores", 2L),
  ...
)
}
\arguments{
\item{dta_room}{Directory where data chunks are housed.}

\item{write_tbl}{The segment of the index table used to write data chunks. Contains date-time floor and ceiling for filtering data chunks.}

\item{dta_sets}{List of data sets to be written as chunks per the write_tbl.}

\item{i_zeros}{The number of leading zeros used to name files.}

\item{ri}{Not implemented.}

\item{rit}{Not yet implemented.}

\item{overwrite}{Must date time records with the same chunk index be overwritten by matching date-time records in the input data sets? Logical: overwrites if \code{TRUE}.}

\item{verbose}{Whether to report back on function progress. Useful for debugging.}

\item{cores}{Number of processing cores to use for writing data chunks, and checking intra chunk time-series integrity.}
}
\value{
The \code{wrtie_tbl} with updated minimum and maximum date times for each row/chunk index.
}
\description{
Internal function. Checks integrity of continuous data whilst writing data chunks. Does not check between chunks (inter chunk) integrity, only intra chunk integrity.
}
\author{
Paul J Gordijn
}
\keyword{chunking}
\keyword{function.}
\keyword{index;}
\keyword{internal}
