% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_sub_circular_mean.R
\name{circular_mean}
\alias{circular_mean}
\title{A wrapper for the circular.mean function}
\usage{
circular_mean(x, u = "degrees", ignore_nas = FALSE, verbose = FALSE, ...)
}
\arguments{
\item{x}{Vector of angles.}

\item{u}{Angle units. Either degrees or radians.}

\item{verbose}{Logical indicating whether to print error messages etc.}
}
\value{
The mean angle.
}
\description{
Find the mean angle/azimuth
}
\author{
Paul J. Gordijn
}
