% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_sub_dt_clean_param_eval.r
\name{clean_param_eval}
\alias{clean_param_eval}
\title{Evaluate and prepare tables for joins}
\usage{
clean_param_eval(
  clean_f = "hampel",
  phen_names = NULL,
  segs = NULL,
  seg_fuzz = NULL,
  seg_na_t = 0.75,
  w_size = 21,
  mad_dev = 3,
  last_rule = FALSE,
  tighten = 0.65,
  full_eval = FALSE,
  station_file = NULL,
  ppsij = NULL,
  f_params = NULL,
  sfc = NULL,
  verbose = FALSE,
  xtra_v = FALSE,
  ...
)
}
\arguments{
\item{clean_f}{Algorithm name. Only "hampel" supported.}

\item{phen_names}{Vector of the phenomena names that will be evaluated by the hampel filter. If NULL the function will not run.}

\item{segs}{Vector of station data table names which contain timestamps used to slice data series into segments with independent outlier detection runs. If left \code{NULL} (default) the run will be from the start to end of data being evaluated.}

\item{seg_fuzz}{String representing the threshold time interval between the list of segment date-time}

\item{seg_na_t}{Fractional tolerace of the amount of NA values in a segment for linear interpolation of missing values.}

\item{w_size}{Window size for the hempel filter. Defaults to ten.}

\item{mad_dev}{Scalar factor of MAD (median absolute deviation). Higher values relax oulier detection. Defaults to the standard of three.}

\item{last_rule}{Whether or not to reapply stored rules in the outlier rule column. TRUE will apply old rules.}

\item{tighten}{Scaling fraction between zero and one that sensitizes the detection of outliers near the head and tail ends of segments. The fraction is multiplied by the mad_dev factor.}

\item{full_eval}{param f_params}

\item{station_file}{Name of the station being processed.}

\item{ppsij}{Data processing \code{pipe_seq} table from which function parameters and data are extracted/evaluated. This is parsed to this function automatically by \code{ipayipi::dt_process()}.
param sf}

\item{sfc}{List of file paths to the temporary station file directory. Generated using \code{ipayipi::open_sf_con2()}.}

\item{verbose}{Logical. Whether or not to report messages and progress.}

\item{xtra_v}{Logical. Whether or not to report xtra messages, progess, plus print data tables.}
}
\description{
Flexible and fast joining of \code{ipayipi} data.
}
\details{
\strong{Join types}
\itemize{
\item \code{full_join}: Merges matching \code{x_tbl} and \code{y_tbl} rows, and retains all unmatching records.
\item \code{left_join}: Matching \code{x_tbl} and \code{y_tbl} rows, plus all unmatched \code{x_tbl} rows.
\item \code{right_join}: Matching \code{x_tbl} and \code{y_tbl} rows, plus all unmatched \code{y_tbl} rows.
\item \code{inner_join}: Only retains matching \code{x_tbl} and \code{y_tbl} rows.
}
}
\author{
Paul J. Gordijn
}
