% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_close_sf_con.r
\name{close_sf_con}
\alias{close_sf_con}
\title{Opens a connection to an ipayipi station file.}
\usage{
close_sf_con(
  pipe_house = NULL,
  station_file = NULL,
  overwrite = TRUE,
  append = FALSE,
  ...
)
}
\arguments{
\item{pipe_house}{List of pipeline directories. \strong{See} \code{ipayipi::ipip_init()} \strong{for details}.}

\item{station_file}{Path to 'ipip' station file.}
}
\value{
Station file connection file paths. To be used in the pipeline.
}
\description{
Package subroutine for opening a station file for work.
}
\details{
This function is an internal function called by others in the
pipeline.
}
\author{
Paul J Gordijn
}
\keyword{add}
\keyword{edit}
\keyword{station;}
\keyword{to}
\keyword{write}
