% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cs_stom}
\alias{cs_stom}
\title{Header harvest: SAEON Terrestrial Observation Monitor}
\format{
\subsection{'cs_stom'}{

An ipayipi data_setup list for imbibing (\code{ipayipi::imbibe_raw_logger_dt()})
data from the SAEON Terrestrial Observation Monitor (STOM).
Data from STOM follows SAEON iPayipi's phenomena nomenclature---when running
\code{phenomena_sts()} set \code{external_phentab = ipayipi::sts_phens}.
a \code{data_setup}.
}
}
\source{
Internal.
}
\usage{
cs_stom
}
\description{
ipayipi data setup list for Cambell Scientific TOA5 format files
}
\keyword{datasets}
