% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cs_toa5}
\alias{cs_toa5}
\title{Header harvest: Cambell Scientific, TOA5}
\format{
\subsection{'cs_toa5'}{

An ipayipi data_setup list for imbibing (\code{ipayipi::imbibe_raw_logger_dt()})
TOA5 formatted data. The 'location' name is not specified. Only generic
information is provided for harvesting of logger file header information.
__See \code{ipayipi::imbibe_raw_logger_dt()} documentation for a describtion of
a \code{data_setup}.
}
}
\source{
Internal.
}
\usage{
cs_toa5
}
\description{
ipayipi data setup list for Cambell Scientific TOA5 format files
}
\keyword{datasets}
