% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_redundant_gw_sub_data_list.r
\name{data_list}
\alias{data_list}
\title{List raw data files files in a directory}
\usage{
data_list(input_dir = NULL, file_type = NULL, recurr = FALSE, ...)
}
\arguments{
\item{input_dir}{The folder, that is, directory in which to search
for solonist xle files.}

\item{file_type}{String. Select from the following options:
"xle", "rain_hobo".}

\item{recurr}{Should the search be recursive i.e., also include
subdirectories.}
}
\value{
a vector of xle file paths.
}
\description{
A function to search and list data files files in a directory.
}
\details{
This is used as a helper function. The search is based on the
file extension for the specified file type. The following strings
are excluded for each of the following file types:
\itemize{
\item \strong{xle} '.lev' (old solonist files); 'Compensat' (compensated
solonist files)
\item \strong{rain_hobo} N/A
Additional files can be added and rules for search exclusion in the
function.
}
}
\examples{
# Generate a list of xle files in the working directory
dir <- "." # define the directory
gw_xle_list(input_dir = dir, file_type = "xle", recurr = FALSE)
}
\author{
Paul J. Gordijn
}
\keyword{data}
\keyword{files}
\keyword{import}
\keyword{into}
\keyword{pipeline.}
\keyword{processing}
