% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gw_f2_driffter2.r
\name{driffter2}
\alias{driffter2}
\title{A function to correct for \strong{instrument drift}.}
\usage{
driffter2(
  input_dir = ".",
  recurr = FALSE,
  rng_side = 3,
  rng_max = 5,
  robust = TRUE,
  prompted = TRUE,
  ...
)
}
\arguments{
\item{input_dir}{The working directory were where the R data
solonist files are stored.}

\item{recurr}{Setting this to TRUE will enable searching
for R data files in sub directories i.e., a recursive search}

\item{rng_side}{The window side (half window) for calculation
drift offsets. The size of this range represents the size of
the number of point measurements around a central value which
is used to estimate drift correction.}

\item{rng_max}{If there are NA values in the window determined by
\strong{rng_side} the window side size will expand until reaching this
maximum value.}

\item{prompted}{If TRUE, a command line prompt will be used to
enable selection of which files in the working directory require
drift correction.}

\item{method}{driffter2() differs from gw_driffter() in that it includes
a method for simple drift correction. Ironically, the simple method
was designed to be more effective for data series with high frequency
large amplitude patterns. The methods are: "mean", "median", and
"fv". "fv" stands for face value.}
}
\value{
R solonist object with corrected outlier values and flags indicating
outlier values plus the ourlier rule used.
}
\description{
The function requires that manual water level, that is,
dipper readings to calculate drift. Drift is assumued to be linear.
The function also requires that datum information has been
appended to the level logger file before drift correction.
}
\details{
Linear drift correction may not be as straight forward as
may be anticipated. This function works in stages at each point in
the water level data series where there are corresponding
calibration measurements.
Firstly, at each dipper reading, 'windows' around the central water
level data point, which corresponds to the (nearest) time when the
dipper reading was taken, are manufactured. The size of a window size
is specified using the 'rng_side' paramenter. The amount by which the
central water level data point is offset is estimated by calculating
the difference in the calibration value and the median or mean (see
'robust' parameter) water level of the specified window. If there is
'leading' and 'tailing' data around the central water level value,
the window size is equal to two times the 'rng_side' parameter. If
there is only leading or trailing data, the window size is equal to
'rng_side' (plus one for the central window value).
Note that if there is missing data in the 'leading' or 'training' data
windows, the window can be set to expand, until sufficient non-null data
is found. The expansion of the window in search of non-null data is set by
the 'rng_max' parameter.
Once the offset values are estimated for each central data point,
graph theory, that is, linear algebra is used to esimtate offsets for
each data point in the full data series. The gradient and y-intercept for
this is estimated using 'half-windows' in segments of the data series
between each central data point (which correspond to when calibration
measures were taken). Determination of window sizes and indicies (i.e.,
index number in the time series) is done using the 'win_rows' function.
For wiggly data (high variability) it is better to used smaller windows
for drift correstion. More needs to be done using this function to explore
the effectiveness of this drift correction method. It would be good to
include a dynamic window size based on the rate of change in water level,
that is, the first derivative.
}
\keyword{correction;}
\keyword{drift}
\keyword{drift;}
\keyword{linear}
