% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_dt_drift_param_eval.R
\name{drift_param_eval}
\alias{drift_param_eval}
\title{Evaluate arguments for linear drift correction.}
\usage{
drift_param_eval(
  phen = NULL,
  wside = 3,
  wside_max = 12,
  robust = TRUE,
  edge_adj = NULL,
  pipe_house = NULL,
  station_file = NULL,
  full_eval = FALSE,
  f_params = NULL,
  ppsij = NULL,
  sfc = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{phen}{String name of phenomena name ('phen_name') to undergo drift correction.}

\item{wside}{The window side (half window) for calculation of drift offsets. The size of this range represents the size of the number of point measurements around a central value which is used to estimate drift correction.}

\item{wside_max}{If there are \code{NA} values in the window given by \code{wside} the window side size will expand until reaching this maximum value.}

\item{robust}{Use the median instead of the mean to calculate the central tendancy of the window of interest. Reduces the influence of outliers on drift parametization.}

\item{station_file}{Name of the station being processed.
details}

\item{full_eval}{Logical that indicates the depth of evaluation. This argument is provided parsed internally. Full evaluation can only be completed within \code{ipayipi::dt_process()} sequence evaluation.}

\item{f_params}{Argument used for parsing \code{pipe_seq} function parameters to \code{drift_param_eval()} when \code{full_eval == TRUE}.}

\item{ppsij}{Data processing \code{pipe_seq} table from which function parameters and data are extracted/evaluated. This is parsed to this function automatically by \code{ipayipi::dt_process()}.}

\item{sfc}{List of file paths to the temporary station file directory. Generated using \code{ipayipi::sf_open_con()}.}

\item{verbose}{Logical. Whether or not to report messages and progress.}
}
\description{
Performs linear drift correction of a single phenomena using calibration measurements.
}
\author{
Paul J. Gordijn
}
