% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h1_dt_calc.R
\name{dt_calc}
\alias{dt_calc}
\title{Perform data processing calculations}
\usage{
dt_calc(
  sfc = NULL,
  pipe_house = NULL,
  station_file = NULL,
  f_params = NULL,
  ppsij = NULL,
  verbose = FALSE,
  xtra_v = FALSE,
  chunk_v = FALSE,
  dt_format = c("Ymd HMOS", "Ymd HMS", "Ymd IMOSp", "Ymd IMSp", "ymd HMOS", "ymd HMS",
    "ymd IMOSp", "ymd IMSp", "mdY HMOS", "mdY HMS", "mdy HMOS", "mdy HMS", "mdy IMOSp",
    "mdy IMSp", "dmY HMOS", "dmY HMS", "dmy HMOS", "dmy HMS", "dmy IMOSp", "dmy IMSp"),
  dt_tz = "Africa/Johannesburg",
  ...
)
}
\arguments{
\item{sfc}{List of file paths to the temporary station file directory. Generated using \code{ipayipi::sf_open_con()}.}

\item{pipe_house}{A pipeline's directory structure. Generated using \code{ipayipi::ipip_house()}.}

\item{station_file}{Name of the station being processed.}

\item{f_params}{Function parameters evaluated by \code{ipayipi::calc_param_eval()}. These are internally parsed to \code{dt_calc()} from \code{dt_process()}.}

\item{ppsij}{Data processing \code{pipe_seq} table from which function parameters and data are extracted/evaluated. This is parsed to this function automatically by \code{ipayipi::dt_process()}.}

\item{verbose}{Logical. Whether or not to report messages and progress.}

\item{xtra_v}{Logical. Whether or not to report xtra messages, progess, plus print data tables.}

\item{chunk_v}{Whether or not to indicate the start of data chunking.}

\item{dt_format}{The function attempts to work out the date-time format from a vector of format types supplied to this argument. The testing is done via \code{lubridate::parse_date_time()}. \code{lubridate::parse_date_time()} prioritizes the tesing of date-time formats in the order vector of
formats supplied. The default vector of date-time formats supplied should work well for most logger outputs.}

\item{dt_tz}{Recognized time-zone (character string) of the data locale. The default for the package is South African, i.e., "Africa/Johannesburg" which is equivalent to "SAST".}
}
\value{
A list containing the processed data sets 'dts_dt'.
}
\description{
Calculations are parsed to data.table.
}
\details{
Note that gap metadata will be transferred to the station file's 'gap' table.

*Testing import of gap info and saving to sfc gaps
}
\author{
Paul J. Gordijn
}
