% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_h1_station_dt_calc_chain.r
\name{dt_calc_chain}
\alias{dt_calc_chain}
\title{Perform data processing calculations}
\usage{
dt_calc_chain(
  station_file = NULL,
  dta_in = NULL,
  input_dt = NULL,
  output_dt = NULL,
  output_dt_preffix = "dt_",
  output_dt_suffix = NULL,
  f_params = NULL,
  dt_working = NULL,
  ...
)
}
\arguments{
\item{station_file}{Name of the station being processed.}

\item{dta_in}{A list of data to be processed.}

\item{input_dt}{}

\item{output_dt}{}

\item{f_summary}{A summary table of function parameters. This
summary table is stored in the station file object. If there
is a summary table for the function, differences between the
\code{f_params} and table are checked and the 'f_summary' table
is updated---triggering a re-calculation of the whole time series instead
of for a slice of the series.}
}
\value{
A list containing the processed data sets 'dts_dt'.
}
\description{
Calculations are parsed to data.table.
}
\author{
Paul J. Gordijn
}
