% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h1_dt_clean.R
\name{dt_clean}
\alias{dt_clean}
\title{Dev function: Apply filter windows to time series data}
\usage{
dt_clean(
  phens = NULL,
  w = 21,
  madf = 3,
  align = "left",
  seg = NULL,
  cush = TRUE,
  clean_f = "hampel",
  owrite = TRUE,
  na_t = 0.25,
  tighten = 0.65,
  station_file = NULL,
  pipe_house = NULL,
  sfc = NULL,
  ppsij = NULL,
  f_params = NULL,
  verbose = FALSE,
  xtra_v = FALSE,
  chunk_v = FALSE,
  ...
)
}
\arguments{
\item{phens}{Vector of the phenomena names that will be evaluated by the hampel filter. If NULL the function will not run.}

\item{w}{Window size for the hempel filter. Defaults to ten.}

\item{madf}{Scalar factor of MAD (median absolute deviation). Higher values relax oulier detection. Defaults to the standard of three.}

\item{align}{Corresponds to data.table's frollapply function's 'align' argument. Options: 'right' -- value being evaluated is on the right with respect to preceeding values; 'left' -- opposite of 'right'; 'centre' -- value under evaluation mid window (if even, biased to the left).}

\item{seg}{Name of column to segment filtering on. Cleaning within each data segment or slice will be run independently.}

\item{cush}{Logical indicating whether to prevent NAs at data start and end dates owing to window/filter alignment. \code{TRUE} will \bold{cush}ion data start and ends where necessary by adjust window alignment properties (within segments). \code{FALSE} does not prevent NA values.}

\item{clean_f}{Algorithm name. Only "hampel" supported.}

\item{owrite}{Logical. Should outliers be overwritten in the data? In any case a table of original and replacement values is generated by this function.}

\item{na_t}{Threshold (as a proportion of window length) of NA values tolerated. If the number of NAs exceeds the threshold, filtering will not be performed. Must be supplied as a vector or list with subsequent values corresponding to each filter 'run'.}

\item{tighten}{Not yet implemented. Scaling fraction between zero and one that sensitizes the detection of outliers near the head and tail ends of segments. The fraction is multiplied by the mad factor.}

\item{station_file}{Name of the station being processed.}

\item{pipe_house}{If the \code{pipe_house} argument is provided the \code{pipe_house$d4_ipip_room} will be used instead of the \code{input_dir}.}

\item{sfc}{List of file paths to the temporary station file directory. Generated using \code{ipayipi::sf_open_con()}.}

\item{ppsij}{Data processing \code{pipe_seq} table from which function parameters and data are extracted/evaluated. This is parsed to this function automatically by \code{ipayipi::dt_process()}.}

\item{f_params}{Function parameters evaluated by \code{ipayipi::calc_param_eval()}. These are internally parsed to \code{dt_calc()} from \code{dt_process()}.}

\item{verbose}{Logical. Whether or not to report messages and progress.}

\item{xtra_v}{Logical. Whether or not to report xtra messages, progess, plus print data tables.}

\item{chunk_v}{Logical. Print data chunking messages. Useful for debugging/digging into chunking methods.}

\item{seg_fuzz}{Not yet implemented. String representing the threshold time interval between the list of segment date-time}
}
\description{
Function under development. Uses the non-parametric hampel filter to detect anomalies and impute values in a univariate series of numeric data.
}
\details{
This function employs a user-customised hampel filter to check data for outliers and impute missing values with the median of the filter window. Function under development.
\itemize{
\item Need to add generated table details to the phen_ds summary table!
}
}
\author{
Paul J. Gordijn
}
\keyword{data,}
\keyword{detection,}
\keyword{imputation,}
\keyword{outlier}
\keyword{univariate}
\keyword{value}
