% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_sub_dt_dta_open_filter.r
\name{dt_dta_filter}
\alias{dt_dta_filter}
\title{dt: Opens data table}
\usage{
dt_dta_filter(dta_link = NULL, ppsij = NULL, ...)
}
\arguments{
\item{dta_link}{Link from \code{ipayipi::sf_dta_read()}.}

\item{ppsij}{Pipeline processing table step info.}
}
\value{
Filtered data link.
}
\description{
Internal function that filters data by start and end date-times provided by the processing pipeline ppsij.
}
\author{
Paul J. Gordijn
}
