% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_sub_dt_dta_open.r
\name{dt_dta_open}
\alias{dt_dta_open}
\title{dt: Opens data table}
\usage{
dt_dta_open(dta_link = NULL, start_dttm = NULL, end_dttm = NULL, ...)
}
\arguments{
\item{dta_link}{Link from \code{ipayipi::sf_dta_read()}.}

\item{start_dttm}{Earliest data filter date-time.}

\item{end_dttm}{Latest data filter date-time.}
}
\value{
Data in date.table format.
}
\description{
Internal function that handles the variety of ways data are harvested; returning data from either chunked or unchunked sources ready for evaulation in 'dt' processing steps.
}
\author{
Paul J. Gordijn
}
