% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_h1_station_dt_harvest.r
\name{dt_harvest}
\alias{dt_harvest}
\title{'dt' processing pipeline: harvest data from another station or table.}
\usage{
dt_harvest(
  station_file = station_file,
  f_params = NULL,
  time_interval = NULL,
  sf = NULL,
  dt_working = NULL,
  ...
)
}
\arguments{
\item{station_file}{File path of the station being processed.}

\item{f_params}{A vector of phenomena name to be extracted from the
harvested data table. If NULL all column names are extracted.}

\item{time_interval}{The desirsed time_tinterval associated with the
extracted data. If this is \code{NULL} then the table is extracted as is.}

\item{hsf_table}{The path of the directory in which to search for the
external data and/or data table.}

\item{ext_station}{The name (or keyword) of the station from which to extract
data.}

\item{output_dt}{The output table name where harvested data is extracted too.}
}
\description{
Used to extract data from another table, station, or source.
}
\author{
Paul J. Gordijn
}
