% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_h1_station_dt_join.r
\name{dt_join}
\alias{dt_join}
\title{Perform data processing calculations}
\usage{
dt_join(
  join = "full_join",
  x_tbl = NULL,
  y_tbl = NULL,
  y_phen_names = NULL,
  x_key = "date_time",
  y_key = "date_time",
  nomatch = NA,
  mult = "all",
  roll = FALSE,
  rollends = FALSE,
  allow.cartesian = FALSE,
  time_seq = NULL,
  fuzzy = NULL,
  station_file = NULL,
  dta_in = NULL,
  input_dt = NULL,
  output_dt = NULL,
  output_dt_preffix = "dt_",
  output_dt_suffix = NULL,
  f_params = NULL,
  dt_working = NULL,
  ppsij = ppsij,
  ...
)
}
\arguments{
\item{station_file}{Name of the station being processed.}

\item{dta_in}{Arugment for pipeline processing. Instead of providing the 'x' and 'y' tables to \code{x_tbl} and \code{y_tbl} The pipeline feeds data to this argument. A list of data to be processed.}

\item{input_dt}{}

\item{output_dt}{}

\item{f_summary}{A summary table of function parameters. This
summary table is stored in the station file object. If there
is a summary table for the function, differences between the
\code{f_params} and table are checked and the 'f_summary' table
is updated---triggering a re-calculation of the whole time series instead
of for a slice of the series.}
}
\value{
A list containing the processed data sets 'dts_dt'.
}
\description{
Calculations are parsed to data.table.
}
\author{
Paul J. Gordijn
}
