% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_f0_station_dt_process.r
\name{dt_process}
\alias{dt_process}
\title{Pipeline data processing}
\usage{
dt_process(
  station_file = NULL,
  ipip_room = ".",
  pipe_process = NULL,
  time_intervals = NULL,
  output_dt_preffix = "dt_",
  output_dt_suffix = NULL,
  overwrite_pipe_memory = TRUE,
  ...
)
}
\arguments{
\item{station_file}{Name of the station being processed.}

\item{ipip_room}{Directory where station data is stored.}

\item{time_intervals}{Vector of strings describing time intervals.
Each element of the vector corresponds to the aggregation of data
into a new table. Each string element should begin with an integer,
then the standard period, e.g., "1 month" or "10 mins". The period must
be recognised by \code{ipayipi::sts_interval_name()}.}

\item{output_dt_preffix}{The output table preffix which defaults to
"dt_".}

\item{output_dt_suffix}{A custom suffix to be appended to the output
tables name.}

\item{overwrite_pipe_memory}{Logical. If TRUE then extant pipeline steps,
which are summarised in the 'pipe_process_summary' data table (\emph{see
details}), are modified by arguments in the pipe_process argument.}
}
\description{
Processes data in sequential steps
}
\details{
This function forms the basis of setting up a sequential data
processing pipeline. This allows the extraction and preparation of
raw, or other data from a data table in an 'ipayipi' station file,
and further processing of this data.

Different portions of the pipeline are summarised in the
'pipe_process_summary' data table. The sequential steps in the processing
pipeline are numbered by the field \code{dt_1}.
\enumerate{
\item \code{dt_n}: Sequential number of data tables produced by the pipeline.
Seperate data tables will be processed following this order.
\item \code{dt_name}: The name of the data processing table.
\item \code{dtp_n}: Sequential numbering of pipeline process within a data table.
\item \code{f}: The processing function.
\item 'start_dttm': Earliest date time for which data have been processed in
this table.
\item 'end_dttm': Latest date time for which data have been processed in this
table.
}
\itemize{
\item \emph{\strong{Aggregating data by time periods}}:
}
}
\author{
Paul J. Gordijn
}
\keyword{data}
\keyword{pipeline;}
\keyword{processing}
\keyword{processing;}
\keyword{steps}
