% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_h0_station_dt_process_batch.r
\name{dt_process_batch}
\alias{dt_process_batch}
\title{Append logger data files to 'station file'}
\usage{
dt_process_batch(
  pipe_house = NULL,
  pipe_seq = NULL,
  unwanted_tbls = "_tmp",
  output_dt_preffix = "dt_",
  output_dt_suffix = NULL,
  overwrite_pipe_memory = FALSE,
  station_ext = ".ipip",
  wanted = NULL,
  unwanted = NULL,
  prompt = FALSE,
  verbose = FALSE,
  cores = getOption("mc.cores", 2L),
  ...
)
}
\arguments{
\item{pipe_house}{List of pipeline directories. \strong{See}
\code{ipayipi::ipip_init()} \strong{for details}.}

\item{station_ext}{The extension of the station file. Defaults to 'ipip'.}

\item{prompt}{Prompt logical passed for selection of 'nomvet' files.}

\item{verbose}{Print some details on the files being processed? Logical.}

\item{cores}{Number of CPU's to use for processing in parallel. Only applies when working on Linux.}

\item{overwrite_sf}{If \code{TRUE} then original data (station data) is
disgarded in favour of new data. If TRUE both data sets will be evaluated
and where there are NA values, a replacement, if available, will be used
to replace the NA value. Defaults to FALSE.}

\item{by_station_table}{If TRUE then multiple station tables will not be
kept in the same station file. Defaults to FALSE.}

\item{sts_file_ext}{Extension of standardised files in the 'nomvet_room'.}
}
\description{
Appends phenomena data and metadata into continuous or
discontinuous data streams in the form of station files.
}
\details{
This function searches for all standardised files in the \code{nomvet_room}
and existing 'station files' in the \code{ipip_room}. Files from the
\code{nomvet_room} that have not been appended to station files will be
appended to appropriate stations. Appending is done via the
\code{ipayipi::append_station()}, \code{ipayipi::append_phen_data()}, and
\code{ipayipi::append_tables()} functions.

If the data is continuous the function ensures continuous date-time values
between missing data chuncks. Missing values are fillled as NA. Event-based
or discontinuous data are not treated as continuous data streams.

New phenomena are seemlessly included in the appending process. Moreover,
each phenomena is evaluated in turn allowing for a choise of overwring old
data with new data or \emph{vice versa}.
}
\author{
Paul J. Gordijn
}
\keyword{append}
\keyword{data}
\keyword{data;}
\keyword{logger}
\keyword{meteorological}
\keyword{pipeline;}
