% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_dta_availability0.r
\name{dta_availability0}
\alias{dta_availability0}
\title{Plot the coverage of ipayipi station files}
\usage{
dta_availability0(
  pipe_house = NULL,
  station_ext = ".ipip",
  gap_problem_thresh_s = 6 * 60 * 60,
  event_thresh_s = 10 * 60,
  start_dttm = NULL,
  end_dttm = NULL,
  plot_tbls = NULL,
  meta_events = "meta_events",
  verbose = FALSE,
  wanted = NULL,
  unwanted = NULL,
  recurr = FALSE,
  prompt = FALSE,
  keep_open = TRUE,
  ...
)
}
\arguments{
\item{pipe_house}{List of pipeline directories. \strong{See} \code{ipayipi::ipip_house()} \strong{for details}.}

\item{station_ext}{Required argument. File extension of the station files in the pipe_house directory 'ipip_room'.}

\item{gap_problem_thresh_s}{\emph{Parsed to} \code{ipayipi::gap_eval()}.}

\item{event_thresh_s}{\emph{Parsed to} \code{ipayipi::gap_eval()}.}

\item{start_dttm}{For plotting data.}

\item{end_dttm}{For plotting data.}

\item{plot_tbls}{Vector of table names to include when plotting data availability.}

\item{meta_events}{\emph{Parsed to} \code{ipayipi::gap_eval()}.}

\item{verbose}{Logical. Whether or not to report messages and progress.}

\item{wanted}{A string of keywords to use to filter which stations are selected for processing. Multiple search kewords should be seperated with a bar ('|'), and spaces avoided unless part of the keyword.}

\item{unwanted}{Similar to wanted, but keywords for filtering out unwanted stations.}

\item{recurr}{Should the function search recursively into sub directories for hobo rainfall csv export files? TRUE or FALSE.}

\item{prompt}{Should the function use an interactive file selection function otherwise all files are returned. TRUE or FALSE.}

\item{keep_open}{Logical. Keep \emph{hidden} 'station_file' open for ease of access. Defaults to \code{TRUE}.}

\item{cores}{Number of CPU's to use for processing in parallel. Only applies when working on Linux.}
}
\value{
A list containing a plot, which shows the availability of data, the gap data as formatted for plotting. The plot is produced using \code{ggplot2::ggplot()}.
}
\description{
Uses ggplot2 to show the availability of data for a select group of station's and their raw data tables. \emph{Data availability is based on overall data logging time not NA values}.
}
\details{
Gap data for each station and respective tables are extracted from the station, using \code{ipayipi::gap_eval()}. Gap tables are combined with the full record of data in a data summary table and plotted using ggplot2. Note that 'gaps' can be edited by imbibing metadata into a stations record, \emph{see} \code{ipayipi::gap_eval()} for details.
}
\author{
Paul J. Gordijn
}
\keyword{data}
\keyword{data;}
\keyword{missing}
\keyword{over}
\keyword{plotting;}
\keyword{reporting;}
\keyword{view;}
