% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_dta_flat_pull_discnt.R
\name{dta_flat_pull_discnt}
\alias{dta_flat_pull_discnt}
\title{Pulls discontinuous data}
\usage{
dta_flat_pull_discnt(
  input_dir = ".",
  pipe_house = NULL,
  tbl_names = NULL,
  gaps = TRUE,
  output_dir = NULL,
  wanted = NULL,
  unwanted = NULL,
  out_csv = FALSE,
  out_tab_name = NULL,
  out_csv_preffix = "",
  recurr = TRUE,
  prompt = FALSE,
  verbose = FALSE,
  xtra_v = FALSE,
  chunk_v = FALSE,
  ...
)
}
\arguments{
\item{input_dir}{The directory in which to search for stations form which to extract time-series data.}

\item{pipe_house}{If the \code{pipe_house} argument is provided the \code{pipe_house$d4_ipip_room} will be used instead of the \code{input_dir}.}

\item{tbl_names}{Vector of table names in a station files from which to extract data. Add items to the vector such that only the first matching table from a station is selected. Table names are selected via the \code{\%in\%} argument.}

\item{gaps}{If \code{TRUE} then a logical table of the same dimensions as the extracted flat data is saved to disk The same file name as for the data but with a '_gaps' suffix is used. Where gaps in the data have been detected in the pipeline process respective data rows are set to \code{TRUE}; \code{FALSE} indicates no data gap. Gap data is produced by running \code{ipayipi::gap_eval_batch()}.}

\item{output_dir}{The output directory where an output csv file is saved.}

\item{wanted}{A strong containing keywords to use to filter which stations are selected for processing. Multiple search kewords should be seperated with a bar ('|'), and spaces avoided unless part of the keyword.}

\item{unwanted}{Similar to wanted, but keywords for filtering out unwanted stations.}

\item{out_csv}{Logical. If TRUE a csv file is exported to the output directory.}

\item{out_tab_name}{If left \code{NULL} (default), the output table name will be \code{paste0(out_csv_preffix, '_', ri)}, where ri is determined using the \code{ipayipi::sts_interval_name()} or provided by the \code{ri} argument.}

\item{out_csv_preffix}{Preffix for the output csv file. The phenomena name and then time interval by which the data are summarised are used as a suffix.}

\item{recurr}{Whether to search recursively through folders. Defaults to TRUE.}

\item{prompt}{Logical. If \code{TRUE} a prompt will be called so that the user can interactively select station files.}

\item{verbose}{Whether to report on progress. Logical.}

\item{xtra_v}{Extra verbose. Logical.}

\item{chunk_v}{More messages when chunking data? Logical.}

\item{phen_name}{The field/column/name of the phenomena which to join into a single data table.}

\item{file_ext}{The extension of the stations from where data will be extracted. Defaults to ".ipip".}
}
\value{
A list containing 1) the summarised data in a single data.table, 2) a character string representing the time interval by which the data has been summarised, 3) the list of stations used for the summary, 4) the name of the table and the name of the field for which data was summarised, & 5) the station file extension uesd for querying data.
}
\description{
Queries \strong{discontinuous} data from multiple stations and converts to 'long' format.
}
\details{
Note this function only extracts data from decompressed station files. If a station has not be decompressed the function will take time decompressing.
}
\author{
Paul J. Gordijn
}
\keyword{data}
\keyword{data;}
\keyword{long-format}
\keyword{pipeline;}
\keyword{query.}
\keyword{summarise}
\keyword{time-series}
