% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_dta_flat_push.R
\name{dta_flat_push}
\alias{dta_flat_push}
\title{Pushes flattened station data into station files}
\usage{
dta_flat_push(
  stations = NULL,
  flat_dta = NULL,
  tbl_names = NULL,
  phen_name = NULL,
  ri = NULL,
  preds = TRUE,
  recurr = TRUE,
  file_ext = ".ipip",
  ...
)
}
\arguments{
\item{stations}{Vector of file paths to each station to which data will be pushed. File paths must be reletive to the current working directory.}

\item{flat_dta}{Long format data to be pushed back to respective stations and tables replacing extant station data.}

\item{tbl_names}{Vector of table names in a station files from which to extract data. Add items to the vector such that only one table per station is selected. Table names are selected via the \code{\%in\%} argument.}

\item{phen_name}{The field/column/name of the phenomena which to join into a single data table.}

\item{ri}{Record-interval string. This function will guess the record interval used to generate a flat table. However, the guess may not be possible if there is insufficient data.}

\item{file_ext}{The extension of the stations from where data will be extracted. Defaults to ".ipip".}

\item{output_dir}{The output directory where an output csv file is saved.}

\item{wanted}{A strong containing keywords to use to filter which stations are selected for processing. Multiple search kewords should be seperated with a bar ('|'), and spaces avoided unless part of the keyword.}

\item{unwanted}{Similar to wanted, but keywords for filtering out unwanted stations.}

\item{out_csv}{Logical. If TRUE a csv file is exported to the output directory.}

\item{out_csv_preffix}{Preffix for the output csv file. The phenomena name and then time interval by which the data are summarised are used as a suffix.}

\item{prompt}{Logical. If \code{TRUE} a prompt will be called so that the user can interactively select station files.}
}
\value{
A list containing 1) the summarised data in a single data.table, 2) a character string representing the time interval by which the data has been summarised, 3) the list of stations used for the summary, 4) the name of the table and the name of the field for which data was summarised.
}
\description{
Effectively the reverse of \code{ipayipi::dta_flat_pull()}.
}
\details{
Need details here
}
\author{
Paul J. Gordijn
}
\keyword{data}
\keyword{data.}
\keyword{data;}
\keyword{long-format}
\keyword{pipeline;}
\keyword{summarise}
\keyword{time-series}
