% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_dta_list.R
\name{dta_list}
\alias{dta_list}
\title{Generate file list}
\usage{
dta_list(
  input_dir = ".",
  file_ext = NULL,
  wanted = NULL,
  unwanted = NULL,
  prompt = FALSE,
  single_out = FALSE,
  recurr = FALSE,
  verbose = FALSE,
  xtra_v = FALSE,
  ...
)
}
\arguments{
\item{input_dir}{Folder in which to search for R solonist data files.}

\item{file_ext}{Character string of the file extension of the input data files. E.g., ".csv" for hobo rainfall file exports. This input character string should contain the period as in the previous sentence.}

\item{wanted}{Regex string of files to select for listing. Seperate search tags by using the bar character '|'.}

\item{unwanted}{Regex string of files to filter out the listing. Seperate search tags by using the bar character '|'.
in the import.}

\item{prompt}{Set to TRUE for interactive mode. Note this will not work if embedded in a parallel processing instance.}

\item{single_out}{Logical. If \code{TRUE} then only a single item in the list of files will be returned. When \code{prompt} is \code{TRUE} then this can be achieved interactively.}

\item{recurr}{Should the function search recursively i.e., thorugh sub-folders as well - \code{TRUE}/\code{FALSE.}}

\item{verbose}{Logical. Print some details and progress of function progress?}

\item{xtra_v}{Logical. Should some 'x'tra messaging be done? Use to help diagnose problems, and for guidance.}
}
\value{
A vector of selected files.
}
\description{
Interactively or not, create a list (actually vector) of files in a specified directory.
}
\author{
Paul J. Gordijn
}
\keyword{Data}
\keyword{R}
\keyword{data}
\keyword{files;}
\keyword{hobo}
\keyword{solonist}
\keyword{vector;}
