% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_sub_extend_long.r
\name{dttm_extend_long}
\alias{dttm_extend_long}
\title{Extend date time series data}
\usage{
dttm_extend_long(data_sets = NULL, ri = NULL, remove_overlap = FALSE, ...)
}
\description{
Internnal function to generate time series data tables to
fill in missing date time values in a list of non-overlapping data sets
with their date and time value in a column named "date_time".
}
\author{
Paul J. Gordijn
}
