% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{file_read_meta}
\alias{file_read_meta}
\title{File reading metadata}
\format{
\subsection{'file_read_meta'}{

iPayipi file_read_meta table used for setting defaults when attempting to
read/imbibe raw data files. This table is used internally in:
\code{ipayipi::imbibe_raw_batch()}. If the \code{file_ext} and/or \code{col_dlm}
arguments are \code{NULL} this table assists with reading in data in the
correct format. Additional formats can be inserted into this data object.
}

\subsection{'flag_tbl'}{

iPayipi flag_tbl table used as input for ... when attempting to
}
}
\source{
Internal

Internal
}
\usage{
file_read_meta

file_read_meta
}
\description{
File defaults used when imbibing raw logger data

Flagging thresholds and other checks for meteorological data
}
\keyword{datasets}
