% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/y_f0_gap_eval.R
\name{gap_eval}
\alias{gap_eval}
\title{Generates 'gap' (missing data) table}
\usage{
gap_eval(
  pipe_house = NULL,
  station_file = NULL,
  gap_problem_thresh_s = 6 * 60 * 60,
  event_thresh_s = 10 * 60,
  phen_eval = TRUE,
  phens = NULL,
  meta_events = "meta_events",
  tbl_n = "^raw_*",
  verbose = FALSE,
  xtra_v = FALSE,
  chunk_v = FALSE,
  sfc = NULL,
  ...
)
}
\value{
A table describing data 'gaps'
}
\description{
The period between the time a logger/sensor is stopped, removed, or discommissioned, and when a logger/sensor beings recording data again, counts as missing data. However, for event-based data, where the frequency of recordings is temporarily erratic, identifying gaps is more tricky. This function sets some rules when identifying gaps in event-based time-series data. Furthermore, gaps can be manually declared in event metadata and integrated into the gap table.
}
\details{
There are a number of problems that arise in time-series data. One of the simplest issues are data gaps, that is, periods where no data were recorded by a logger, or an individual sensor attached to a logger. Moreover, data can be declared 'missing' if it is erraneous. This function helps users identify gaps in continuous and discontinuous data types. Below is an overview of this function.
\enumerate{
\item Derive gap table from raw data data summary information.
If the gap period is longer than the record interval in continuous data the gap is declared problematic, i.e., \code{problem_gap == TRUE}. For discontinuous data, where the gap period extends beyond the \code{gap_problem_thresh_s} the gap is \emph{'automatically'} declared to be problematic.
\item Integrate event metadata into the gap table (\code{meta_events} table).
If a 'problem gap' has been declared in event metadata then this information is transferrred into the gap table. Also we can add notes to gaps declared automatically in the step above. Note that any gap declared in the metadata requires appropriate metadata describing the gap. These include:
}
\itemize{
\item 'eid': a unique event identification number (integer),
\item 'qa': Quality assessment. Logical. Only evaluates \code{qa == TRUE},
\item 'station/location station': character string identifying the staiton name,
\item 'date_time': The date and time of the event (POSIX date time with format to match the raw data date_time),
\item 'event_type': Stanardised character string. For declaring a gap period use 'manual_gap'. For adding notes to a gap period use 'gap_notes' --- \emph{see notes on this below},
\item 'phen': If not provided this will default to 'logger', or else a specific phenomena name can be provided (standardised from the phens table). If this value is 'logger' we will assume the logger (with all sensors attached) was not functioning over the declared period,
\item 'table_name': A specific table the gap period is applied to. If no value is provided this will default to all raw data,
\item 'gap_problem_thresh_s': Can be provided to overwrite the \code{gap_problem_thresh_s} provided in the function arguments,
\item 'event_thresh_s': The event threshhold in seconds (_see start_dttm and end_dttm fields below),
\item 'problem_gap': Logical. If \code{TRUE} then the gap period declared is considered to be problematic. If \code{FALSE} a gap period will not be considered problematic. This is useful for discontinuous data. For example, if it was not raining during the period our logger had been stopped --- the 'automatically' identified gap, between the stop and start of the logger recording rainfall events (discontinuous record interval) --- the gap is not problematic. Therefore, we can declare \code{problem_gap} to be \code{FALSE} in the event metadata. Appropriate notes will be appended from the event metadata to the gap table in all cases.
\item 'notes': Character field for general note taking,
\item 'start_dttm': *Start date time of the manual gap period, and
\item 'end_dttm': *End date time of the manual gap period.
}
\itemize{
\item note that if start and end date times are not provided the 'date_time' field must be provided. Using adding and substracting the \code{event_thresh_s} to \code{date_time} will be used to generate \code{start_dttm} and \code{end_dttm} fields in this case.
}

The final gap table summarises gap data merged with the above event metadata.
}
\author{
Paul J. Gordijn
}
\keyword{data}
\keyword{data;}
\keyword{event}
\keyword{gaps;}
\keyword{missing}
\keyword{pipeline;}
