% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gw_f3_alpine_butter.r
\name{gw_alpine_butter}
\alias{gw_alpine_butter}
\title{Generate dummy dipper readings}
\usage{
gw_alpine_butter(
  input_dir = ".",
  wanted = NULL,
  unwanted = FALSE,
  prompt = FALSE,
  recurr = FALSE,
  ...
)
}
\arguments{
\item{input_dir}{The working directory where the standardised 'ipayipi'
water level files are located.}

\item{wanted}{Vector of strings listing files that should not be
included in the import.}

\item{unwanted}{Vector of strings listing files that should not be included
in the import.}

\item{prompt}{If TRUE, a command line prompt will be used to
enable selection of which files in the working directory require
drift correction.}

\item{recurr}{Should the function work recurrsively in the parent
directory.}
}
\description{
Generates dummy dipper readings using linear interpolation and
adds these to the 'log_retreive' table. These will be used for in subsequent
drift correction.
}
\author{
Paul J. Gordijn
}
