% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gw_d_baro_comp_batch.r
\name{gw_baro_comp_batch}
\alias{gw_baro_comp_batch}
\title{Batch compensate function}
\usage{
gw_baro_comp_batch(
  input_dir = NULL,
  out_csv = TRUE,
  overwrite_baro = TRUE,
  overwrite_comp = TRUE,
  start_t = NULL,
  end_t = NULL,
  join_tolerance = NULL,
  na_thresh = 5,
  overwrite_t_bt_level = FALSE,
  prompt = FALSE,
  wanted = NULL,
  unwanted = NULL,
  recurr = NULL
)
}
\arguments{
\item{input_dir}{Folder where to get R solonist data files.}

\item{out_csv}{Logical. Defaulted to TRUE.
If there is no baro file specified in the rdta_log this
will spit out a csv that can be edited.}

\item{overwrite_baro}{Logical. Defaulted to TRUE.
This option is described in the "gw_baro_comp" function.}

\item{overwrite_comp}{Logical. Defaulted to TRUE.
This option is described in the "gw_baro_comp" function.}

\item{start_t}{This option is described in the "gw_baro_comp" function.}

\item{end_t}{This option is described in the "gw_baro_comp" function.}

\item{join_tolerance}{Join tolerance between the barometric
and level data (for an overlap join).}

\item{na_thresh}{Feeder to the gw_baro_comp function.}

\item{overwrite_t_bt_level}{Need to populate detail here.}

\item{wanted}{Vector of strings listing files that should not be
included in the import.}

\item{unwanted}{Vector of strings listing files that should not be included
in the import.}

\item{recurr}{Logical. Argument fed to the prompted component of the
rsol_list function.}

\item{prompted}{Logical and defaults to FALSE.
If TRUE, a command line dialog will be used to
select data files for compensation.}
}
\value{
Screen print out of all the level logger data that has been
barometrically compensated in this function call.
}
\description{
This function feeds the gw_baro_comp function
files for processing. The function requires a pregenerated
rdta_log has been generated from the associated data pipeline.
}
\author{
Paul J. Gordijn
}
\keyword{barometric}
\keyword{compensation}
\keyword{level}
\keyword{water}
