% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gw_d_sub_clear_tabs.R
\name{gw_clear_tabs}
\alias{gw_clear_tabs}
\title{Clear R data solonist tables}
\usage{
gw_clear_tabs(
  input_dir = ".",
  clear_baro = FALSE,
  recurr = FALSE,
  baros = TRUE,
  ...
)
}
\arguments{
\item{input_dir}{Directory where the function will search for
R data solonist files.}

\item{clear_baro}{TRUE or FALSE. Defaults to FALSE. TRUE will clear
the barometric data in the "log_baro_log", "log_baro_data" and all
data in the "log_t" file. FALSE will clear the "log_t" table of all data.}

\item{recurr}{TRUE or FALSE. Defaults to FALSE. TRUE will enable a
recursive search for R data solonist files in the "input_dir".}

\item{baros}{TRUE or FALSE. Defaults to FALSE. If TRUE, barometric
data will be cleared as well.}
}
\value{
Screen note that the tables were cleared.
}
\description{
A function to clear all data tables in the standardised
R groundwater data files except the actual untransformed water level
data. Running in the pipeline will enable interactive selection of files
for which tables will be cleared.
}
\details{
Clears all logger data in the R data barologger table (and
associated baro logger tables) and/or the transformed logger data
table named 'log_t'. Enables a fresh start to data processing.
}
\author{
Paul J. Gordijn
}
\keyword{clear}
\keyword{data}
\keyword{delete}
\keyword{rows;}
