% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gw_f3_data_handler.r
\name{gw_data_handler}
\alias{gw_data_handler}
\title{Read data handling values}
\usage{
gw_data_handler(
  input_dir = ".",
  wanted = NULL,
  unwanted = FALSE,
  prompt = FALSE,
  recurr = FALSE,
  ...
)
}
\arguments{
\item{input_dir}{The working directory where the standardised 'ipayipi'
water level files are located.}

\item{wanted}{Vector of strings listing files that should not be
included in the import.}

\item{unwanted}{Vector of strings listing files that should not be included
in the import.}

\item{prompt}{If TRUE, a command line prompt will be used to
enable selection of which files in the working directory require
drift correction.}

\item{recurr}{Should the function work recurrsively in the parent
directory.}
}
\value{
Saves data with a table of manually declared outliers. List of station names of processed data.
}
\description{
Read in manually entered values which are used for correcting
anomalous data or removing outliers. The data handle information is
appended to the standardised data files and saved.
}
\author{
Paul J. Gordijn
}
\keyword{anomaly}
\keyword{cleaning,}
\keyword{data}
\keyword{outliers}
\keyword{outliers,}
\keyword{prior-informed}
\keyword{removal,}
