% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gw_f1_naartjie_clean.r
\name{gw_naartjie_clean}
\alias{gw_naartjie_clean}
\title{Interpolate and flag outliers plus missing values.}
\usage{
gw_naartjie_clean(
  input_dir = ".",
  recurr = FALSE,
  seg_na_t = 0.75,
  w_size = 21,
  mad_dev = 3,
  last_rule = FALSE,
  tighten = 0.65,
  wanted = NULL,
  unwanted = NULL,
  prompt = TRUE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{input_dir}{Working directory where the R data solonist
files are stored.}

\item{recurr}{Logical. Whether to search for R data files in
recursively.}

\item{seg_na_t}{Fractional tolerace of the amount of NA values
in a segment for linear interpolation of missing values.}

\item{w_size}{Window size for the hempel filter. Defaults to ten.}

\item{mad_dev}{Scalar factor of MAD. Higher values relax oulier detection.
Defaults to the standard of three.}

\item{last_rule}{Whether or not to reapply stored rules in the outlier
rule column. TRUE will apply old rules}

\item{tighten}{Scaling fraction between zero and one that sensitizes the
detection of outliers near the head and tail ends of segments. The fraction
is multiplied by the mad_dev factor.}

\item{wanted}{Vector of strings listing files that should not be
included in the import.}

\item{unwanted}{Vector of strings listing files that should not be included
in the import.}

\item{prompt}{Logical. Invoke prompt to select which data files in the
working directory are processed.}
}
\value{
A vector of xle file paths.
}
\description{
Function to detect outliers and remove them from the
data plus predict replacement values. Cleaning the
data like this is done sequentially through each
segment of the data as divided by logger download events and missing
data.
}
\author{
Paul J. Gordijn
}
\keyword{detection,}
\keyword{interpolation.}
\keyword{outlier}
\keyword{value}
