% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gw_e1_physiCally.R
\name{gw_physiCally}
\alias{gw_physiCally}
\title{Update manual calibration measurements}
\usage{
gw_physiCally(
  input_dir = ".",
  recurr = FALSE,
  overwrite_logs = TRUE,
  first_dip = TRUE,
  prompt = FALSE,
  wanted = NULL,
  unwanted = NULL,
  ...
)
}
\arguments{
\item{input_dir}{The working directory where the R data ground water
files are stored with the retrieve and datum logs.}

\item{recurr}{Should the function work recurrsively in the parent
directory.}

\item{overwrite_logs}{Defaults to TRUE which will overwrite all datum
and retrieve log information in the R data ground water files.}

\item{first_dip}{If this option is TRUE then the dipper readings taken
before the start of the barometrically corrected data are included if
they are within 12 hours of the start of the barometrically corrected data.}

\item{prompted}{If this is set to TRUE then an on screen prompt will be used
to interactively select files to process in this function.}
}
\value{
Screen print out of the R data files that have had their manual
calibration readings and datum atrributes updated.
}
\description{
This function updates the relevant boreholes with
all 'physical' or manual readings used to correct water level
height and drift. This is done by reading in the datum and
log retrieve logs. Syncing with the logs is only done in a
one-way direction, that is, if the R data ground water files are
updated internally -- this will not be transferred to the csv log.
}
\author{
Paul J. Gordijn
}
\keyword{altitude,}
\keyword{calibration,}
\keyword{casing}
\keyword{collar}
\keyword{datum,}
\keyword{dipper}
\keyword{height}
\keyword{height,}
\keyword{readings,}
