% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gw_c_sub_rdta_append.r
\name{gw_rdta_append}
\alias{gw_rdta_append}
\title{Append Rsol files}
\usage{
gw_rdta_append(
  new_file = NULL,
  old_file = NULL,
  InasRDS = TRUE,
  new_dir = ".",
  old_dir = ".",
  save_rdta = FALSE,
  RDS_name = NULL
)
}
\arguments{
\item{new_file}{R data solonist file to add to an existing file.}

\item{old_file}{Existing R data solonist file to which the new data
is appended to.}

\item{InasRDS}{Logical. Is the input data format RDS (file extension ".rds")}

\item{new_dir}{Directory in which the appended file is saved to.}

\item{old_dir}{Directory in which the existing or old R data
solonist file is located.}

\item{save_rdta}{Logical. Save the output file in RDS format?}

\item{RDS_name}{Output RDS file name.}
}
\value{
Appended data file.
}
\description{
This is a feeder function and not tested thoroughly
outside of the package pipeline. The function appends R solonist
data files and is used in the gw_xle_to_R() function.
}
\examples{
## This function is for use within the pipeline, especially the 'xle_to_R'
# function.
# Will add in an example when there is data included in the package. 
}
\author{
Paul J Gordijn
}
\keyword{append;}
\keyword{data}
\keyword{pipeline}
