% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gw_c_sub_rdta_list.R
\name{gw_rdta_list}
\alias{gw_rdta_list}
\title{Pipeline function - lists R data files}
\usage{
gw_rdta_list(input_dir = ".", recurr = FALSE)
}
\arguments{
\item{input_dir}{Direcytory for which to generate a list of .rds files.}

\item{recurr}{Logical. Should the function search recursively i.e., through
sub folders.}
}
\value{
Vector of R solonist data files in a directory.
}
\description{
Designed for use within the package pipeline. A simple function
that generates a list of RDS files in a directory. The function
discriminates "rdta_log.rds" and "transfer_log.rds" files - other files with
an ".rds" file extension will not be discriminated.
}
\examples{
## This function is a helper function not designed for isolated use.
## Get list of solonist data files which have been converted to
# 'rds' format.
input_dir <- "./solR_dta" # set directory where processed data is maintained.
gw_rdta_list(input_dir = input_dir, recurr = FALSE)
# Note we did not search recursively through folders by setting recurr
# to FALSE
}
\author{
Paul J Gordijn
}
\keyword{RDS}
\keyword{pipeline;}
