% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gw_c_sub_rdta_log_mk.r
\name{gw_rdta_log_mk}
\alias{gw_rdta_log_mk}
\title{Generate xle R data log file}
\usage{
gw_rdta_log_mk(input_dir = NULL, recurr = FALSE)
}
\arguments{
\item{input_dir}{Directory for which to generate a log.}

\item{recurr}{Logical. Should the log be generated recursively i.e.,
for sub-directories.}
}
\value{
A metadata table of R data files.
}
\description{
Creates a log file of all RDS data files in a directory.
Designed for use within the package pipeline.
}
\details{
The log this function creates extracts metadata from each
site where a logger is deployed. The metadata includes the site location
and borehole/piezometer name, serial numbers of the loggers that have been
deployed, the logger model, the start and end date times, the file name,
the location (directory) of the file, the barometer that should be used
to compensate the logger data (where relevant) and the class of the R
object, that is, either a 'baro_file' (barologger) or a 'level_file'
(water level and).
}
\examples{
## This function is a helper function not designed for isolated use.
## Get list of solonist data files which have been converted to
# 'rds' format.
input_dir <- "./solR_dta" # set directory where processed data is maintained.

## Generate log file
rdta_log_mk(input_dir = input_dir, recurr = FALSE)
}
\author{
Paul J. Gordijn
}
\keyword{log}
\keyword{pipeline,}
