% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gw_e_sub_read_datum_log.r
\name{gw_read_datum_log}
\alias{gw_read_datum_log}
\title{Read datum log csv}
\usage{
gw_read_datum_log(input_dir = NULL, recurr = FALSE, ...)
}
\arguments{
\item{input_dir}{Directory in which the R solonist data files and datum log
csv files are located.}

\item{recurr}{Should sub-directories be included in the search for the datum
log csv file - TRUE/FALSE.}
}
\description{
Function to read in a datum log csv file and checks variable
formats. The datum log contains the borehole metadata with survey heights
i.e. altitude, collar/casing height and depth to well etc. This data is
appended to the \strong{R} groundwater data files.
}
\author{
Paul J. Gordijn
}
\keyword{altitude}
\keyword{datum,}
