% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gw_b3_read_nomtab_csv.r
\name{gw_read_nomtab_csv}
\alias{gw_read_nomtab_csv}
\title{Read xle nomenclature csv}
\usage{
gw_read_nomtab_csv(wait_room = NULL, file = NULL, ...)
}
\arguments{
\item{wait_room}{Directory into which new data is copied into -- the start
of the data processing pipeline.}

\item{file}{name of the csv file to be read as the nomenclature table.}
}
\value{
Saves the read nomenclature table in 'rds' format. Prints the
nomenclature table.
}
\description{
Read in a user edited nomenclature file and save as a usable
nomtab file in the 'waiting room' folder. Part of the xle to R solonist
data pipeline.
}
\details{
\enumerate{
\item Read in csv file from specified folder.
\item Check for NA values in the table.
\item If the nomenclature table has no NA's:
- convert to tibble.
- save as rds in the 'waiting room' folder.
\item If there are NA values:
- stop the function and.
- produce warning, i.e. the user needs to fill in the NA values.
}
}
\examples{
## read in or import an updated 'nomtab' csv. 'nom' being nomenclature and
# 'tab' being table.
## set variables
wait_room <- "./temp" # directory (temporary) with xle data
csv_file <- "./nomtab_20210415_09h42.csv"

## read in the nomenclature table
gw_read_nomtab_csv(wait_room = wait_room, file = csv_file)
}
\author{
Paul J Gordijn
}
\keyword{Naming}
\keyword{convention,}
\keyword{nomenclature}
