% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gw_d_sub_read_rdta_log.r
\name{gw_read_rdta_log}
\alias{gw_read_rdta_log}
\title{Reads in R data log csv file}
\usage{
gw_read_rdta_log(input_dir = NULL, file = NULL, ...)
}
\arguments{
\item{input_dir}{Directory in which both the csv and R data files
are located.}

\item{file}{File name of the csv to be read.}
}
\value{
A data.table of the logger files in a directory.
}
\description{
This function just reads in the rdta_log csv file and converts
to rds if all barologger file names were specified using an editor of user
preferance.
}
\details{
The 'batch_baro' function uses this log file to search for
specified barologger files for the barometric compensation. When running
the 'batch_baro' function.
}
\author{
Paul J Gordijn
}
\keyword{data}
\keyword{log}
