% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gw_e_sub_read_retrieve_log.r
\name{gw_read_retrieve_log}
\alias{gw_read_retrieve_log}
\title{Function to read in the events table}
\usage{
gw_read_retrieve_log(input_dir = NULL, recurr = FALSE, ...)
}
\arguments{
\item{input_dir}{Directory in which the R solonist data files and retrieve
log csv files are located.}

\item{recurr}{Should sub-directories be included in the search for the datum
log csv file - TRUE/FALSE.}
}
\description{
This function reads the dipper and casing height readings
(retrieve log) which are used for drift correction of water levels.
This data is appended to the \strong{R} groundwater data files.
}
\author{
Paul J. Gordijn
}
\keyword{dipper}
\keyword{readings}
