% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gw_g_remove_site.R
\name{gw_remove_site}
\alias{gw_remove_site}
\title{Remove sites from processed data and logs}
\usage{
gw_remove_site(solr_room = NULL, recurr = FALSE, ...)
}
\arguments{
\item{solr_room}{The repository where the processed solonist data is located.}

\item{recurr}{Should the search for data include subrepositories i.e. be
recursive? Logical}
}
\description{
This function removes the site from the processed data folder
and records of this site in log files.
}
\details{
This function will help remove the data file and associated history
of a site from the processed data repository. The function may be useful for
simply removing a site from the repository or before rerunning the site
series_ function.
}
\author{
Paul J Gordijn
}
