% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gw_d_sub_rsol_list.R
\name{gw_rsol_list}
\alias{gw_rsol_list}
\title{Generate R data solonist file list}
\usage{
gw_rsol_list(
  input_dir = ".",
  recurr = FALSE,
  baros = FALSE,
  prompt = FALSE,
  ...
)
}
\arguments{
\item{input_dir}{Folder in which to search for R solonist data files.}

\item{recurr}{Should the function search recursively
i.e., thorugh sub-folders as well - TRUE/FALSE.}

\item{baros}{Should the function include barometric files
in the final list - TRUE/FALSE.}

\item{prompt}{Set to TRUE for interactive mode.}
}
\value{
A vector of selected R data solonist files.
}
\description{
Interactively or not, create a list (actually vector)
of R data solonist files in a specified directory.
}
\author{
Paul J Gordijn
}
\keyword{R}
\keyword{data}
\keyword{solonist}
\keyword{vector}
