% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gw_c1_site_series.R
\name{gw_site_series}
\alias{gw_site_series}
\title{Ammulgamate site data into time series}
\usage{
gw_site_series(solr_room = NULL, nomvet_room = NULL, ...)
}
\arguments{
\item{solr_room}{Directory, that is, folder in the package pipeline
that holds the R solonist data. Files will be appended to the R
solonist data files in this folder.}

\item{nomvet_room}{Folder in the data pipeline where the standardised
xle data is stored/archived.}
}
\value{
Vector of files transferred and/or appended to R data in the
designated repository.
}
\description{
This function explores the 'nomvet' data file log and
determines which xle files to amulgamate to generate master files
for each different site (or borehole).
}
\examples{
## Batch processing of groundwater or water level data
# Set the directory in where the r data files must be pushed/appended to.
solr_room <- "./solr_room"
# Set the directory where standardized xle files are stored.
nomvet_room <- "./nomvet_room"

# Batch run xle to R and append where appropriate.
gw_site_series(solr_room = solr_room, nomvet_room = nomvet_room)
}
\author{
Paul J Gordijn
}
\keyword{append,}
\keyword{pipeline}
