% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gw_f3_sub_ties.R
\name{gw_ties}
\alias{gw_ties}
\title{Generates dummy dipper points (ties).}
\usage{
gw_ties(
  file = NULL,
  tie_type = NULL,
  tie_datetime = NULL,
  dt_format = c("Ymd HMS", "Ymd IMSp", "ymd HMS", "ymd IMSp", "mdY HMS", "mdy IMSp",
    "dmY HMS", "dmy IMSp"),
  dt_tz = "Africa/Johannesburg",
  ...
)
}
\arguments{
\item{file}{Standardised 'ipayipi' water level file. The file must be
updated with the latest calibration measurements using
\code{ipayipi::gw_physiCally()}.}

\item{tie_type}{Must be supplied with one of the following options
(strings):
\enumerate{
\item "dummy" --- This option generates a dummy-dipper reading using
linear interpolation. Only one date-time value must be provided; this will
be the point at which the dummy dipper (calibration) measurement will be
created. In the linear interpolation the two closest surrounding
calibration measurements will be used. Any dummy-dipper reading generated
will be used in subsequent drift correction.
}}

\item{dt_format}{The function guesses the date-time format from a vector of
format types supplied to this argument. The 'guessing' is done via
\code{lubridate::parse_date_time()}. \code{lubridate::parse_date_time()} prioritizes
the 'guessing' of date-time formats in the order vector of formats
supplied. The default vector of date-time formats supplied should work
well for most logger outputs.}

\item{dt_tz}{Recognized time-zone (character string) of the data locale. The
default for the package is South African, i.e., "Africa/Johannesburg" which
is equivalent to "SAST".}

\item{recurr}{Setting this to TRUE will enable searching
for R data files in sub directories i.e., a recursive search.}

\item{prompted}{If TRUE, a command line prompt will be used to
enable selection of which files in the working directory require
drift correction.}
}
\value{
Table of dummy dipper readings, or tie points.
}
\description{
Generated tie points are used as artificial calibration
points (dipper readings) for drift correction.
}
\details{
Tie points can be used to correct for level shifts in time series
data. This function will attempt to automatically generate tie points using:
\enumerate{
\item Available calibration measurements,
\item The drift corrected water level, or
\item Something else ...
}
}
\keyword{correction;}
\keyword{dipper}
\keyword{drift}
\keyword{drift;}
\keyword{dummy}
\keyword{linear}
\keyword{points;}
\keyword{readings}
\keyword{tie}
