% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gw_plot_dygraph.r
\name{gw_vis_dy}
\alias{gw_vis_dy}
\title{Visualisation of ground water corrections with interactive widgets.}
\usage{
gw_vis_dy(
  file = NULL,
  vis_type = 1,
  dippr = FALSE,
  group = NULL,
  show_events = TRUE,
  conduct = FALSE,
  rain_data = NULL,
  ...
)
}
\arguments{
\item{file}{A level file object read into R.}

\item{vis_type}{Integer options 1, 2, 3, or 4. 1 - plots only the drift
corrected water level; 2 - plots only drift neutral time series with
outliers; 3 - combination of 1 and 2. 4 - plots barologger data alongside
drift neutral series; cannot include rainfall (\code{rain_data} must be set to
NULL, and conduct to FALSE).}

\item{dippr}{If TRUE this add dipper readings to the graphs.}

\item{group}{use a common string to syncronize the a-axes of separate graphs.}

\item{show_events}{Show when logger data was downloaded.}

\item{conduct}{Logical. If \code{TRUE} conductivity data (if present) will be
plotted on the secondary axis.}

\item{rain_data}{A list with three elements. If provided the
value will be passed to \code{ipayipi::rain_aggs()} --- see documentation
--- and the rainfall data  will be included on the graphs secondary axis.
If this argument is provided it will override display of elecroconductivity
on the secondary axis. The first element must contain the aggregation period
and the second a search keyword that will be used to select a rainfall
station.}
}
\description{
This function uses the html based `dygraphs' package to
generate interactive graphs that can be included in markdown documents.
There are several options provided for plotting transformed groundwater
data.
}
\details{
Plotting is optimized for evaluating transformed groundwater data
processed using \code{ipayipi}. The transformed data should be compared to the
\code{drift-neutral} data to visually assess the effectiveness of data
processing. Drift-netral data has not been drift corrected, and therefore,
indicates the effectiveness of barometric compensation.
}
\author{
Paul J. Gordijn
}
\keyword{Ground}
\keyword{graph,}
\keyword{visualisation,}
\keyword{water}
