% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gw_plot_dygraph2.R
\name{gw_vis_dy2}
\alias{gw_vis_dy2}
\title{Visualisation of multiple groundwater level series.}
\usage{
gw_vis_dy2(
  input_dir = ".",
  ri = NULL,
  wanted = NULL,
  unwanted = NULL,
  rain_data = NULL,
  recurr = TRUE,
  prompt = FALSE,
  pipe_house = NULL,
  ...
)
}
\arguments{
\item{input_dir}{Defaults to the current working directory. Only used if the \code{pipe_house} object is \code{NULL}.}

\item{ri}{Record interval string used to generate time sequence for merging data by date-time. Parsed to \code{dta_flat_pull0()}.}

\item{wanted}{Regex string used to filter stations when searching for groundwater data to plot.}

\item{unwanted}{Regex string to filter off unwanted stations when searching for groundwater stations from which to extract data.}

\item{rain_data}{A data.table with daily rainfall data. The rainfall column must be headed 'rain_tot', and the 'date-time' column must be named 'date_time'.}

\item{recurr}{Whether to search recursively within the \code{input_dir} or \code{pipe_house$ipip_room}. Logical. Defaults to FALSE.}

\item{prompt}{Logical. Use a prompt to select stations for plotting? Defaults to FALSE.}

\item{pipe_house}{If the \code{pipe_house} is provided stationse will be searched for in the \code{pipe_house$ipip_room} directory.}
}
\description{
Plots interactive graph for visualising trends in water level and daily rainfall patterns.
}
\details{
Plotting is optimized for evaluating transformed groundwater data processed using \code{ipayipi}. Daily rainfall data must have no missing values within the data set (i.e., leading and training NA values are fine, but NAs between the start and end date-times will be problematic). Plotting is done via \code{dygraphs}.
}
\author{
Paul J. Gordijn
}
\keyword{Ground}
\keyword{graph,}
\keyword{grouped}
\keyword{series}
\keyword{visualisation,}
\keyword{water}
