% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gw_b1_xle_import.r
\name{gw_xle_import}
\alias{gw_xle_import}
\title{xle import function}
\usage{
gw_xle_import(input_dir = NULL, wait_room = NULL, recurr = TRUE, ...)
}
\arguments{
\item{input_dir}{Directory where the files will be copied from.}

\item{wait_room}{Folder where ground water files can be temporarily stored
and prepared for processing.}

\item{recurr}{Should the function recurrsively search in subdirectories
for solonist files (.xle)?}
}
\value{
Screen printout of the xle files copied into the
designated directory.
}
\description{
Copy new data and prepare for preprocessing into the pipeline
}
\details{
The objectives of this function are:
\enumerate{
\item Copy raw file to a 'waiting room' folder where the raw data is
temporarily stored.
At his stage I have not enabled operation with .lev (old Solonist
files). These can be exported as xle files from within Solonist.
.lev files are a type of tab seperated value format.
For more on this see the sub function to list .xle files.
\item Standardise data encoding to UTF8 which is more usable.
- raw files that have moved through here must not be re-processed
to save time.
\item NB! File renaming is based on the logger metadata NOT the .xle
filename. Moreover, because the files are renamed with start and
end dataes if the same logger is deployed twice on a given day with
the same project ID and location it will be overwritten. It is
therefore important to change these logger settings before
deployment or after saving the file.
}
The name of the directory to be copied to must be customized to suite the
user. Under my regime I will use a data drive on my network.
The folder from which the raw files are being copied from is
interactively selected by the user.
}
\examples{
## import xle files into the groundwater level data pipeline. This is done
## by bringing these into a temporary folder or "waiting room".
## new_data_dir <- "./202104_sibhayi" # diectory from which to import data
## wait_room <- "./temp" # temporary directory into which to introduce data

## use the import function. Note 'recurr' was set to TRUE so that the
## function would import relevant xle files from subdirectories of the
## 'input_dir'.
gw_xle_import(input_dir = new_data_dir, wait_room = wait_room, recurr = TRUE)
}
\author{
Paul J Gordijn
}
\keyword{data}
\keyword{pipe}
\keyword{solonist,}
\keyword{xle,}
\keyword{xml,}
