% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gw_b_sub_xle_list.R
\name{gw_xle_list}
\alias{gw_xle_list}
\title{List xle files in a directory}
\usage{
gw_xle_list(input_dir = NULL, recurr = FALSE, ...)
}
\arguments{
\item{input_dir}{The folder, that is, directory in which to search
for solonist xle files.}

\item{recurr}{Should the search be recursive i.e., also include
subdirectories.}
}
\value{
a vector of xle file paths.
}
\description{
A function to search and list .xle files in a directory.
}
\details{
This is used as a helper function. The search is based on the
file extension .xle. Older solonist levelogger files with extension
.lev are excluded.
}
\examples{
# Generate a list of xle solonist files in the working directory
dir <- "." # define the directory
gw_xle_list(input_dir = dir, recurr = FALSE)
}
\author{
Paul J. Gordijn
}
\keyword{files}
\keyword{solonist}
\keyword{xle}
