% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gw_b_sub_xle_log.R
\name{gw_xle_log}
\alias{gw_xle_log}
\title{xle log}
\usage{
gw_xle_log(log_dir = NULL, xlefiles = NULL, place = NULL, ...)
}
\arguments{
\item{log_dir}{Path to the folder for which to generate a log. This must be
the folder which contains the xle files.}

\item{xlefiles}{A vector of file names passed to this function.}

\item{place}{Directory in which this log function operates.}
}
\value{
A data.table of all xle files in a directory with basic
logger and data metadata.
}
\description{
A function to generate an \strong{xle file log} of all present files
in a folder.
}
\details{
In addition to the log, that is, an inventory of xle files listed in the
directory the following metadata is extracted and appended in the log:
\itemize{
\item{"Location name"}{That is the site name where the borehole/piezometer
is deployed}
\item{"Borehole name"}{In the nomenclature under the hierarchy of the
'Location name'}
\item{"SN"}{That is Serial Number of the solonist or other logger}
\item{"Start"}{The date when the logger was programmed to begin collecting
data}
\item{"End"}{The date when the logger was programmed to stop collecting
data}
\item{"File name"}{The original name of the solonist xle file from which the
respective data was extracted.}
\item{"Place"}{The folder/directory from which the data was extracted/
originally located}
}
}
\examples{
# create an inventory of the xle files in the working directory
# first use internal gw_xle_list function to get vector of xle files in a
# directory.
xlefiles <- gw_xle_list(input_dir = dir, recurr = FALSE)
gw_xle_log(log_dir = ".", xlefiles = xlefiles, place = ".")
}
\author{
Paul J Gordijn
}
\keyword{log}
