% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gw_b4_xle_rename.r
\name{gw_xle_rename}
\alias{gw_xle_rename}
\title{A function to standardise xle file nomenclature}
\usage{
gw_xle_rename(wait_room, out_csv = TRUE, ...)
}
\arguments{
\item{wait_room}{Directory where the xle files are located with the
nomenclature table.}

\item{out_csv}{Logical. Defaults to TRUE which initiates a nomenclature csv
file output in the 'wait_room'.}
}
\value{
Screen printout of re-named files.
}
\description{
Using the xle nomenclature standarising table generated
from the xle_nomenclature function, this function renames xle files.
}
\details{
\enumerate{
\item Call the main nomenclature function, 'xle_nomenclature'
to make sure that nomenclature is ready.
\item Check for NA values in the table.
\item If there are NA values:
\itemize{
\item stop the function and
\item produce warning.
\item If the nomenclature table has no NA's:
\item rename files according to the nomenclature table.
}
}
}
\examples{
## rename and standardize xle file names in the waiting room directory
wait_room <- "./temp" # directory (temporary) with xle data

gw_xle_rename(wait_room = wait_room, out_csv = TRUE)

## Note that with 'out_csv' set to TRUE a csv will be generated with the
# current date-time of the running of this function.
# Once the csv has been updated the 'read_nomtab()' function can be used to
# read/import the csv.

gw_read_nomtab_csv(wait_room = wait_room, file = csv_file)
}
\author{
Paul J Gordijn
}
\keyword{files,}
\keyword{nomenclature}
\keyword{rename}
