% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gw_b5_xle_transfer.R
\name{gw_xle_transfer}
\alias{gw_xle_transfer}
\title{Archive xle files}
\usage{
gw_xle_transfer(wait_room = NULL, nomvet_room = NULL, ...)
}
\arguments{
\item{wait_room}{Folder path where the standardised xle files are stored.}

\item{nomvet_room}{Folder path where the standardised xle files will be
moved. The xle files should remained archived in this folder.}
}
\value{
List of dupplicate files if existant.
}
\description{
A function to transfer standardised xle files into an
archive folder.
}
\details{
Export renamed files into a central data folder called , 'nomvet'.
This function forms part of the xle to R data pipeline and will function
best where the suggested folders are created and utilised.
\enumerate{
\item Before cutting files to the 'nomvet' folder check for duplicates.
- generate a 'wait_log' table with file names and start and end
dates for the level logger data.
\item Update or create a 'nomvet_log' file that lists all stored .xle
files in the 'nomvet' folder.
\item Compare the 'wait_log' and 'nomvet_log' files and create:
- a list of potential duplicate files.
- list of non-duplicate files.
\item Copy the non-duplicate files across to the 'nomvet' folder.
\item Delete the non-duplicate files in the 'wait_room' folder.
\item If duplicates exist generate a message plus list/table of
detected duplicates.
}
NB! If your files are being transferred into the 'nomvet_room' you may
need to update the 'nomvet_log.rds' in the 'nomvet_room'. Do this using
gw_xle_log() and save the result in the nomvet folder. This should only be
necessary if you have removed xle files from the 'nomvet_room'.
}
\examples{
## Transfer the xle files that have updated nomenclature to a folder
# which contains xle data files used to generate continuous records.
## Set the directory (temporary) with xle data for which the nomenclature
# has been standardized.
wait_room <- "./temp"
## Set the directory to where the standardized xle files should be cut to
# (i.e. the files in the waiting room will be deleted.)
nomvet_room <- "./nomvet"

## Transfer the xle files in the temporary folder.
gw_xle_transfer(wait_room = wait_room, nomvet_room = nomvet_room)

## Note that a log file, 'nomvet_log.rds', will be updated as files are
# transferred.
}
\author{
Paul J Gordijn
}
\keyword{cut;}
\keyword{file;}
\keyword{move}
\keyword{xle;}
