% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_hampel.R
\name{hampel}
\alias{hampel}
\title{iPayipi Generalised Hampel filter}
\usage{
hampel(
  x = NULL,
  w = 21,
  d = 3,
  na_t = 0.25,
  align = "right",
  series_mad = NULL,
  ...
)
}
\arguments{
\item{x}{Values on which to perform the hampel filter evaluation.}

\item{w}{Total width of the filter window.}

\item{d}{The number of deviations from the central tendacy of the window used to detect outlier values.}

\item{align}{String matching one of the following options: 1) 'right' where the value under evaluation is on the right of a window; 2) 'left' where the value being evaluated is on the left of the window; and 3) 'centre' meaning that the value being evaluated is in the centre of the window. If the window length is even, then the 'centre' is biased to the left.}

\item{series_mad}{The median absolute deviation of the entire series being evaluated.}
}
\description{
Applies the non-linear (non-parametric) hampel filter to elements of a vector. The hampel is used to detect ourliers and produce estimates for imputation. If the focal value is an NA value or an 'outlier' the median is returned.
}
\section{General:}{
The hampel filter has been widely adopted for outlier detection in time-series data. The filter is non-parametric with overall robust behaviour (Pearson et al. 2016). Unlike the Tukey filter that has been based on standard deviation from the mean, the Hampel assesses median-absolute deviation (MAD) within the window of interest (ref). The filter is fully customisable and can be tued by adjusting the window size (\code{w}) and the amount of accceptable deviation from the MAD (\code{d}). If the focal value falls outside the range of the window median, plus minus MAD * d, it is considered an outliler and the median returned.
\subsection{\code{align}:}{

In addition to tuning \code{w} and \code{d} the focal value of the window can be set as \code{centre}, \code{left}, or \code{right}. See the parameter description for more.
}

\subsection{NA values:}{

The filter will remove NA values when determining the median and MAD, and continue to function with these up until the proportional-tolderance level of NA's in a window is surpassed, or there are no non-NA values in a window. The default proportion of NAs in a window, until where the function will operate is 0.25. When this level of NAs is reached the focal value in a window is not assessed in relation to the MAD and retured as is. When the focal value is NA the median is returned.
}
}

\section{Implosion sequences:}{
Owing to the failure of the Hampel where MAD approximates 0 when there is little to no deviation in data (i.e., an implosion sequence) (Pearson et al. 2016); this function temporariliy adjusts window data by adding minimal, normally-distributed random noise. This noise is approximated with a thousanth of the series MAD around a mean of zero. By adding this noise this customisation avoids the implosion sequence caveat of the Hampel filter.
}

\section{Advanced implementation:}{
This filter can be recursively deployed using \code{dt_clean()}. Moreover, deployement in a data stream can be segmented where appropriate, e.g., for data with level shifts that require correction post outlier detection---segmentation can be implemented in between levels.
}

\section{References:}{
Pearson, R.K., Y. Neuvo, J. Astola, and M. Gabbouj. 2016. Generalized Hampel filters. EURASIP Journal on Advances in Signal Processing, 2016(87). \href{https://asp-eurasipjournals.springeropen.com/articles/10.1186/s13634-016-0383-6}{doi:10.1186/s13634-016-0383-6}
}

\author{
Paul J Gordijn
}
\keyword{data;}
\keyword{detection;}
\keyword{imputation;}
\keyword{missing}
\keyword{outlier}
\keyword{patching;}
