% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_sub_hampel_f.R
\name{hampel_f}
\alias{hampel_f}
\title{Hampel outlier detection and filter}
\usage{
hampel_f(srs = NULL, w_width = 21, x_devs = 3, tighten = 1, ...)
}
\arguments{
\item{srs}{Input data in the form of a data table or data frame with one
heading as the "bt_level_m" and the other date-time series as POSIXT date
time (not time series).}

\item{w_width}{The desired width of the window used to evaluate a central
value.}

\item{x_devs}{The number of deviations from the central tendacy of the
window used to detect outlier values.}

\item{tighten}{multiplier of the \emph{x_devs} parameter only applied at
vector sides or begining and end. Designed to increase the sensitivity of
outlier detection at the vector sides.}
}
\value{
list containing tables including the interpolated values and outlier
flags.
}
\description{
Applies the non-linear hampel filter to detect ourliers and
produce estimates for interpolation of missing or outlier values.
The window size used to detect outliers and perform interpolation is
customizable. The window formation or relative position of the window
to a center value is adjustable.
}
\details{
This function is key to avoiding correcting drift based on outlier
values.
}
\author{
Paul J Gordijn
}
\keyword{detection;}
\keyword{interpolation;}
\keyword{missing}
\keyword{outlier}
\keyword{patching;}
\keyword{values;}
