% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_sub_dt_harvest_phens.r
\name{harvest_phens}
\alias{harvest_phens}
\title{Aggregate organisation for phenomena}
\usage{
harvest_phens(phen = NULL)
}
\description{
Function to organise a list or a data.table of
aggregation parameters.
}
