% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_c3_header_sts.r
\name{header_sts}
\alias{header_sts}
\title{Standardises ipayipi header information}
\usage{
header_sts(
  pipe_house = NULL,
  prompt = FALSE,
  recurr = TRUE,
  wanted = NULL,
  unwanted = NULL,
  file_ext_in = ".ipr",
  file_ext_out = ".iph",
  ...
)
}
\arguments{
\item{pipe_house}{List of pipeline directories. \strong{See}
\code{ipayipi::ipip_init()} \strong{for details}.}

\item{prompt}{Should the function use an interactive file selection function
otherwise all files are returned. TRUE or FALSE. Defaults to FALSE.}

\item{recurr}{Should the function search recursively into sub directories
for hobo rainfall csv export files? TRUE or FALSE.}

\item{wanted}{A strong containing keywords to use to filter which stations
are selected for processing. Multiple search kewords should be seperated
with a bar ('|'), and spaces avoided unless part of the keyword.}

\item{unwanted}{Similar to wanted, but keywords for filtering out unwanted
stations.}
}
\description{
Uses the nomenclature table, standardises the recording
interval string, start and end date times in the file header, and finally
the file name is copied from the standardised station name, that is, the
'standardised title'. The standardised station file header nomenclature
is retained in an ipayipi station file 'data_summary' table.
}
\details{
This function calls \code{ipayipi::nomenclature_sts()} which will take
take the user interactively through a process of standardising station
names and titles.
Station phenomena can only be checked once header nomenclature has been
standardised.

Note that there is no unstandardised 'uz' location. Whatever location
provided to in nomenclature table will be used in naming conventions.
The 'location' field can be provided edited during the
\code{ipayipi::nomenclature_sts()} and \code{ipayipi::read_nomtab_csv()}
functionality.
}
\author{
Paul J. Gordijn
}
\keyword{Cambell}
\keyword{Scientific;}
\keyword{automatic}
\keyword{batch}
\keyword{data;}
\keyword{file}
\keyword{header}
\keyword{information}
\keyword{meteorological}
\keyword{nomenclature;}
\keyword{process;}
\keyword{standardisation;}
\keyword{station;}
\keyword{weather}
