% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hobo_rain}
\alias{hobo_rain}
\title{Header harvest: Exported hobo rainfall files}
\format{
\subsection{'hobo_rain'}{

iPayipi data_setup list for imibing (\code{ipayipi::imbibe_raw_logger_dt()})
Hobo logger data from a Hobo rainfall logger pendant. The setup includes
harvseting of the loggers serial number using the search pattern
'LGN.S.N..'. The 'location' name is not specified. Only generic
information is provided for harvesting of logger file header information.
__See \code{ipayipi::imbibe_raw_logger_dt()} documentation for a describtion of
a \code{data_setup}.
}

\subsection{'file_read_meta'}{

iPayipi file_read_meta table used for setting defaults when attempting to
read/imbibe raw data files. This table is used in:
\code{ipayipi::imbibe_raw_batch()}. If the \code{file_ext} and/or \code{col_dlm}
arguments are \code{NULL} this table assists with reading in data in the
correct format.
}
}
\source{
Internal

Internal
}
\usage{
hobo_rain

hobo_rain
}
\description{
ipayipi data setup for Hobo logger files exported from a 'rainfall pendant'

File defaults used when imbibing raw logger data
}
\keyword{datasets}
