% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_sub_dt_hsf_param_eval.r
\name{hsf_param_eval}
\alias{hsf_param_eval}
\title{Harvest station data}
\usage{
hsf_param_eval(
  station_file = NULL,
  hsf_station = NULL,
  harvest_dir = ".",
  hsf_table = "raw",
  time_interval = NULL,
  phen_names = NULL,
  recurr = TRUE,
  harvest_station_ext = ".ipip",
  prompt = FALSE,
  single_out = TRUE,
  full_eval = FALSE,
  f_summary = NULL,
  ppsi = NULL,
  sf = NULL,
  ...
)
}
\arguments{
\item{station_file}{The path and name of the main station to
which data is being harvested.}

\item{hsf_station}{The path and name of the station, or a
keyword, to search for the station from which data will be
harvested. \emph{If \code{NULL} the \code{hsf_station} is set to
\code{station_file}, i.e., data will be harvested from the
\code{station_file}.}}

\item{hsf_table}{The name of the table from which to harvest
data. If the keyword "raw" is used data will extracted from
stations 'raw' data.}

\item{time_interval}{The desired time interval of the \code{output_dt}.
The full evaluation will extract phenomena from raw data tables
based accordingly.}

\item{phen_names}{Names of phenomena to extract from the specified
tables. If \code{NULL} all phenomena possible will be harvested. If the
desired \code{time_interval} is shorter than the time interval or
frequency of recordings of the data to be harvested, phenomena will
not be harvested.}

\item{recurr}{Logical. Whether to search recursively in directories
for the \code{hsf_station}. Parsed to \code{ipayipi::dta_list()}.}

\item{harvest_station_ext}{Parsed to \code{ipayipi::dta_list()}. Defaults
to ".ipip". Must include the period (".").}

\item{prompt}{Parsed to \code{ipayipi::dta_list()}. Set to \code{TRUE} to use
interactive harvest station selection.}

\item{single_out}{Forces through an interactive process the singling
out of a harvest station. Useful for example where \code{ipayipi::dta_list()}
returns from than one option.}

\item{full_eval}{Defaults to \code{FALSE}.}

\item{ppsi}{Summary pipe process table for the function. If provided this
data will be used to overwrite similar arguments provided in the function.
This must be provided for the full function evaluation.}

\item{hsf_dir}{The directory in which to search for the
\code{hsf_station}.}
}
\description{
Sets up a data harvest from a desired station file.
}
