% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/b_imbibe_raw_logger_dt.R
\name{imbibe_raw_logger_dt}
\alias{imbibe_raw_logger_dt}
\title{Imbibes logger data exports}
\usage{
imbibe_raw_logger_dt(
  pipe_house = NULL,
  file_path = NULL,
  file_ext = NULL,
  col_dlm = NULL,
  dt_format = c("Ymd HMOS", "Ymd HMS", "Ymd IMOSp", "Ymd IMSp", "ymd HMOS", "ymd HMS",
    "ymd IMOSp", "ymd IMSp", "mdY HMOS", "mdY HMS", "mdy HMOS", "mdy HMS", "mdy IMOSp",
    "mdy IMSp", "dmY HMOS", "dmY HMS", "dmy HMOS", "dmy HMS", "dmy IMOSp", "dmy IMSp"),
  dt_tz = "Africa/Johannesburg",
  record_interval_type = "continuous",
  max_rows = 1000,
  data_setup = NULL,
  remove_prompt = FALSE,
  logg_interfere_type = "on_site",
  verbose = FALSE,
  xtra_v = FALSE,
  ...
)
}
\arguments{
\item{file_path}{Path and name of file (excluding the file extension).}

\item{file_ext}{The file extension defaults of the raw logger data files. This can be left as \code{NULL} so 'all' files but those with extensions that cannot be imbibed (".ipr|.ipi|.iph|.xls|.rps|.rns|.ods|.doc").}
}
\value{
A list of class "ipayipi_raw_data" that contains a 'data_summary', 'phens' (phenomena), and 'raw_data' tables (data.table).
}
\description{
Function to read in 'flat' loggers files into R. A first step towards processing data in \code{ipayipi}.
}
\details{
\strong{See} \code{\link[=imbibe_raw_batch]{imbibe_raw_batch()}}.
}
\author{
Paul J. Gordijn
}
