% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_sub_a_imbibe_xml.r
\name{imbibe_xml}
\alias{imbibe_xml}
\title{Imbibe xle file}
\usage{
imbibe_xml(
  file_path = NULL,
  data_setup = NULL,
  dt_format = c("Ymd HMOS", "Ymd HMS", "Ymd IMOSp", "Ymd IMSp", "ymd HMOS", "ymd HMS",
    "ymd IMOSp", "ymd IMSp", "mdY HMOS", "mdY HMS", "mdy HMOS", "mdy HMS", "mdy IMOSp",
    "mdy IMSp", "dmY HMOS", "dmY HMS", "dmy IMOSp", "dmy IMSp"),
  dt_tz = "Africa/Johannesburg",
  ...
)
}
\arguments{
\item{file_path}{Recognized time-zone (character string) of the data locale. The default for the package is South African, i.e., "Africa/Johannesburg" which is equivalent to "SAST".}

\item{data_setup}{The solonist default data setup for 'xle' files is stored as package data: \code{ipayipi::solonist}. This is the only data setup that is in the early stages of testing for Solonist data loggers.}

\item{dt_format}{The function attempts to work out the date-time format from a vector of format types supplied to this argument. The testing is done via \code{lubridate::parse_date_time()}. \code{lubridate::parse_date_time()} prioritizes the tesing of date-time formats in the order vector of formats supplied. The default vector of date-time formats supplied should work well for most logger outputs.}

\item{dt_tz}{Recognized time-zone (character string) of the data locale. The default for the package is South African, i.e., "Africa/Johannesburg" which is equivalent to "SAST".}
}
\description{
Reads xml formatted solonist files into r for 'ipayipi'
}
\details{
Only continuous (not event type) Solonist data files have been tested with this function. There is a default \code{data_setup} that is used for reading Solonist xle files. The \code{XML} package is used to read xml formatted Solonist 'xle' files.
}
\author{
Paul J Gordijn
}
\keyword{#return}
\keyword{data;}
\keyword{format;}
\keyword{imbibe}
\keyword{solonist}
\keyword{xml}
