% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/x_ipayipi2csv.R
\name{ipayipi2csv}
\alias{ipayipi2csv}
\title{ipayipi station to 'csv'}
\usage{
ipayipi2csv(
  input_dir = ".",
  pipe_house = NULL,
  output_dir = NULL,
  file_ext = ".ipip",
  wanted_tabs = NULL,
  prompt = FALSE,
  wanted = NULL,
  unwanted = NULL,
  recurr = TRUE,
  verbose = FALSE,
  xtra_v = FALSE,
  ...
)
}
\arguments{
\item{input_dir}{The directory where standardised ipayipi data files to be are stored.}

\item{pipe_house}{List of pipeline directories. \strong{See} \code{\link[=ipip_house]{ipip_house()}} \strong{for details}.}

\item{output_dir}{The outpur directory where the standardised ipayipi data object will be saved/exported to.}

\item{file_ext}{Character string of the file extension of the input data files. E.g., ".csv" for hobo rainfall file exports. This input character string should contain the period as in the previous sentence.}

\item{wanted_tabs}{Data items/tables that should be exported to csvs.  '\%ilike\%': ergo seperate multiple search phrases by '|'.}

\item{prompt}{Set to TRUE for interactive mode. Note this will not work if embedded in a parallel processing instance.}

\item{wanted}{Regex string of files to select for listing. Seperate search tags by using the bar character '|'.}

\item{unwanted}{Regex string of files to filter out the listing. Seperate search tags by using the bar character '|'.
in the import.}

\item{recurr}{Should the function search recursively i.e., thorugh sub-folders as well - \code{TRUE}/\code{FALSE.}}

\item{verbose}{Logical. Print some details and progress of function progress?}

\item{xtra_v}{Logical. Should some 'x'tra messaging be done? Use to help diagnose problems, and for guidance.}
}
\value{
A list of station files which had tables exported to csv files.
}
\description{
Exports standardised ipayipi data objects to csv format.
}
\details{
This function will only export tables within a list item, not a list of tables within the list.
}
\author{
Paul J. Gordijn
}
\keyword{R}
\keyword{batch}
\keyword{csv}
\keyword{data;}
\keyword{export}
\keyword{export.}
\keyword{format;}
\keyword{save}
