% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_ipayipi_agg.R
\name{ipayipi_agg}
\alias{ipayipi_agg}
\title{Summarize rainfall data by multiple time periods}
\usage{
ipayipi_agg(
  input_file = NULL,
  output_pref = "agg_",
  aggs = c("day", "month"),
  ignore_nas = FALSE,
  raw_tbl = NULL,
  raw_tbl_preffix = "raw_",
  ...
)
}
\arguments{
\item{input_file}{Output from \code{clean_rain_hobo}}

\item{output_pref}{A string for the file name prefix. The time periods given
in the \code{aggs} argument will form the suffex. The default preffix used in
the data pipeline is 'agg_'. The 'agg_' preffix must be used for further
processing.}

\item{aggs}{vector of strings giving the time periods to use for aggregating
the data.}

\item{ignore_nas}{Choose whether to ignore gaps and other \code{NA} values when
aggregating the data by sum.}

\item{csv_out}{If TRUE the function will write a csv to the working
for each set of aggregated daunitsta.}
}
\value{
Returns the "SAEON_rain_data" object with appended rainfall data
aggregations.
}
\description{
Reads rainfall data from the \code{clean_rain_hobo()} function
then aggregates the data by specified time periods.
}
\details{
Depreciated function.
The function uses \code{data.table} to summarize the data by the time
periods of interest (generated with the help of \code{lubridate}). The data
processing and writing to csv can be done using parallel processing by
setting the cores argument to greater than one (but not exceeding the
number of cores on the system).
The data will be saved as a 'data.table' in the \strong{ipayipi} rainfall object
with the given \code{output_pref} --- if this value is changed the processing
of data within the pipeline will be obscured.
}
\author{
Paul J Gordijn
}
\keyword{Internal;}
\keyword{aggregation}
\keyword{bucket}
\keyword{data}
\keyword{gauge,}
\keyword{hoboware,}
\keyword{rain}
\keyword{tipping}
