% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_sub_ipayipi_data_log.R
\name{ipayipi_data_log}
\alias{ipayipi_data_log}
\title{Generate a log of data files within a directory.}
\usage{
ipayipi_data_log(log_dir = NULL, file_ext = ".ipi", ...)
}
\arguments{
\item{hobo_in}{Converted hobo rainfall files as produced by
\code{rain_hobo_conversion()}.}

\item{nomvet_room}{The directory where the standardised hobo rainfall files
are archived.}
}
\value{
A table describing station file data within a directory, and the file extension of the station files.
}
\description{
Once the hobo rainfall file exports have been standardised,
and have their nomenclature checked, this function saves the data from
each import into the 'nomvet_room'.
}
\details{
Aim: Archive each checked hobo rainfall file in a secure
directory, that is, the '\strong{nomvet_room}', which is part of the 'ipayipi'
data pipeline. Once archived the data will be prepared for further
processing. Here are some of the specifics:
\enumerate{
\item Generate inventory of standardised files that have been processed by
\code{rain_hobo_conversion()} --- this object is 'fed' to the function via
the \code{hobo_in} argument.
\item Create (if none) or produce an inventory of files in the 'nomvet_room'.
\item Compare the inventories to identify which new files should be
transferred into the 'nomvet_room'.
\item Transfer files---saving these in RDS format in the 'nomvet_room'.
}
}
\author{
Paul J. Gordijn
}
\keyword{archive}
\keyword{data}
\keyword{data;}
\keyword{pipeline;}
\keyword{save}
