% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_sub_ipayipi_flat_dta.r
\name{ipayipi_flat_dta}
\alias{ipayipi_flat_dta}
\title{Pulls multiple station data into a single file}
\usage{
ipayipi_flat_dta(
  input_dir = NULL,
  file_ext = ".ipi",
  tab_name = NULL,
  phen_name = NULL,
  output_dir = NULL,
  prompt = FALSE,
  wanted = NULL,
  unwanted = NULL,
  out_csv = FALSE,
  out_csv_preffix = "",
  recurr = FALSE,
  ...
)
}
\arguments{
\item{input_dir}{The directory in which to search for stations form which
to extract weather data.}

\item{file_ext}{The extension of the stations from where data will be
extracted.}

\item{tab_name}{The name of the table in a station file (list object) from
which to extract data.}

\item{phen_name}{The field/column/name of the phenomena which to join into
a single data table.}

\item{output_dir}{The output directory where an output csv file is saved.}

\item{prompt}{Logical. If \code{TRUE} a prompt will be called so that the user
can interactively select station files.}

\item{wanted}{A strong containing keywords to use to filter which stations
are selected for processing. Multiple search kewords should be seperated
with a bar ('|'), and spaces avoided unless part of the keyword.}

\item{unwanted}{Similar to wanted, but keywords for filtering out unwanted
stations.}

\item{out_csv}{Logical. If TRUE a csv file is exported to the output
directory.}

\item{out_csv_preffix}{Preffix for the output csv file. The phenomena name
and then time interval by which the data are summarised are used as a
suffix.}
}
\value{
A list containing 1) the summarised data in a single data.table, 2)
a character string representing the time interval by which the data has
been summarised, 3) the list of stations used for the summary, 4) the name
of the table and the name of the field for which data was summarised.
}
\description{
Summarises data by the longest time interval for a particular
phenomena wihtin the specified data table in an \code{ipayipi} station object.
The function can be used to summarise time-series data from a number of
stations and the output saved to a csv file.
}
\author{
Paul J. Gordijn
}
\keyword{data}
\keyword{data;}
\keyword{pipeline;}
\keyword{summarise}
\keyword{time-series}
