% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_sub_ipip_init.r
\name{ipip_init}
\alias{ipip_init}
\title{Setup an \code{ipayipi} data pipeline directory system}
\usage{
ipip_init(
  work_dir = ".",
  source_dir = NULL,
  wait_room = NULL,
  nomvet_room = NULL,
  ipip_room = NULL,
  raw_room = NULL,
  ...
)
}
\arguments{
\item{work_dir}{The folder, that is, directory in which to create the
pipeline directories.}
}
\value{
List of 'room' directories.
}
\description{
Creates the folders for the ipayipi data processing pipeline.
}
\details{
This function automates the creation of four folders/directories
that are requried for bulk processing of files in the ipayipi groundwater
pipeline structure. The directories are:
\itemize{
\item \strong{wait_room} --- temporary directory where raw data imports are
standardised.
\item \strong{nomvet_room} --- for housing standardized native solonist data files.
\item \strong{ipip_room} --- transforming and processing standardized native data.
\item \strong{raw_room} --- for archiving 'raw' data files.
\item \strong{source_dir} --- directory in which new raw data files are 'sourced'.
\emph{NB}! Existing directories (and their contents) with similar names are not
deleted/removed by this function.
}
}
\examples{
# Inititate pipeline
dir <- "." # define the working directory
pipe_init(work_dir = dir)
}
\author{
Paul J. Gordijn
}
\keyword{creation,}
\keyword{directory}
\keyword{folder}
\keyword{initiate}
\keyword{pipeline,}
\keyword{structure.}
