% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_sub_dt_join_param_eval.r
\name{join_param_eval}
\alias{join_param_eval}
\title{Evaluate and prepare tables for joins}
\usage{
join_param_eval(
  join_type = NULL,
  x_table = NULL,
  y_table = NULL,
  y_phen_names = NULL,
  y_table_key = NULL,
  x_table_key = NULL,
  full_eval = FALSE,
  f_summary = NULL,
  ppsij = NULL,
  sf = sf,
  ...
)
}
\description{
Flexible and fast joining of \code{ipayipi} data.
}
\author{
Paul J. Gordijn
}
