% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_a_logger_data_import_batch.r
\name{logger_data_import_batch}
\alias{logger_data_import_batch}
\title{Import batches of logger data into the ipayipi 'waiting room'.}
\usage{
logger_data_import_batch(
  source_dir = ".",
  wait_room = NULL,
  prompt = FALSE,
  recurr = FALSE,
  wanted = NULL,
  unwanted = NULL,
  file_ext = ".dat",
  ...
)
}
\arguments{
\item{source_dir}{The path directory from where Cambell logger data files
are located.}

\item{wait_room}{Path to the 'wait_room' directory.}

\item{prompt}{Should the function use an interactive file selection function
otherwise all files are returned. TRUE or FALSE.}

\item{recurr}{Should the function search recursively into sub directories
for hobo rainfall csv export files? TRUE or FALSE.}

\item{wanted}{Vector of strings listing files that should not be
included in the import.}

\item{unwanted}{Vector of strings listing files that should not be included
in the import.}

\item{file_ext}{The file extension defaults to ".dat". Other file types could
be incorporatted if required.}
}
\description{
Locates, copies, transforms in R format, then pastes logger
data files into the 'wait_room'.
}
\details{
Copies logger data files into a directory where further data
standardisation will take place in the 'ipayipi' data pipeline.
}
\author{
Paul J. Gordijn
}
\keyword{automatic}
\keyword{batch}
\keyword{data}
\keyword{data;}
\keyword{files;}
\keyword{hydrological}
\keyword{import}
\keyword{logger}
\keyword{meteorological}
\keyword{process;}
\keyword{station;}
\keyword{weather}
