% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_b_cl_read.r
\name{met_cl_read}
\alias{met_cl_read}
\title{Read cambell logger data exports}
\usage{
met_cl_read(
  file_path = NULL,
  file_title_row = 1,
  phen_name_row = 2,
  phen_units_row = 3,
  phen_measure = 4,
  phen_first_col = 3,
  data_row_start = 5,
  dt_col = 1,
  dt_format = "\%y-\%m-\%d \%H:\%M:\%S",
  dt_tz = "Africa/Johannesburg",
  id_col = 2,
  file_ext = ".dat",
  file_sep_dlm = NULL,
  row_names = FALSE,
  ...
)
}
\description{
Function to read in Cambell logger '.dat' files in R.
}
\author{
Paul J. Gordijn
\enumerate{
\item Read in file.
\item Standardise units and generate file metadata in preparation for further
data checks and station standardisation.
}
}
