% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_g0_meta_read.R
\name{meta_read}
\alias{meta_read}
\title{Open and prepare event metadata database.}
\usage{
meta_read(
  input_dir = ".",
  meta_file = NULL,
  col_dlm = ",",
  col_names = NULL,
  col_types = NULL,
  dt_format = c("Ymd HMS", "Ymd IMSp", "ymd HMS", "ymd IMSp", "mdY HMS", "mdy IMSp",
    "dmY HMS", "dmy IMSp", "Ymd HMOS", "Ymd IMOSp", "ymd HMOS", "ymd IMOSp", "mdY HMOS",
    "mdy IMOSp", "dmY HMOS", "dmy IMOSp"),
  dt_tz = "Africa/Johannesburg",
  d_format = "ymd",
  T_format = "hms",
  sheet = NULL,
  output_dir = ".",
  output_name = NULL,
  event_thresh = 600,
  ...
)
}
\arguments{
\item{input_dir}{Directory in which to search for the metadata database.}

\item{meta_file}{File name of the standardised rainfall metadata database or a meta_file table object.}

\item{col_dlm}{The column delimter of the 'flat' logger file which is fed to \code{data.table::fread()} or \code{base::read.csv()}.`}

\item{col_names}{Names of the columns (as a vector of strings. If provided then this is used to check against the data being read. If a column is missing/misspelt then data will not be read.}

\item{col_types}{Inspired by \code{googlesheets4} --- string of data indicating the data tye of each column. The following formats are supported:
\itemize{
\item l: Logical,
\item i: Integer,
\item d or n: Numeric, or "double",
\item D*: Date (parsed as a date --- format as d_format,
\item t*: Time of day (time is parsed as a lubridate period --- format
as per \code{t_format}),
\item T: POSIXct date-time (format as per \code{dt_format}), &
\item c: Character.
\item f: Factor.
}
\itemize{
\item not yet implemented.
}}

\item{dt_format}{The function attempts to work out the date-time format from a vector of format types supplied to this argument. The testing is done via \code{lubridate::parse_date_time()}. \code{lubridate::parse_date_time()} prioritizes the tesing of date-time formats in the order vector of formats supplied. The default vector of date-time formats supplied should work well for most logger outputs.}

\item{dt_tz}{Recognized time-zone (character string) of the data locale. The default for the package is South African, i.e., "Africa/Johannesburg" which is equivalent to "SAST".}

\item{d_format}{String indicating the order of the year month and date to be parsed by \code{lubridate}. Not yet implemented.}

\item{T_format}{String indicating time format, e.g., hms --- see lubridate. Not yet implemented.}

\item{output_dir}{Directory where meta file should be saved.}

\item{output_name}{Name of the file for saving as an RDS file. The default extension is ".rmds".}

\item{event_thresh}{The event threshold provided in seconds. Argument used to prepare event database for gap and pseudo event evaluation by populating NA values in the 'event_thresh_s' field by the argument value provided here. The defult is ten minutes, that is 10 * 60 seconds. \emph{see} \code{ipayipi::gap_eval()}.}
}
\value{
Standardised data table of events.
}
\description{
Reads standardised metadata and appends records the matching station.
}
\details{
Reads in an events database or sheet in 'csv' format. Checks that column names have been standardised. Transforms the date-time columns to a standardised format --- ** this format and timezone must match that used by the data pipeline **.

The metadata table rows must have an assicatiated time stamp with header "date_time". Additional date-time columns are permitted e.g., start and end date_times, as, 'start_dttm' and 'end_dttm', respectively.
}
\author{
Paul J. Gordijn
}
\keyword{data}
\keyword{data;}
\keyword{field}
\keyword{logger}
\keyword{metadata;}
\keyword{notes}
\keyword{pipeline;}
\keyword{processing;}
\keyword{supplementary}
