% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_h1_sub_dt_join_mhlanga.r
\name{mhlanga}
\alias{mhlanga}
\title{Join data sets together}
\usage{
mhlanga(
  x_tbl = NULL,
  y_tbl = NULL,
  join = "full_join",
  x_key = "date_time",
  y_key = "date_time",
  fuzzy = NULL,
  inq = NULL,
  y_phen_names = NULL,
  nomatch = NA,
  mult = "all",
  roll = FALSE,
  rollends = FALSE,
  allow.cartesian = FALSE,
  time_seq = TRUE,
  ...
)
}
\arguments{
\item{x_tbl}{The left, that is, 'x' table.}

\item{y_tbl}{The right, that is, 'y' table.}

\item{join}{One of the following join types:
\enumerate{
\item \emph{full_join}. Joins the 'x' and 'y' tables using keyed column. Retains all the matching and unmatching rows of 'x' and 'y' tables.
\item \emph{left_join}. Adds the matching rows of 'x' to 'y' by keyed columns.
\item \emph{right_join}. Adds the matching rows of 'y' to 'x' by keyed columns.
\item \emph{inner_join}. Retains only rows of 'x' and 'y' that match by keyed columns.
\strong{Keys}
Keys are the columns of the 'x' (left) and 'y' (right) tables used for finding matches for a join. Because \code{ipayipi} is based on date-time data the default keys are assumed to be "date_time" if not provided. Unlike conventional joins where rows are only matched if there is an exact match, for 'fuzzy' or 'non-equi' joins are flexible. When using a 'fuzzy' join an inequality sign must be provided in the key list. \emph{See below}_.
}}

\item{x_key}{Name(s) of the column(s) to key by the 'x' table. Vector of length one or two for conventional or fuzzy or non-equi joins, respectiely.}

\item{y_key}{Same as above but for the 'y' or right table.}

\item{fuzzy}{Numeric duration of time used to generate first and second 'x' and 'y' key columns. If provided fuzzy columns will override the \code{x_key} and \code{y_key} columns with auto generated default names. The default names for these are xd1 and xd2, and yd1 and yd2, for 'x' and 'y' tables, respectively. Columns are d1 and d2 columns are calculated around the "date_time" (POSIXct) columns specified in the key arguments. By default this will be "date_time" for both 'x' and 'y' tables. Provided as a single value or vector of two.
For 'fuzzy' joins where the two keys are provided for each 'x' and 'y' table the join syntax is compiled like this:
'1st x key' 'x inequality' '1st y key', '2nd y key' 'x inequality' '2nd x key'.}

\item{inq}{Inequality signs, vector of two (i.e., ">", ">=", "<", "<=") for fuzzy (non-equi) joins. Defaults to NULL.}

\item{y_phen_names}{Names of phenomena (columns) to retain post the join.
\emph{Custom data.table arguments}
\code{data.table} has great documentation for these arguments.}

\item{nomatch}{In outer joins the rows for which no match was found can be retained by setting this argument to NA. If this is set to NULL 'no match' rows are not retained. By default an inner_join has this argument set to NULL. For other joins this is set to NA. Setting the value here will override the default.}

\item{mult}{Where multiple rows of the right table match the left, this argument controls how more than one match is handled. If specified as "all" (default), all matching rows are returned. Otherwise the "first" or "last" matching rows will be returned.}

\item{roll}{Can't get better than the \code{data.table} documentation here.}

\item{rollends}{Can't get better than the \code{data.table} documentation here.}

\item{allow.cartesian}{Can't get better than the \code{data.table} documentation here.}

\item{time_seq}{If data is continuous this can be set to TRUE which will trigger time-series sensitive joining of data. Time-series sensitive joining on ensures the date-time series is continuous, i.e., the record intervalis kept constant, and unique (no duplicate time stamps).}

\item{input_dt}{}

\item{output_dt}{}

\item{f_summary}{A summary table of function parameters. This
summary table is stored in the station file object. If there
is a summary table for the function, differences between the
\code{f_params} and table are checked and the 'f_summary' table
is updated---triggering a re-calculation of the whole time series instead
of for a slice of the series.}
}
\value{
A list containing the processed data sets 'dts_dt'.
}
\description{
.
}
\author{
Paul J. Gordijn
}
