% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_sub_message.r
\name{msg}
\alias{msg}
\title{Internal function for returning message or not}
\usage{
msg(x = NULL, verbose = TRUE)
}
\arguments{
\item{x}{A message string.}

\item{verbose}{If TRUE the message will print, and \emph{vice versa} for FALSE.}
}
\description{
An alias to return a message.
}
\author{
Paul J. Gordijn
}
