% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_c1_nomenclature_sts.r
\name{nomenclature_sts}
\alias{nomenclature_sts}
\title{Check ipayipi data file nomenclature}
\usage{
nomenclature_sts(
  wait_room = NA,
  check_nomenclature = TRUE,
  csv_out = TRUE,
  file_ext = "ipr",
  ...
)
}
\arguments{
\item{wait_room}{The working directory where imported meteorological data
files are stored.}

\item{check_nomenclature}{Should the file naming be check. This helps ensure
that files are named consistently and that data from different stations are
not appended erraneously.}

\item{file_ext}{The extension of the file for which the nomenclature is being
assessed.}

\item{out_csv}{Logical. If TRUE a csv file is made in the working
directory if there are files with unrecognised nomenclature.}
}
\value{
returns a csv nomenclature file. Screen output notes whether
the nomenclature table is complete.
}
\description{
A function to check standardisation of meteorological data
objects to be run before introducing data into the pipeline.
}
\author{
Paul J. Gordijn
}
\keyword{nomenclature}
