% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_sub_open_sf_con.r
\name{open_sf_con}
\alias{open_sf_con}
\title{Opens a connection to an ipayipi station file.}
\usage{
open_sf_con(pipe_house = NULL, station_file = NULL, tmp = FALSE, ...)
}
\arguments{
\item{pipe_house}{Optional. Used to generate filepath to station file name. List of pipeline directories. \strong{See} \code{ipayipi::ipip_init()} \strong{for details}.}

\item{station_file}{Path (with file name and extension) to 'ipip' station file.}

\item{tmp}{Logical that defaults to FALSE. If TRUE then the station file is ignored. If TRUE then the temporary station is updated with information from the station file.}
}
\value{
Station file connection file paths. To be used in the pipeline.
}
\description{
Package subroutine for opening a station file for work.
}
\details{
In order to speed processing of data 'ipayipi' decompresses a station file list of tables/values into a 'hidden' temporary folder in the same working directory as a particular station file. Running this function will extract station file tables/values to a 'hidden' folder in the working directory. This function searches for_or_ creates a 'hidden' directory with station data and lists the file paths for each item in the directory. This list of directories for station data is termed a 'station file connection'. Functions using this connection will only update the station file on disk ('non-hidden' file) when updated information is written using \code{ipayipi::write_station()}.
}
\author{
Paul J Gordijn
}
\keyword{add}
\keyword{edit}
\keyword{station;}
\keyword{to}
\keyword{write}
