% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_sub_open_sf_con2.r
\name{open_sf_con2}
\alias{open_sf_con2}
\title{Establishes a connection to an ipayipi station file.}
\usage{
open_sf_con2(
  pipe_house = NULL,
  station_file = NULL,
  tmp = FALSE,
  chunk_i = NULL,
  i_zeros = 5,
  rit = NULL,
  ri = NULL,
  verbose = FALSE,
  tv = NULL,
  cores = getOption("mc.cores", 2L),
  xtra_v = FALSE,
  ...
)
}
\arguments{
\item{pipe_house}{Optional. Used to generate filepath to station file name. List of pipeline directories. \strong{See} \code{ipayipi::ipip_house()} \strong{for details}.}

\item{station_file}{Path (with file name and extension) to 'ipip' station file.}

\item{tmp}{Logical that defaults to FALSE..If \code{FALSE} then the temporary station is updated with new tables from the station file (old tables in the temporary station file are not overwritten). If \code{TRUE} then the temporary station file is not updated with extra/new tables in station file is ignored.}

\item{chunk_i}{Time interval, as a string, used to chunk time series data. Only tables selected using this regex, "^raw_\emph{|^dt_}", are chunked. A chunk index is creater to facilitate rapid read/writing of data.}

\item{i_zeros}{Used to name chunked files; the number of leading zeros to include before a chunk number index.}

\item{verbose}{Logical. Whether or not to report messages and progress.}

\item{tv}{String vector of the names of values or tables to read from the station file. The strings provided here are used to filter items in the \code{sfc} object. If left NULL (default) all tables will be opened into the 'tmp' folder. Table names are matched with data.tables \%ilike\% operator.}

\item{cores}{Number of cores to farm processing off to. This assists with increasing the speed of read-write capabilities dealing with chunks.}

\item{buff_period}{Time periods represented by a character string used to buffer the min and max date-times around index creation. Should ideally be at least 10 years. If new data extends before or beyond the index dates then data will be rechunked, and a new index generated.}
}
\value{
Station file connection file paths. To be used in the pipeline.
}
\description{
Package subroutine for opening a station file for work. Chuncks station file data with 'date_time' column and generates index. Reduces data read into memory when processing by accessing chunks filtered by date and time.
}
\details{
This function will not overwrite temporary station tables/data. If there is no chunked versino of the station file available, the station file ("ipip" file) will be chunked, and the connection established.
}
\author{
Paul J Gordijn
}
\keyword{add}
\keyword{edit}
\keyword{station;}
\keyword{to}
\keyword{write}
