% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_sub_dt_process_p_step.r
\name{p_step}
\alias{p_step}
\title{Unpack data processing step}
\usage{
p_step(
  dt_n = 1,
  dtp_n = 1,
  f = NULL,
  f_params = NULL,
  input_dt = NA,
  output_dt = NA,
  output_dt_preffix = "dt_",
  output_dt_suffix = NULL,
  time_interval = NA,
  ...
)
}
\arguments{
\item{dtp_n}{The numeric (integer) order of the processing step.}

\item{f}{A string to be evaluated in the data.table environment.}

\item{input_dt}{The name of the input data table (character string).}

\item{output_dt}{The name of the output data table (character string).}
}
\description{
Internal function used to unpack a data pipeline processing
step. Essentially converts parameters into a list.
}
\author{
Paul J. Gordijn
}
