% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_sub_misc_padr.r
\name{padr}
\alias{padr}
\title{Message generic}
\usage{
padr(
  core_message = NULL,
  wdth = NULL,
  pad_char = c("-"),
  pad_extras = c("|", "", "", "|"),
  force_extras = FALSE,
  justf = c(0, 0)
)
}
\arguments{
\item{core_message}{String of variable length. The message to be padded by
special characters and
justified.}

\item{wdth}{Integer or NULL. The width of the message line to be printed
in characters. If NULL the width is set to the console width. The default
console width in R is 80 characters.}

\item{pad_char}{String. The character to be used to pad a message.
Default "-".}

\item{pad_extras}{Vector of four characters placed around padding on the
left and right padding of the core message. Characters given in order
from right to left: 1. Left preffix, 2. left suffix, 3. left preffix,
4. right suffix.}

\item{force_extras}{Logical. If TRUE then the \verb{pad_extras' are implemented even if the total nuber of characters on the message line exceeds the }wdth'.}

\item{justf}{Vector with two integers. The first integer indicates
the message justification (-1 = left justified, 0 = centered, 1 = right
justified). The second integer in the input vector gives the number of
characters by which the `core_message' should be offset to the left or
right.}
}
\value{
Padded string that can be printed to the terminal/console.
}
\description{
Formats message padding and alignment.
}
\details{
This function is designed to be used within other functions. The
function takes a message and aligns it with padding and a preffix and
suffix.
}
\author{
Gordijn, Paul J.
}
\keyword{message,}
\keyword{pad,}
\keyword{padding,}
\keyword{print.}
\keyword{string,}
