% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_sub_phens_vars_sts.r
\name{phen_vars_sts}
\alias{phen_vars_sts}
\title{Transforms variable types}
\usage{
phen_vars_sts(phen_table = NULL, dta_in = NULL, ...)
}
\arguments{
\item{phen_table}{iPayipi standardised phenomena table.}

\item{dta_in}{Input table with standardised phenomena names.}
}
\value{
Data table with transformed variables.
}
\description{
Uses the variable types specified in the phenomena table to
standardise the variable type in ipayipi data tables.
}
\author{
Paul J. Gordijn
}
\keyword{data}
\keyword{pipeline;}
\keyword{transform}
\keyword{type.}
