% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_c4_phenomena_chk.r
\name{phenomena_chk}
\alias{phenomena_chk}
\title{Update meteorological file phenomena}
\usage{
phenomena_chk(
  wait_room = NA,
  check_phenomena = TRUE,
  file_ext_in = ".iph",
  csv_out = TRUE,
  ...
)
}
\arguments{
\item{wait_room}{The working directory where imported meteorological data
files are stored.}

\item{check_phenomena}{Should the file naming be check. This helps ensure
that files are named consistently and that data from different stations are
not appended erraneously.}

\item{nomtab_import}{Is there an update to the extant (or non-existant)
nomenclature table? This parameter can be used to input these updates.}

\item{out_csv}{Logical. If TRUE a csv file is made in the working
directory if there are files with unrecognised phenomena.}
}
\value{
returns a csv phenomena file. Screen output notes whether
the nomenclature table is complete.
}
\description{
A function to check standardisation of meteorological data
phenomena to be run before introducing data into the pipeline.
}
\author{
Paul J. Gordijn
}
\keyword{nomenclature}
