% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_c5_phenomena_read_csv.r
\name{phenomena_read_csv}
\alias{phenomena_read_csv}
\title{Update meteorological file phenomena}
\usage{
phenomena_read_csv(input_dir = ".", file = NULL, ...)
}
\arguments{
\item{input_dir}{The directory where imported meteorological data
files are stored with the phenomena table.}

\item{file}{The (base) file name of the csv import. If NULL then the
function searches for the most recently modified csv file to read (and
deletes older phenomena csv files).}
}
\value{
returns a copy of the finalized phenomena table
}
\description{
A function to check standardisation of meteorological data
phenomena to be run before introducing data into the pipeline.
}
\author{
Paul J. Gordijn
}
\keyword{measurement}
\keyword{phenomena;}
\keyword{standardisation;}
