% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_c6_phenomena_sts.r
\name{phenomena_sts}
\alias{phenomena_sts}
\title{Standardises data phenomena in 'ipayipi' format.}
\usage{
phenomena_sts(
  wait_room = NULL,
  prompt = FALSE,
  recurr = TRUE,
  unwanted = NULL,
  file_ext_in = ".iph",
  file_ext_out = ".ipi",
  ...
)
}
\arguments{
\item{wait_room}{Path to the 'wait_room' directory.}

\item{prompt}{Should the function use an interactive file selection function
otherwise all files are returned. TRUE or FALSE.}

\item{recurr}{Should the function search recursively into sub directories
for hobo rainfall csv export files? TRUE or FALSE.}

\item{unwanted}{Vector of strings listing files that should not be included
in the import.}

\item{file_ext_in}{The file extension defaults to ".iph". Other file types
could be incorporatted if required.}
}
\description{
Standardises data phenomena in 'ipayipi' format.
}
\author{
Paul J. Gordijn
}
\keyword{automatic}
\keyword{batch}
\keyword{data;}
\keyword{file}
\keyword{meteorological}
\keyword{process;}
\keyword{standardisation;}
\keyword{standardise}
\keyword{station;}
\keyword{transform}
\keyword{variables}
\keyword{variables;}
\keyword{weather}
