% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_sub_dt_pipe_process.r
\name{pipe_process}
\alias{pipe_process}
\title{Organise and update a data processing pipeline}
\usage{
pipe_process(
  pipe_seq = NULL,
  pipe_memory = NULL,
  overwrite_pipe_memory = TRUE,
  pipe_eval = TRUE,
  update_pipe_data = FALSE,
  ...
)
}
\arguments{
\item{pipe_seq}{Pipeline processing steps generated using a list of
\code{ipayipi::p_dt()}. Both or either the \code{pipe_seq} or \code{pipe_memory} must be
supplied.}

\item{pipe_memory}{A station processing pipeline summary in table format.
If supplied this will be updated via comparison with pipe_seq.}

\item{overwrite_pipe_memory}{If TRUE then extant pipeline processing
summaries are overwritten, and \emph{vice versa}.}

\item{start_dttm}{Earliest date time for which data have been processed in
this table.}

\item{end_dttm}{Latest date time for which data have been processed in
this table.}
}
\value{
A standardised pipe process summary and a tag indicating whether an
update has been made to the pipeline. If an update has been made the
updates have to the entire length/duration, not only a time slice, of the
processing pipelines data.
}
\description{
Organise and update a data processing pipeline
}
\author{
Paul J. Gordijn
}
