% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_sub_dt_pipe_process_pipe_seq.r
\name{pipe_seq}
\alias{pipe_seq}
\title{Presentation of pipe processes in sequence}
\usage{
pipe_seq(
  p = NULL,
  pipe_eval = TRUE,
  output_dt_preffix = "dt_",
  output_dt_suffix = NULL,
  ...
)
}
\arguments{
\item{pipe_eval}{Logical. If TRUE then basic checks are done on the
processing steps to ensure all parameters are filled and correct.}
}
\value{
A standardised data table of class "pipe_seq".
}
\description{
Accepts a list of parameters corresponding to pipeline
processing steps, each compiled by \code{ipayipi::p_dt}.
}
\author{
Paul J. Gordijn
}
