% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_plot_bar_agg.r
\name{plot_bar_agg}
\alias{plot_bar_agg}
\title{Plot data by period aggregate}
\usage{
plot_bar_agg(
  input_dir = ".",
  agg = "1 month",
  phen_name = NULL,
  tbl_search_key = "dt_1_months_agg",
  show_gaps = FALSE,
  wanted = NULL,
  unwanted = NULL,
  x_lab = "Date-time",
  y_lab = NULL,
  file_ext = ".ipip",
  prompt = FALSE,
  recurr = TRUE,
  pipe_house = NULL,
  ...
)
}
\arguments{
\item{input_dir}{Folder in which to search for standardised station data. Defaults to the current directory.}

\item{agg}{The time period of data aggregation to be plotted given as a character string. The character string is evaluated by \code{ipayipi::sts_interval_name()} to ensure compatibility.}

\item{phen_name}{The standardised phenomena name. This needs to be an exact match to the desired phenomena.}

\item{tbl_search_key}{A character string used to filter station tables within which the \code{phen_name} is found. This is a search string filters station tables using the \verb{\%ilike\%} operator. Multiple search strings should be separated using the '|' symbol (regex apply).}

\item{wanted}{Character string of the station keyword. Use this to filter out stations that should not be plotted alongside eachother. If more than one search key is included, these should be separated by the bar character, e.g., \code{"mcp|manz"}, for use with data.table's \verb{\%ilike\%} operator.}

\item{unwanted}{Vector of strings listing files that should not be included in the import.}

\item{x_lab}{The x-axis label to be rendered in the 'ggplot'.}

\item{y_lab}{The y-axis label.}

\item{file_ext}{The station file extension. Inherits the 'ipayipi' defult of '.ipip'.}

\item{prompt}{Logical. If \code{TRUE} the function will prompt the user to identify stations for plotting. The wanted and unwanted filters still apply when in interactive mode (\code{TRUE}).}

\item{recurr}{Should the function search recursively? I.e., thorugh sub-folders as well --- \code{TRUE}/\code{FALSE}. Defaults to \code{FALSE}.}

\item{pipe_house}{If provided the 'ipip_room' will be used as the input directory. \emph{See} \code{ipip_house()}.}
}
\value{
A list containing the plot data, a ggplot object, and the aggregation period used to summarise data.
}
\description{
Plots data that has been processed by 'ipayipi' into a defined aggregation period.
}
\details{
When searching for station data only the first station table that matches tbl_search_key is extracted. Extracted tables will not be used if they don't have either a "date-time" column, or the specified 'phen_name'.
\emph{NA} values in data aggretation periods: if the 'gid' field (for gap identification) was created using \code{ipayipi::dt_agg()} a column in the returned dataa with the suffix '_nas' will be calculated whereby NA values are returned if a gap id ('gid') was present (\emph{see also} \code{ipayipi::gap_eval()}).
}
\author{
Paul J. Gordijn
}
\keyword{aggregates;}
\keyword{data}
\keyword{over}
\keyword{phenomena}
\keyword{plotting;}
\keyword{reporting;}
\keyword{view;}
