% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_plot_m_anomaly.R
\name{plot_m_anomaly}
\alias{plot_m_anomaly}
\title{Pulls multiple station data into a 'flat' file}
\usage{
plot_m_anomaly(
  input_dir = ".",
  phen_name = NULL,
  phen_units = "",
  output_dir = NULL,
  wanted = NULL,
  unwanted = NULL,
  out_csv = FALSE,
  out_tab_name = NULL,
  out_csv_preffix = "",
  recurr = TRUE,
  file_ext = ".ipip",
  prompt = FALSE,
  verbose = FALSE,
  xtra_v = FALSE,
  pipe_house = NULL,
  ...
)
}
\arguments{
\item{input_dir}{The directory in which to search for stations form which to extract time-series data.}

\item{phen_name}{The field/column/name of the phenomena which to join into a single data table.}

\item{output_dir}{The output directory where an output csv file is saved.}

\item{wanted}{A strong containing keywords to use to filter which stations are selected for processing. Multiple search kewords should be seperated with a bar ('|'), and spaces avoided unless part of the keyword.}

\item{unwanted}{Similar to wanted, but keywords for filtering out unwanted stations.}

\item{out_csv}{Logical. If TRUE a csv file is exported to the output directory.}

\item{out_csv_preffix}{Preffix for the output csv file. The phenomena name and then time interval by which the data are summarised are used as a suffix.}

\item{recurr}{Whether to search recursively through folders. Defaults to TRUE.}

\item{file_ext}{The extension of the stations from where data will be extracted. Defaults to ".ipip".}

\item{prompt}{Logical. If \code{TRUE} a prompt will be called so that the user can interactively select station files.}

\item{verbose}{Whether to report on progress.}

\item{xtra_v}{Extra verbose. Logical.}

\item{pipe_house}{If the \code{pipe_house} argument is provided the \code{pipe_house$ipip_room} will be used instead of the \code{input_dir}.}

\item{tab_names}{Vector of table names in a station files from which to extract data. Add items to the vector such that only the first matching table from a station is selected. Table names are selected via the \code{\%in\%} argument.}

\item{wide}{Logical. If FALSE output data will be pivoted to wide format. Defaults to TRUE.}

\item{ri}{Record-interval string. This function will guess the record interval used to generate a flat table. However, the guess may not be possible if there is insufficient data.}
}
\value{
A list containing 1) the summarised data in a single data.table, 2) a character string representing the time interval by which the data has been summarised, 3) the list of stations used for the summary, 4) the name of the table and the name of the field for which data was summarised.
}
\description{
Queries continuous data of common record intervals.
}
\details{
Note this function only extracts data from decompressed station files. If a station has not be decompressed the function will take time decompressing.
}
\author{
Paul J. Gordijn
}
\keyword{data}
\keyword{data.}
\keyword{data;}
\keyword{long-format}
\keyword{pipeline;}
\keyword{summarise}
\keyword{time-series}
