% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotdy_cleanr.R
\name{plotdy_cleanr}
\alias{plotdy_cleanr}
\title{plot hampel filter run results}
\usage{
plotdy_cleanr(
  input_dir = ".",
  phen_name = NULL,
  phen_units = "",
  outliers = TRUE,
  output_dir = NULL,
  wanted = NULL,
  unwanted = NULL,
  tbl_names = NULL,
  legend_options = "onmouseover",
  gaps = TRUE,
  mn = NULL,
  mx = NULL,
  slide = 1,
  dflt_nobs = NULL,
  verbose = FALSE,
  xtra_v = FALSE,
  chunk_v = FALSE,
  station_file = NULL,
  pipe_house = NULL,
  ...
)
}
\description{
plot hampel filter run results
}
