% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotdy_driftr.R
\name{plotdy_driftr}
\alias{plotdy_driftr}
\title{plot hampel filter run results}
\usage{
plotdy_driftr(
  station_file = NULL,
  pipe_house = NULL,
  tbl_name = NULL,
  phen_key = NULL,
  legend_options = "onmouseover",
  mn = NULL,
  mx = NULL,
  slide = 1,
  dflt_nobs = NULL,
  verbose = FALSE,
  xtra_v = FALSE,
  chunk_v = FALSE,
  ...
)
}
\description{
plot hampel filter run results
}
