% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotgg_chks.R
\name{plotgg_chks}
\alias{plotgg_chks}
\title{plot hampel filter run results in ggplot}
\usage{
plotgg_chks(
  input_dir = ".",
  output_dir = NULL,
  wanted = NULL,
  unwanted = NULL,
  tbl_names = NULL,
  start_dttm = NULL,
  end_dttm = NULL,
  slide = 1,
  dflt_nobs = NULL,
  verbose = FALSE,
  xtra_v = FALSE,
  chunk_v = FALSE,
  station_file = NULL,
  pipe_house = NULL,
  ...
)
}
\description{
plot hampel filter run results in ggplot
}
