% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotgg_cleanr.R
\name{plotgg_cleanr}
\alias{plotgg_cleanr}
\title{plot hampel filter run results in ggplot}
\usage{
plotgg_cleanr(
  input_dir = ".",
  phen_name = NULL,
  phen_units = "",
  outliers = TRUE,
  output_dir = NULL,
  wanted = NULL,
  unwanted = NULL,
  tbl_names = NULL,
  gaps = TRUE,
  start_dttm = NULL,
  end_dttm = NULL,
  slide = 1,
  dflt_nobs = NULL,
  verbose = FALSE,
  xtra_v = FALSE,
  chunk_v = FALSE,
  station_file = NULL,
  pipe_house = NULL,
  ...
)
}
\description{
plot hampel filter run results in ggplot
}
