% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rain_to_met_format.r
\name{rain2met_format}
\alias{rain2met_format}
\title{Convert rain imported standardised nomvet file into generic format}
\usage{
rain2met_format(
  input_dir = ".",
  recurr = FALSE,
  wanted = NULL,
  unwanted = NULL,
  prompt = FALSE,
  ...
)
}
\arguments{
\item{input_dir}{Folder in which to search for standardised SAEON hobo
rainfall files.}

\item{recurr}{Should the function search recursively? I.e., thorugh
sub-folders as well --- \code{TRUE}/\code{FALSE}.}

\item{wanted}{Character string of the station keyword. Use this to
filter out stations that should not be plotted alongside eachother. If more
than one search key is included, these should be separated by the bar
character, e.g., \code{"mcp|manz"}, for use with data.table's \verb{\%ilike\%} operator.}

\item{unwanted}{Vector of strings listing files that should not be included
in the import.}

\item{prompt}{Logical. If \code{TRUE} the function will prompt the user to
identify stations for plotting. The wanted and unwanted filters
still apply when in interactive mode (\code{TRUE}).}
}
\description{
Opens hobo rainfall files and coverts these into a new format
compatible with the generic rainfall files.
}
\author{
Paul J. Gordijn
}
\keyword{data}
\keyword{files;}
\keyword{graphs;}
\keyword{hobo}
\keyword{management;}
