% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rain_i_aggregate.r
\name{rain_agg}
\alias{rain_agg}
\title{Summarize rainfall data by multiple time periods}
\usage{
rain_agg(
  input_file = NULL,
  output_pref = "agg_",
  aggs = c("5 mins", "day", "month"),
  csv_out = FALSE,
  ignore_nas = FALSE,
  agg_offset = c("0 secs", "0 secs"),
  ...
)
}
\arguments{
\item{input_file}{Output from \code{clean_rain_hobo}}

\item{output_pref}{A string for the file name prefix. The time periods given
in the \code{aggs} argument will form the suffex. The default preffix used in
the data pipeline is 'agg_'. The 'agg_' preffix must be used for further
processing.}

\item{aggs}{vector of strings giving the time periods to use for aggregating
the data.}

\item{csv_out}{If TRUE the function will write a csv to the working
for each set of aggregated data.}

\item{ignore_nas}{Choose whether to ignore gaps and other \code{NA} values when
aggregating the data by sum.}

\item{agg_offset}{A vector of two strings that can be coerced to a
\code{lubridate} period. These can be used to offset the date-time from
which aggregations are totalled. For example, for rainfall totals estimated
from 8 am to 8pm the \code{agg_offset} should be set to c(8 hours, 8 hours).}
}
\value{
Returns the "SAEON_rain_data" object with appended rainfall data
aggregations.
}
\description{
Reads rainfall data from the \code{clean_rain_hobo()} function
then aggregates the data by specified time periods.
}
\details{
The function uses \code{data.table} to summarize the data by the time
periods of interest (generated with the help of \code{lubridate}). The data
processing and writing to csv can be done using parallel processing by
setting the cores argument to greater than one (but not exceeding the
number of cores on the system).
The data will be saved as a 'data.table' in the \strong{ipayipi} rainfall object
with the given \code{output_pref} --- if this value is changed the processing
of data within the pipeline will be obscured.
}
\author{
Paul J Gordijn
}
\keyword{aggregation}
\keyword{bucket}
\keyword{data}
\keyword{gauge,}
\keyword{hoboware,}
\keyword{rain}
\keyword{tipping}
