% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rain_h_false_tips.r
\name{rain_false_tips}
\alias{rain_false_tips}
\title{Create rainfall data table and flag 'false tips'}
\usage{
rain_false_tips(input_file = NULL, false_tip_thresh = 10 * 60, ...)
}
\arguments{
\item{input_file}{A data object standardised as a SAEON 'ipayipi' hobo
rainfall file.}

\item{false_tip_thresh}{The time buffer in seconds around logger
interferance events (e.g., logger connections/downloads) in which any
tips (logs) should be considered false. The default used by SAEON is ten
minutes, that is, 10x60=600 seconds. *Event thresholds for other types
of \verb{false tips} are specified during metadata collection.}
}
\value{
Object of class "SAEON_rainfall_data" with an updated 'fasle_tip'
table.
}
\description{
Identifies likely erraneous rainfall recordings. The tips
of a tipping bucket raingauge are classed as \code{FALSE} if they occur
within a specified time period, called the `event threshold', around an
interference event, e.g., when the data from the logger was downloaded.
In the process of identifying false tips the 'interference event' table
is updated or created and a rainfall data table created after removing
'false tips'.
}
\details{
\enumerate{
\item Creates table of false tips
}
\itemize{
\item "auto" false tips include tips within the threshold of a "logged"
event where there was 'interference with the logger'.
\item "auto_double_tip".
\item "manual" false tips imported from
}
\enumerate{
\item Checks for an event table---check for rain events and manually declared
false tips.
\item Generates rainfall table
\item Warning printout
}
}
\author{
Paul J. Gordijn
}
\keyword{bucket}
\keyword{cleaning,}
\keyword{data}
\keyword{false}
\keyword{gauge,}
\keyword{hoboware,}
\keyword{rain}
\keyword{tipping}
\keyword{tips}
