% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rain_h_false_tips_batch.r
\name{rain_false_tips_batch}
\alias{rain_false_tips_batch}
\title{Flag 'false tips' --- batch mode}
\usage{
rain_false_tips_batch(
  rainr_room = NULL,
  false_tip_thresh = 10 * 60,
  unwanted = NULL,
  wanted = NULL,
  prompt = TRUE,
  recurr = FALSE,
  ...
)
}
\arguments{
\item{rainr_room}{Directory in which the standardised hobo rainfall files
are located.}

\item{false_tip_thresh}{The time buffer in seconds around logger
interferance events (e.g., logger connections/downloads) in which any
tips (logs) should be considered false. The default used by SAEON is ten
minutes, that is, 10x60=600 seconds.}

\item{prompt}{Logical. If \code{TRUE} the function will enable interactive
selection of hobo rainfall files for processing in this function.}

\item{recurr}{Logical. Should the listing of standardised hobo rainfall
files be done by searching recursively through the input directory, that
is, the \code{rainr_room}.}
}
\value{
List of class "SAEON_rainfall_data" with an updated 'fasle_tip'
table. Updated files are saved (old files are overwritten) in the
specificed input directory.
}
\description{
Processes standardised hobo rainfall files in batch mode to
create/update a rainfall data table, and a table of 'false tips'.
}
\details{
\enumerate{
\item Generated list of available standardised rainfall stations in the
\code{rainr_room} using \code{ipayipi::dta_list}.
\item Uses \code{ipayipi::rain_false_tips} to generate rainfall data with
false tips removed, and a 'false tip' table.
\item Overwrites old files by saving over these.
}
}
\author{
Paul J. Gordijn
}
\keyword{bucket}
\keyword{cleaning,}
\keyword{data}
\keyword{false}
\keyword{gauge,}
\keyword{hoboware,}
\keyword{rain}
\keyword{tipping}
\keyword{tips}
