% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rain_g_gaps.r
\name{rain_gaps}
\alias{rain_gaps}
\title{Highlight missing data at a station}
\usage{
rain_gaps(
  input_file = NULL,
  gap_problem_thresh_s = 6 * 60 * 60,
  event_thresh_s = 10 * 60 * 60,
  ...
)
}
\arguments{
\item{input_file}{Standardised \strong{ipayipi} rainfall file.}

\item{gap_problem_thresh_s}{The threshold duration of a gap beyond which
the gap is considered problematic, and therefore, the data requires
infilling or 'patching'. The default is six hours given in seconds, i.e.,
6 * 60 * 60.}

\item{event_thresh_s}{A gap can also be specified in the event metadata by
only providing a single date-time stamp and an 'event_threshold_s'. By
taking the date-time stamp the function approximates the start and end-time
of the data gap by adding and subtracting the given \code{event_thresh_s} to the
date-time stamp, respectively. If the \code{event_thresh_s} is supplied in the
event data then the value supplied therein is used in preference to the
value provided in this function. The default event threshold given here
in seconds is ten minutes, i.e., 10 * 60 seconds.}
}
\value{
The input file (standardised R rainfall station) with a table
describing data 'gaps'
}
\description{
The period between each logger download, stopping, removal,
or discommissioning and the next time a logger beings recording data again
counts as missing data if it rained over this period.
}
\details{
There are a number of different types of data problems that may be
encountered whilst collecting rainfall data. Confusion has to be avoided
between declaring that no rainfall occurred or there were no recordings of
rainfall over a period of time. This function helps users identify these
different types problems with the data taking the following steps to
identify and classify data 'gaps' (i.e., missing data).
\enumerate{
\item Trim any manually declared data gaps by the start and end dates of the
extant rainfall data (start and end dates taken from the data summary).
\item Generate a table of data 'gaps' between each download and deploy event.
This type of data gap, i.e., the 'gap_type' is categorized as 'auto' in
the final output. The rest of the gap types dealt with are manually
declared data gaps.
\item If there has been event data appended to the station data the function
will search therein for any manually declared gaps, that is, the
'event_type' 'logger_declare_error_segment'. If these segments extend
'over' any 'auto' gaps they are trimmed to preserve the record of the
'auto' gap. If any 'auto' gaps fall within an manually declared error
segment then, they too, are considered a problem gap, even if the duration
of the 'auto' gap is short enough to not be considered a problem gap.
\item Within the pipeline notes can be appended to a gap using the
'logger_gap_notes' 'event_type'. To do this a date-time stamp must be
provided which falls within the gap of interest. When appending the notes
\emph{\strong{the function will check whether the 'problem_gap' field was marked}}.
\emph{If this field was FALSE then the gap is not considered a problematic gap}
, i.e., the user is condident that no rainfall occurred during the 'gap'
period. If the \emph{\strong{raining}} field was TRUE in the 'event_data' then the
gap is considered problematic.
\item A final 'gap' table is provided in the R data object. This table will
be used to highlight no data when aggregating the data. Furthermore, in
infilling exercises, the gaps provide a means for highlighting where
'patching' is required. Below is a description of the gap table fields:
}\tabular{ll}{
   Column \tab Description \cr
   gid \tab Unique identifier for each gap. \cr
   euid \tab Event ID number imported from manually declared error segments. \cr
   gap_start \tab The start date-time of a gap. \cr
   gap_end \tab The end date-time of the gap. \cr
   dt_diff_s \tab The time duration in seconds of the gap. \cr
   gap_problem_thresh_s \tab The threshold duration beyond which a gap is \cr
}


\if{html}{\out{<div class="sourceCode">}}\preformatted{considered problematic.|
}\if{html}{\out{</div>}}

|problem_gap    |Logical field indicating whether a gap is considered to be
problematic.|
|Notes    | Notes describing the gap. The function will append any notes
imported from the event data, if existent. |
}
\author{
Paul J. Gordijn
}
\keyword{data}
\keyword{data;}
\keyword{event}
\keyword{gaps;}
\keyword{missing}
\keyword{pipeline;}
