% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rain_g_gaps_batch.r
\name{rain_gaps_batch}
\alias{rain_gaps_batch}
\title{Bulk checking and summarising of data gaps}
\usage{
rain_gaps_batch(
  input_dir = NULL,
  recurr = FALSE,
  prompt = FALSE,
  wanted = NULL,
  unwanted = NULL,
  gap_problem_thresh_s = 6 * 60 * 60,
  event_thresh_s = 10 * 60,
  ...
)
}
\arguments{
\item{input_dir}{The path to the directory in which standardised 'ipayipi'
rainfall data files are located.}

\item{recurr}{Search recursively through the input directory for
standardised hobo rainfall station files.}

\item{prompt}{Logical parameter. If \code{TRUE} then the function will
promt the user to interactively select files for which to append metadata
to.}

\item{wanted}{Vector of strings listing files that should not be
included in the import.}

\item{unwanted}{The function uses this vector of character strings used
exclude files being processed which contain one of the search strings. The
strings are passed to \code{ipayipi::dta_list()}.}

\item{gap_problem_thresh_s}{The threshold duration of a gap beyond which
the gap is considered problematic, and therefore, the data requires
infilling or 'patching'. The default is six hours given in seconds, i.e.,
6 * 60 * 60.}

\item{event_thresh_s}{A gap can also be specified in the event metadata by
only providing a single date-time stamp and an 'event_threshold_s'. By
taking the date-time stamp the function approximates the start and end-time
of the data gap by adding and subtracting the given \code{event_thresh_s} to the
date-time stamp, respectively. If the \code{event_thresh_s} is supplied in the
event data then the value supplied therein is used in preference to the
value provided in this function. The default event threshold given here
in seconds is ten minutes, i.e., 10 * 60 seconds.}

\item{search_patterns}{Character string containing search strongs/words to
use to filter the stations bing processed. Multiple options/strings should
be seperated by the bar symbol ("|").}
}
\value{
Saved data files and a list of data sets which have their
respective metadata appended to them.
}
\description{
Processes selected or all starndardised files within a
directory, and identifies manually declared and automatically
detected data 'gaps'.
}
\details{
This function is designed for batch processing of rainfall data
which has been standardised in the \strong{ipayipi} data pipeline.
\enumerate{
\item Extracts standardised rainfall data files from 'rainr_room' using
search strings and/or an interactive selection method involving a prompt.
\item Checks and classifies data gaps using
\code{ipayipi::rain_data_gaps()}.
\item Appends table describing data gaps to each station.
\item Saves the station data with updated gap table.
}
}
\author{
Paul J. Gordijn
}
\keyword{batch}
\keyword{data}
\keyword{data;}
\keyword{field}
\keyword{gaps;}
\keyword{metadata;}
\keyword{missing}
\keyword{notes;}
\keyword{pipeline;}
\keyword{processing;}
\keyword{rainfall}
\keyword{supplementary}
