% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rain_c_hobo_clean.r
\name{rain_hobo_clean}
\alias{rain_hobo_clean}
\title{Clean hobo rainfall logger data}
\usage{
rain_hobo_clean(
  input_format = ".csv",
  input_file = NULL,
  input_file_name = NULL,
  false_tip_thresh = 10 * 60,
  tip_value = 0.254,
  dt_format = c("Ymd HMOS", "Ymd HMS", "Ymd IMOSp", "Ymd IMSp", "ymd HMOS", "ymd HMS",
    "ymd IMOSp", "ymd IMSp", "mdY HMOS", "mdY HMS", "mdy HMOS", "mdy HMS", "mdy IMOSp",
    "mdy IMSp", "dmY HMOS", "dmY HMS", "dmy IMOSp", "dmy IMSp"),
  tz = "Africa/Johannesburg",
  temp = TRUE,
  instrument_type = NA,
  ptitle = NA,
  ...
)
}
\arguments{
\item{input_format}{Format of the input file. Can be one of two formats:
\enumerate{
\item either a csv file which has been exported from hoboware, or 2., \emph{a
special case for reading a rainfall \verb{master file' (}mf') from excel
which has been imported in R using \code{ipayipi::rain_master()}}.
}}

\item{input_file}{The path and file name of the csv file to be read. \emph{If
the \code{input_format} is 'mf' then a segment (list item) from a master
rainfall file is passed to this parameter.}}

\item{input_file_name}{Parameter only used \emph{if} input format is 'mf'---
otherwise this parameter should be set to \code{NULL}. When importing using
the 'mf' format the file_name parameter needs to be set to the excel
file from which rainfall data was extracted. The given value will be
captured as the 'import_file_name' in the 'data_summary' produced for
the import data. Will be appended to metadata.}

\item{false_tip_thresh}{Numeric field; units in seconds. The time buffer
in seconds around logger interferance events (e.g., logger connections/
downloads) in which any tips (logs) should be considered false. The
default used by SAEON is ten minutes, that is, 10x60=600 seconds.}

\item{tip_value}{Numeric field; units in millimeters. The amount of
rainfall (in mm) a single tip of the tipping bucket represents. This
defaults so 0.254 mm.}

\item{dt_format}{The date-time format of the input data. \strong{Take care to
input data time formats}. NB! This may appear different depending on your
computer system and software settings.  The default used is "\%x \%r" ---
check \code{?base::strptime} for a list of available formats.}

\item{tz}{Timezone format of the datetime values. For South Africa we use
the character string "Africa/Johannesburg" or "Africa/Johannesburg" for
short.}

\item{temp}{Has the hobo logger been recording temperature? Logical
variable. Defaults to FALSE.}

\item{instrument_type}{The type of instrument wherein the hobo logger was
deployed. For the commonly used \emph{Texus tipping-bucket rainguage} use the
code "txrg".}

\item{ptitle}{The plot title should only be supplied if there is no plot
title in the hobo file which is being exported, or when using
\code{rain_master()}. The supplied title will overwrite the export files
plot title. Defaults to \code{NA}.}
}
\value{
list of class "SAEON_rainfall_data". Contains data and logger
metadata.
}
\description{
Reads Hoboware csv outputs from tipping bucket raingauges and
and performs some basic data cleaning. \emph{The function has a special mode
for which input parameters may vary}.
}
\details{
This function reads into R a csv output file from Hoboware software.
The interference events, that is, when loggers are, "detached", "attached",
"connected" to a host, and when a "file ends", as exported from Hoboware
are summarized and stored in a seperate table in the output list. \emph{The date-
time format of the files are also standardised and this should match the
csv file being imported to prevent incorrect date-time value conversions.}
The function performs a number of checks on the data to make sure that it is
in the correct format:
\enumerate{
\item If there is no plot title supplied in the \code{ptitle} input parameter then
the function will check for a plot title in the hobo file export which is
being read. The plot title should \emph{only be supplied by the user} for a
specific reason, e.g., using \code{rain_master()}. The plot title is the
station--instrument name and in hobo file exports is populated by the user
and programed to be in the first line of the first row and column of the
data table (csv or tsv) export.
If there is no plot title in the document the function requires the user to
have entered this value in the function input.
\item The function will check that all the correct columns have been supplied
in the hobo file export. If there is a column missing then the function
will print a warning message in the terminal. If critical columns are
missing or if there are no columns in addition to the critical ones, the
function will terminate. The following columns are considered critical:
"id", "date_time", and "cumm_rain".
To detect the columns in the rainfall file being imported the following
synonyms (search words) are used (note that the temperature column is
optional).
}\tabular{lll}{
   Column \tab Synonyms \tab Description \cr
   id \tab # \tab unique identifier from import file. \cr
   date_time \tab date; dt \tab The date and time. \cr
   cumm_rain \tab rainfall; rain; precipitation \tab Cummulative rainfall. \cr
   temp \tab temp \tab Temperature in degrees celcius (optional). \cr
   detached \tab detached \tab Detachment of logger from device. \cr
   attached \tab attached \tab Attachment of logger to device. \cr
   host \tab host \tab Connection to a host. \cr
   file_end \tab 'end of file' \tab The end of a download file. \cr
}


If non-critical columns are missing then they are populated with \code{NA}
values (except the temperature column which is optional).
3. If the date-time format has been incorrectly specified the resulting
coersion of the date-time column using the \code{base::as.POSIXct()} can
produce \code{NA} values. If this occurs the function will terminate and a
message printed in the terminal.

\strong{Note that when using this function in "mf" mode only one station can be
processed at a time.}
}
\author{
Paul J. Gordijn
}
\keyword{bucket}
\keyword{cleaning}
\keyword{data}
\keyword{gauge,}
\keyword{hoboware,}
\keyword{rain}
\keyword{tipping}
