% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rain_c_hobo_conversion_batch.r
\name{rain_hobo_conversion_batch}
\alias{rain_hobo_conversion_batch}
\title{Batch processing and standardisation of hobo rainfall exports}
\usage{
rain_hobo_conversion_batch(
  wait_room = NULL,
  out_csv = TRUE,
  input_format = ".csv",
  false_tip_thresh = 10 * 60,
  tip_value = 0.254,
  dt_format = c("Ymd HMOS", "Ymd HMS", "Ymd IMOSp", "Ymd IMSp", "ymd HMOS", "ymd HMS",
    "ymd IMOSp", "ymd IMSp", "mdY HMOS", "mdY HMS", "mdy HMOS", "mdy HMS", "mdy IMOSp",
    "mdy IMSp", "dmY HMOS", "dmY HMS", "dmy IMOSp", "dmy IMSp"),
  tz = "Africa/Johannesburg",
  temp = FALSE,
  instrument_type = NA,
  mf_list = NA,
  ...
)
}
\arguments{
\item{wait_room}{The working directory where the hobo csv files are
temporarily stored. This is also where the pipelines nomenclature table
is stored and used for associated standardisation.}

\item{out_csv}{Logical. If \code{TRUE} a csv file is made in the working
directory if there are files with unrecognised nomenclature.}

\item{input_format}{Currently only ".csv" is supported. \emph{Note that the period
(\strong{"."}) must be included in the extension}.}

\item{false_tip_thresh}{Numeric field; units in seconds. The time buffer
in seconds around logger interferance events (e.g., logger connections/
downloads) in which any tips (logs) should be considered false. The
default used by SAEON is ten minutes, that is, 10x60=600 seconds.}

\item{tip_value}{Numeric field; units in millimeters. The amount of
rainfall (in mm) a single tip of the tipping bucket represents. This
defaults so 0.254 mm.}

\item{dt_format}{The datetime format of the csv (or other) file. Note
that \strong{for batch processing all files being read must have the same
datetime format}. The default used is "\%x \%r" --- check the
\code{base::strptime} function for a list of available formats.}

\item{tz}{Timezone format of the datetime values. For South Africa we use
the character string "Africa/Johannesburg" or "Africa/Johannesburg" for
short.}

\item{temp}{Has the hobo logger been recording temperature? Logical
variable. Defaults to \code{FALSE}.}

\item{instrument_type}{The type of instrument wherein the hobo logger was
deployed. For the commonly used \emph{Texus tipping-bucket rainguage} use the
code "txrg".}

\item{mf_list}{Output from \code{ipayipi::rain_master()} if reading rainfall
files from a 'master file'. If an object is supplied here then files
are automatically processed in the "\emph{mf}" mode.}
}
\value{
If the nomenclature table was updated successfully this function
will return a list containing the updated table, a list of successfully
converted hobo rainfall files, and a list of failed conversions.
}
\description{
Converts a batch of hobo rainfall file exports to R format
then performs some standardisastion of the data and metadata. Some initial
data cleaning operations are also performed by
\code{ipayipi::rain_hobo_clean()}. *Data can also be imported from
\code{ipayipi::rain_master()}.
}
\details{
Once data has been imported into the 'nomvet_room' using
\code{ipayipi::rain_hobo_import()}, \code{ipayipi::rain_hobo_conversion()} checks
the naming conventions of the files being imported against a nomenclature
table. This is a critical step before going ahead with appending data as
the naming standards determine which data stations have their data appended
into a continuous series. To get this right the nomenclature is verified
using \code{ipayipi::rain_nomenclature()} whereby a specific data stream
(or station) is identified by the combination of unique \emph{import file names}
(location and station), \emph{logger serial numbers}, and the \emph{'plot title'}
which is specified in Hoboware (or alternatively provided by the user when
using the \code{ipayipi::rain_master()} import functionality). If there is
incomplete nomenclature data the function will not continue until this
information (\code{ipayipi::rain_read_nomtab_csv()} can be used to do this)
has been updated. In the process the 'nomtab.rds' file, that is, the
standardised nomenclature table, will be updated in the 'wait_room'.
\emph{Master file} input mode: data can be imported into this function from
\code{ipayipi::rain_master()}. The following options must be set for reading
files using this functionality.
\itemize{
\item \code{mf_list} This parameter must be supplied with an output from
\code{ipayipi::rain_master()}.
\strong{Note that when using this function in "mf" mode only one station can be
processed at a time.}
}
}
\author{
Paul J. Gordijn
}
\keyword{batch}
\keyword{data}
\keyword{nomenclature;}
\keyword{pipeline}
\keyword{preparation;}
\keyword{processing}
\keyword{standardization;}
