% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rain_b_hobo_import_batch.r
\name{rain_hobo_import_batch}
\alias{rain_hobo_import_batch}
\title{Import batches of hobo data into the 'wait_room'}
\usage{
rain_hobo_import_batch(
  source_dir = ".",
  wait_room = NULL,
  prompt = FALSE,
  recurr = FALSE,
  wanted = NULL,
  unwanted = NULL,
  file_ext = ".csv",
  ...
)
}
\arguments{
\item{source_dir}{The path directory from where hobo rainfall files are to
be read.}

\item{wait_room}{Path to the 'wait_room' directory.}

\item{prompt}{Should the function use an interactive file selection function
otherwise all files are returned. TRUE or FALSE.}

\item{recurr}{Should the function search recursively into sub directories
for hobo rainfall csv export files? TRUE or FALSE.}

\item{wanted}{Vector of strings listing files that should not be
included in the import.}

\item{unwanted}{Vector of strings listing files that should not be included
in the import.}

\item{file_ext}{The file extension defaults to ".csv". Other file types could
be incorporatted if required.}
}
\value{
Terminal printout of files transferred indicating the success of
reading the data in the appropriate format.
}
\description{
Introduces hobo rainfall data 'csv' exports into the \emph{ipayipi}
rainfall data pipeline.
}
\details{
This function simply copies the csv hobo exports and pastes these
into the specified 'wait_room' folder. Duplicate file names are handled by
appending a number to the end of file names when pasting these.
}
\author{
Paul J. Gordijn
}
\keyword{bucket}
\keyword{cleaning,}
\keyword{data}
\keyword{false}
\keyword{gauge,}
\keyword{hoboware,}
\keyword{import}
\keyword{pipeline,}
\keyword{rain}
\keyword{tipping}
\keyword{tips,}
