% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rain_d_import_remove.r
\name{rain_import_remove}
\alias{rain_import_remove}
\title{Remove csv files from the 'wait_room'}
\usage{
rain_import_remove(rain_in = NULL, ...)
}
\arguments{
\item{rain_in}{The output from \code{ipayipi::rain_hobo_conversion()}.}
}
\value{
Removes files from the waiting room.
}
\description{
Uses a list of successfully converted hobo rainfall files from
a 'converted_rain_hobo' object generated by
\code{ipayipi::rain_hobo_conversion()} to select files from the 'wait_room' and
remove the imported files.
}
\details{
Extracts all the import file names from the batch of successfully converted
hobo rainfall file exports, which have been extracted from files imported
into the 'wait_room', then deletes the files.
}
\author{
Paul J. Gordijn
}
\keyword{append}
\keyword{batch}
\keyword{data}
\keyword{data;}
\keyword{hobo}
\keyword{processing;}
\keyword{rainfall}
