% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rain_c_master.r
\name{rain_master}
\alias{rain_master}
\title{Extract rainfall data from 'master' files}
\usage{
rain_master(input_dir = ".", m_file = NULL, ptitle = NA, ...)
}
\arguments{
\item{input_dir}{The directory in which the file is stored.}

\item{m_file}{The 'master' filename (excluding the directory).}

\item{ptitle}{The standardised title name of the rain gauge station. This
value has to be supplied.}
}
\value{
List of seperate hoboware rainfall exports in seperate data frames
with no formatting applied.
}
\description{
Reads Hoboware csv outputs that have been collated in an
excel file. The format of the excel file (esp. spacing of files and
column headers need to be considered carefully---see example data).
The excel data needs to be converted to 'csv' format before importing.
}
\details{
This function returns the 'raw' hoboware file csv or text exports
which have been appended in excel. It is critical that a blank line is
placed between each csv export, the 'plot title' is placed below this
blank line, then below the title the column headers from the export. The
'#' needs to be retained (as the column header for the id row) as it
signals the start of a new hobo download file.
The exports that this function generates can be read by clean_rain_hobo().
See the example data: "1C Raingauge_Masterfile_20211216.xlsx".
}
\author{
Paul J. Gordijn
}
\keyword{bucket}
\keyword{data}
\keyword{file}
\keyword{gauge,}
\keyword{hoboware,}
\keyword{import,}
\keyword{master}
\keyword{rain}
\keyword{tipping}
