% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rain_f_meta.r
\name{rain_meta}
\alias{rain_meta}
\title{Append respective event metadata to a station}
\usage{
rain_meta(input_file = NULL, meta_file = NULL, ...)
}
\arguments{
\item{input_file}{Standardised ipayipi rainfalll data set (stored in
the \code{rainr_room}).}

\item{meta_file}{Standardised rainfall metadata database.}
}
\value{
Standardised ipayipi rainfall station object with updated event
metadata.
}
\description{
Reads standardised metadata and appends records the matching
station.
}
\details{
The function takes event metadata which matches the standardised
}
\author{
Paul J. Gordijn
}
\keyword{data}
\keyword{data;}
\keyword{field}
\keyword{metadata;}
\keyword{notes}
\keyword{pipeline;}
\keyword{processing;}
\keyword{rainfall}
\keyword{supplementary}
