% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rain_f_meta_batch.r
\name{rain_meta_batch}
\alias{rain_meta_batch}
\title{Batch append metadata to a station}
\usage{
rain_meta_batch(
  input_dir = NULL,
  meta_file_name = "event_db.rmds",
  recurr = FALSE,
  prompt = FALSE,
  wanted = NULL,
  unwanted = NULL,
  dt_format = "\%Y-\%m-\%d \%H:\%M:\%S",
  tz = "Africa/Johannesburg",
  ...
)
}
\arguments{
\item{input_dir}{Directory where the standardised ipayipi rainfalll
data sets are located (i.e., the \code{rainr_room}).}

\item{meta_file_name}{The name of the standardised rainfall metadata
database.}

\item{recurr}{Search recursively through the input directory for
standardised rainfall files.}

\item{prompt}{Logical parameter. If \code{TRUE} then the function will
promt the user to interactively select files for which to append metadata
to.}

\item{wanted}{Character string containing search strongs/words to
use to filter the station data. Multiple options/strings should be
seperated by the bar symbol ("|").}

\item{unwanted}{The function uses this vector of character strings used
exclude files being processed which contain one of the search strings. The
strings are passed to \code{ipayipi::dta_list()}.}

\item{dt_format}{The date-time format passed to
\code{ipayipi::rain_meta_read()} --- the date-time format of the event
metadata database.}

\item{tz}{The time zone passed to \code{ipayipi::rain_meta_read()} ---
the time zone of the event metadata database must match the data in
question.}
}
\value{
Saved data files and a list of data sets which have their
respective metadata appended to them.
}
\description{
Reads standardised metadata and appends records the matching
station.
}
\details{
This function is designed for batch processing of rainfall data
which has been standardised in the \strong{ipayipi} data pipeline.
\enumerate{
\item Opens and checks the events metadata database using
\code{ipayipi::rain_meta_read()}.
\item Extracts standardised rainfall data files from 'rainr_room' using
search strings and/or an interactive selection method involving a prompt.
\item Appends respective event metadata to each station.
\item Saves the stations with updated event metadata.
}
}
\author{
Paul J. Gordijn
}
\keyword{batch}
\keyword{data}
\keyword{data;}
\keyword{field}
\keyword{metadata;}
\keyword{notes;}
\keyword{pipeline;}
\keyword{processing;}
\keyword{rainfall}
\keyword{supplementary}
