% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rain_f_meta_read.r
\name{rain_meta_read}
\alias{rain_meta_read}
\title{Open and check rainfall event metadata database.}
\usage{
rain_meta_read(
  input_dir = NULL,
  meta_file_name = "event_db.rmds",
  dt_format = "\%Y-\%m-\%d \%H:\%M:\%S",
  tz = "Africa/Johannesburg",
  ...
)
}
\arguments{
\item{input_dir}{Directory in which to search for the metadata database.}

\item{meta_file_name}{File name of the standardised rainfall metadata
database.}

\item{dt_format}{The date-time format of the event metadata database.}

\item{tz}{The time zone of the event metadata database that must match
the data in question.}
}
\value{
Standardised data table of events.
}
\description{
Reads standardised metadata and appends records the matching
station.
}
\details{
Reads in an events database or sheet in 'csv' format. Checks that
column names have been standardised. Transforms the date-time columns to
a standardised format --- ** this format and timezone must match that used
by the data pipeline **.
}
\author{
Paul J. Gordijn
}
\keyword{data}
\keyword{data;}
\keyword{field}
\keyword{metadata;}
\keyword{notes}
\keyword{pipeline;}
\keyword{processing;}
\keyword{rainfall}
\keyword{supplementary}
