% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rain_c_nomenclature.r
\name{rain_nomenclature}
\alias{rain_nomenclature}
\title{Check rainfall file nomenclature in the waiting room}
\usage{
rain_nomenclature(
  wait_room = NA,
  nomvet_room = NA,
  nomtab_import = NA,
  check_nomenclature = FALSE,
  csv_out = TRUE,
  ...
)
}
\arguments{
\item{wait_room}{The working directory where the xle files are stored.}

\item{nomtab_import}{Is there an update to the extant (or non-existant)
nomenclature table? This parameter can be used to input these updates.
This input is generated by \code{ipayipi::hobo_clean()}.}

\item{check_nomenclature}{Should the file naming be check. This helps ensure
that files are named consistently and that data from different stations are
not appended erraneously.}

\item{out_csv}{Logical. If TRUE a csv file is made in the working
directory if there are files with unrecognised nomenclature.}
}
\value{
returns a csv nomenclature file. Screen output notes whether
the nomenclature table is complete.
}
\description{
A function to check standardisation of rainfall data objects
to be run before introducing data into the pipeline.
}
\details{
Aims: Generate and/or update existing nomenclature tables
merging the 'waiting room' data with vetted nomenclature data.
\enumerate{
\item Check data nomenclature.
\item Compare nomenclature from (1) with an existing nomenclature table
(if this exists).
\item Output updated nomenclature table and highlight where manual
edits are required.
\item Save updated nomenclature table as usable if no manual edits are
are required.
}
Note that the double underscore "__" is treated as a special character to
designate special logger types, i.e. salinity/conductivity loggers.
NB! If the embedded xml metadata nomenclature differs from the required
standardisation, these will be changed accordingly. Encoding of
the xle files is also standardised as UTF-8.
}
\examples{
# check the nomenclature of the files in the "waiting room" or any
# other folder with xle files against a standard list of names.
wait_room <- "./temp" # directory (temporary) with xle data
gw_xle_nomenclature(wait_room = wait_room, out_csv = TRUE)
# if there are unrecognized names then the function generates a csv
# in the directory specified. The updated csv can be used to update the
# "nomvet.rds" file (which in the pipeline is stored in the
# "waiting room" folder).
}
\author{
Paul J. Gordijn
}
\keyword{nomenclature}
