% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rain_a_pipe_init.r
\name{rain_pipe_init}
\alias{rain_pipe_init}
\title{Pipe directory intitiate---rainfall}
\usage{
rain_pipe_init(work_dir = ".")
}
\arguments{
\item{work_dir}{The folder, that is, directory in which to create the
pipeline directories.}
}
\value{
Notification of directory creation.
}
\description{
Creates the necessary folders for the ipaypi rainfall
data processing pipeline.
}
\details{
This function automates the creation of three folders/diretories
that are requried for bulk processing of files in the ipayipi groundwater
pipeline structure. The directories are:
\itemize{
\item \strong{wait_room} --- for raw data imports
\item \strong{nomvet} --- for storing standardized native solonist data files
\item \strong{rainrr_dta} --- transforming and processing standardized native data
& processing
}
}
\examples{
# Inititate pipeline
dir <- "." # define the working directory
pipe_init_rain(work_dir = dir)
}
\author{
Paul J. Gordijn
}
\keyword{creation,}
\keyword{directory}
\keyword{folder}
\keyword{initiate}
\keyword{pipeline,}
\keyword{structure.}
