% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rain_plot_accumulation.r
\name{rain_plot_accumulation}
\alias{rain_plot_accumulation}
\title{Plot rainfall accumulation from a batch of rain gauges}
\usage{
rain_plot_accumulation(
  input_dir = ".",
  recurr = FALSE,
  wanted = NULL,
  unwanted = NULL,
  prompt = FALSE,
  ...
)
}
\arguments{
\item{input_dir}{Folder in which to search for standardised SAEON hobo
rainfall files.}

\item{recurr}{Should the function search recursively? I.e., thorugh
sub-folders as well --- \code{TRUE}/\code{FALSE}.}

\item{wanted}{Character string of the station keyword. Use this to
filter out stations that should not be plotted alongside eachother. If more
than one search key is included, these should be separated by the bar
character, e.g., \code{"mcp|manz"}, for use with data.table's \verb{\%ilike\%} operator.}

\item{unwanted}{Vector of strings listing files that should not be included
in the import.}

\item{prompt}{Logical. If \code{TRUE} the function will prompt the user to
identify stations for plotting. The wanted and unwanted filters
still apply when in interactive mode (\code{TRUE}).}
}
\description{
Uses ggplot to generate a plot of rainfall tip accumulation
using standardised rainfall data.
}
\details{
This function will plots rainfall accumulation of a number of
stations (rain gauges). Data is extracted from the selected stations'
'rain_data' data table. The plot can be made into an interactive chart using
the \code{plotly} and \code{ggplot2} plotting interoperability.
}
\author{
Paul J. Gordijn
}
\keyword{cummulative}
\keyword{data}
\keyword{over}
\keyword{plotting;}
\keyword{rainfall;}
\keyword{reporting;}
\keyword{view;}
