% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rain_plot_aggs.r
\name{rain_plot_aggs}
\alias{rain_plot_aggs}
\title{Plot rainfall accumulation from a batch of rain gauges}
\usage{
rain_plot_aggs(
  agg = "month",
  input_dir = ".",
  recurr = FALSE,
  wanted = NULL,
  unwanted = NULL,
  prompt = FALSE,
  ignore_nas = FALSE,
  ...
)
}
\arguments{
\item{agg}{The time period of rainfall aggregation to be plotted given
as a character string. The time period for the aggregation must match one
of those processed for the stations in question by the
\code{ipayipi::rain_agg} or the equivalent for batch processing, that is,
\code{ipayipi::rain_agg_batch}.}

\item{input_dir}{Folder in which to search for standardised SAEON ipayipi
rainfall files.}

\item{recurr}{Should the function search recursively? I.e., thorugh
sub-folders as well --- \code{TRUE}/\code{FALSE}.}

\item{wanted}{Character string of the station keyword. Use this to
filter out stations that should not be plotted alongside eachother. If more
than one search key is included, these should be separated by the bar
character, e.g., \code{"mcp|manz"}, for use with data.table's \verb{\%ilike\%} operator.}

\item{unwanted}{Vector of strings listing files that should not be included
in the import.}

\item{prompt}{Logical. If \code{TRUE} the function will prompt the user to
identify stations for plotting. The wanted and unwanted filters
still apply when in interactive mode (\code{TRUE}).}

\item{ignore_nas}{Default is FALSE. When set to TRUE the function aggregating
data will use na.rm = TRUE to avoid return NA values.}
}
\value{
A list containing the ggplot plot, the data used for plotting, and
the aggregation period (\code{agg}) used for data extraction or summary.
}
\description{
Uses ggplot to generate a plot
of rainfall tip accumulation using standardised rainfall data.
}
\details{
Extracts available data from the selected stations and prepares it
for plotting ggplot.
}
\author{
Paul J. Gordijn
}
\keyword{aggregates;}
\keyword{data}
\keyword{over}
\keyword{plotting;}
\keyword{rainfall}
\keyword{reporting;}
\keyword{view;}
