% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rain_plot_availibility.r
\name{rain_plot_availability}
\alias{rain_plot_availability}
\title{Plot the coverage of standardised hobo rainfall files}
\usage{
rain_plot_availability(
  input_dir = ".",
  recurr = FALSE,
  wanted = NULL,
  unwanted = NULL,
  prompt = FALSE,
  ...
)
}
\arguments{
\item{input_dir}{Folder in which to search for standardised SAEON hobo
rainfall files.}

\item{recurr}{Should the function search recursively? I.e., thorugh
sub-folders as well --- \code{TRUE}/\code{FALSE}.}

\item{wanted}{Character string of the station keyword. Use this to
filter out stations that should not be plotted alongside eachother. If more
than one search key is included, these should be separated by the bar
character, e.g., \code{"mcp|manz"}, for use with data.table's \verb{\%ilike\%} operator.}

\item{unwanted}{Vector of strings listing files that should not be included
in the import.}

\item{prompt}{Logical. If \code{TRUE} the function will prompt the user to
identify stations for plotting. The wanted and unwanted filters
still apply when in interactive mode (\code{TRUE}).}
}
\value{
A list containing a plot, which shows the availability of data, the
plot data and a seperate data table showing the mid-point, in time, of each
problem gap, plus associated gap details. This data could be used to
annotated the plot. The plot is produced using \code{ggplot2::ggplot()}.
}
\description{
Uses dygraphs to plot the availability of data at a group
of hobo rainfall stations.
}
\details{
This function will plot the continuity of data from a number of
stations (rain gauges). The potential continuity of data, from start to end,
is plotted then overlaid with 'problem' data gaps. Data gaps are extracted
from the 'gaps' table in a standardised \code{ipayipi} data object. See the
\code{ipayipi::rain_gaps()} and \code{ipayipi::rain_gaps_batch()} functions.
}
\author{
Paul J. Gordijn
}
\keyword{data}
\keyword{data;}
\keyword{files;}
\keyword{graphs;}
\keyword{hobo}
\keyword{management;}
\keyword{missing}
\keyword{over}
\keyword{plotting;}
\keyword{reporting;}
\keyword{view;}
