% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rain_e_station_append.r
\name{rain_station_append}
\alias{rain_station_append}
\title{Append hobo rainfall station data}
\usage{
rain_station_append(
  rain_station = NULL,
  new_data = NULL,
  overwrite_old = FALSE,
  ...
)
}
\arguments{
\item{new_data}{The data that will be appended to the main 'hobo station'.}

\item{overwrite_old}{Logical. Defaults to FALSE so that data from the
\code{hobo_station} is not over writed by the \code{new_data}.}

\item{hobo_station}{The main data set to which data will be appended.}
}
\description{
Appends two standardised hobo rainfall files.
}
\details{
Special attention is given to the \verb{data summary} table of
\code{SAEON_rain_data} objects. The data summary contains information
which is used in the pipeline to determine which files are appended to
another. Specifically, the standardised ptitle in the data summary is the
identifier for stations, and records in the \code{nomvet_name} field are used to
determine which files have been included in any given \code{SAEON_rain_data}
object.
}
\author{
Paul J. Gordijn
}
\keyword{append}
\keyword{data}
\keyword{data;}
\keyword{hobo}
\keyword{pipeline;}
\keyword{rainfall;}
