% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rain_e_station_append_batch.r
\name{rain_station_append_batch}
\alias{rain_station_append_batch}
\title{Batch appending of ipayipi rainfall data}
\usage{
rain_station_append_batch(
  nomvet_room = NULL,
  rainr_room = NULL,
  overwrite_old = FALSE,
  ...
)
}
\arguments{
\item{nomvet_room}{Directory where the standardised hoho rainfall files
have been archived by the ipayipi rainfall data pieline.}

\item{rainr_room}{Directory where the station's data are stored, data
appended to these, and newer data appended them.}

\item{overwrite_old}{Must older data be overwritten by new data? By
default this option is set to FALSE to preserve old data records.}
}
\description{
The end goal of this function is to update the main station
data files with data stored in the 'nomvet_room'. If new data is
detected then it is appended to a main data file for a specific station
in the 'rainr_room'.
}
\details{
The main stations are identified by the standardised ptitle field. New data
are appended to stations in the 'rainr_room'. If the station does not exist
in the 'rainr_room', but there is data in the 'nomvet_room' a new station is
created in the 'rainr_room'. The following describes the general steps
undertaken by \code{ipayipi::rain_batch_append()}.
\enumerate{
\item Create inventory of files in the 'nomvet_room'. \emph{The inventory is
generated by extracting all the unique file names in the 'nomvet_room'.}
\item Create inventory of files which have been introduced into the
'rainr_room'. *This inventory is generated from the 'nomvet_name' field
in each of the \emph{\strong{stations}} 'data_summary' tables in the 'rainr_room'.
\item Pass the 'new data' to \code{ipayipi::rain_station_append()} and save the
appended outputs in the 'rainr_room'.
}
}
\author{
Paul J. Gordijn
}
\keyword{append}
\keyword{batch}
\keyword{data}
\keyword{data;}
\keyword{hobo}
\keyword{processing;}
\keyword{rainfall}
