% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_c5_read_phentab_csv.R
\name{read_phentab_csv}
\alias{read_phentab_csv}
\title{Update data phenomena}
\usage{
read_phentab_csv(pipe_house = NULL, file = NULL, ...)
}
\arguments{
\item{pipe_house}{List of pipeline directories. \strong{See}
\code{ipayipi::ipip_house()} \strong{for details}.}

\item{file}{The (base) file name of the csv import. If NULL then the
function searches for the most recently modified csv file to read (and
deletes older phenomena csv files).}
}
\value{
returns a copy of the finalized phenomena table
}
\description{
A function to check standardisation of logger data
phenomena to be run before introducing data into the pipeline.
}
\author{
Paul J. Gordijn
}
\keyword{measurement}
\keyword{phenomena;}
\keyword{standardisation}
\keyword{standardisation;}
\keyword{unit}
