% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_sub_record_interval_eval.R
\name{record_interval_eval}
\alias{record_interval_eval}
\title{Evaluates the consistency of record interval date-time values}
\usage{
record_interval_eval(
  dt = NULL,
  dta_in = NULL,
  remove_prompt = FALSE,
  record_interval_type = NULL,
  max_rows = 1000,
  ...
)
}
\arguments{
\item{dt}{vector of POSIXct values with defined time zone.}

\item{dta_in}{Input data must have the same number of rows as dt. If inconsistent time intervals are detected then the input data will be filtered out and returned.}

\item{remove_prompt}{Logical. Activate a readline prompt to choose whether or not filter our records from \code{dta_in} with inconsistent record intervals. If TRUE then the \code{max_rows} argument is ignored.}

\item{record_interval_type}{If there is only one data record the this parameter will be used as the default. This must be either "event_based" or "continuous".}

\item{max_rows}{Maximum length of \code{dt} vector to use for evaluating the record interval. Defaults to 1000. If \code{remove_prompt} is TRUE, this argument is ignored.}

\item{dt_format}{The input date-time format of the time series, e.g., "\%y-\%m-\%d \%H:\%M:\%S". See ?base::strptime() for details.}
}
\value{
List containing a logical list of 1) the adjusted date-time values
that will not include irregular time values if these have been removed
2) interval checks (whether or not the record interval was continuous, if
in deed the \code{record_interval_type} was set to 'continuous'), 3) the modal
record interval given as an object of class 'diff time', 4) the record
interval as a diff time character value, the new filtered data with
inconsistent record interval data removed, 6) the filtered date time
values, and 6) the interval type, e.g., "continuous" or "event_based".
}
\description{
This function tests record intervals from a data.table with a time stamp. The test then describes the data as "continuous" for regular time intervals, "event_based" for irregular or discontinuous recordings. A combination of "continuous" and "event_based" is described as "mixed". Note monthly data are considered to be 'continuous'.
}
\details{
If the record iterval is set to "continuous" then the function checks a series of date-time values for inconsistencies in record intervals. Data records can be removed if the data is provided and the user selects to do so in the \code{remove_prompt}. For data to be considered 'continuous' 98\% of intervals in the data table must be continuous, \emph{and} more than 98\% of date-time stamp second values must equals zero. If between 20 and 98\% (and time values include second values other than zero) of intervals are regular then the 'series' is classified as "mixed". If < 20\% of intervals are regular, then the series is considered to be "event_based".
There is a special case for monthly time difference intervals. These are handled by the \code{mondate} package. If seconds and minutes are equal to zero and the difference between sequential date-time values are 28, 29, 30, and 31 days, then the series is described as a 'continuous' monthly date-time series.
}
\author{
Paul J. Gordijn
only work with the full data set if remove_prompt is T
}
