% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_sub_rici.r
\name{rici}
\alias{rici}
\title{Data extractor by row and column index}
\usage{
rici(ri = NULL, ci = NULL, setup_name = NA, ...)
}
\arguments{
\item{ri}{The row index provided as an integer.}

\item{ci}{The column index provided as an integer.}

\item{setup_name}{The name of the item/data found at the defined index. A   character string.}
}
\value{
A vector of length two containing interger indicies for a row and   column, respectively.
}
\description{
A predefined function to return the row and column index in a standardised way, that is, providing the row, then column index, in that order, respectively, in a vector of length two.
}
\details{
This is an internal function not meant to be used as stand alone   feature.
}
\author{
Paul J. Gordijn
}
