% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_sub_sf_dta_chunkr.r
\name{sf_dta_chunkr}
\alias{sf_dta_chunkr}
\title{Chunks data by time period}
\usage{
sf_dta_chunkr(
  dta_room = NULL,
  dta_sets = NULL,
  tn = NULL,
  rit = NULL,
  ri = NULL,
  chunk_i = NULL,
  rechunk = FALSE,
  buff_period = "10 years",
  i_zeros = 5,
  verbose = FALSE,
  xtra_v = FALSE,
  ...
)
}
\arguments{
\item{dta_room}{File path to the directory where the data will be and/or is chunked and indexed.}

\item{dta_sets}{List of data sets, belonging to the same data series to chunk in the \code{dta_room}.}

\item{tn}{Name of the data table to be chunked. The tables data time-series qualities must be consistent with ipayipi's 'continuous', 'event_based', or 'mixed' series qualities.
_Time-series properties:
Time-series properties are not required. If not provided this will be extracted from data summary information, or via \code{ipayipi::record_interval_eval()}.}

\item{rit}{The record interval type: 'continuous', 'mixed', or 'event_based'.}

\item{ri}{The record interval of the data. A string that will be parsed to \code{ipayipi::sts_interval_name()}. This is used to evaluate the integrity of data chunks time-series consistency; missing date-time values will be filled in as NA values. This applies to mixed and continuous data streams, not event-based where there is no consistent record interval. If the record interval is not provided the function will attempt to define this. If this fails the function will use a default chunking index of two years.}

\item{chunk_i}{The chunking interval. Must be a string representing a time period. The string is standardised by \code{ipayipi::sts_interval_name()}. If no chunk interval is provided the function will estimate an appropriate chunking interval based on the record interval.}

\item{buff_period}{Time periods represented by a character string used to buffer the min and max date-times around index creation. Should ideally be at least 10 years. If new data extends before or beyond the index dates then data will be rechunked, and a new index generated.}

\item{i_zeros}{Used to name chunked files; the number of leading zeros to include before a chunk number index.}

\item{verbose}{Logical. Whether or not to report messages and progress.}

\item{cores}{Number of cores to farm processing off to. This assists with increasing the speed of read-write capabilities dealing with chunks.}
}
\description{
Chunks in indexes a list of data sets belonging to the same time-series or table into a specified location. Data must have a 'date_time' column---used for chunking.
}
\details{
This function is an internal function called by others in the pipeline. Its main funciton is to standardise how data is chunked and maintain time-series data integrity. Chunking is done per a 'chunking index' that seperates data chunks by a date floor and ceiling. The index number serves as the identifier for a particular chunk. In addition to the chunk table index the overall min and max date-time of the data series is provided in an index list.
}
\author{
Paul J Gordijn
}
\keyword{big}
\keyword{chunking;}
\keyword{data}
\keyword{data;}
\keyword{indexing;}
