% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_sub_sf_dta_chunkr_sub_rechnk.r
\name{sf_dta_chunkr_sub_rechnk}
\alias{sf_dta_chunkr_sub_rechnk}
\title{Rechunks data}
\usage{
sf_dta_chunkr_sub_rechnk(
  dta_room = NULL,
  dts_min = NULL,
  dts_max = NULL,
  indx = NULL,
  chunk_i = NULL,
  rechunk = FALSE,
  buff_period = "10 years",
  i_zeros = 5,
  verbose = TRUE,
  xtra_v = FALSE,
  ...
)
}
\arguments{
\item{dta_room}{File path to the directory where the data will be and/or is chunked and indexed.}

\item{chunk_i}{The chunking interval. Must be a string representing a time period. The string is standardised by \code{ipayipi::sts_interval_name()}. If no chunk interval is provided the function will estimate an appropriate chunking interval based on the record interval.}

\item{i_zeros}{Used to name chunked files; the number of leading zeros to include before a chunk number index.}

\item{verbose}{Logical. Whether or not to report messages and progress.}

\item{dta_sets}{List of data sets, belonging to the same data series to chunk in the \code{dta_room}.}

\item{tn}{Name of the data table to be chunked. The tables data time-series qualities must be consistent with ipayipi's 'continuous', 'event_based', or 'mixed' series qualities.}

\item{rit}{The record interval type: 'continuous', 'mixed', or 'event_based'.}

\item{ri}{The record interval of the data. A string that will be parsed to \code{ipayipi::sts_interval_name()}. This is used to evaluate the integrity of data chunks time-series consistency; missing date-time values will be filled in as NA values. This applies to mixed and continuous data streams, not event-based where there is no consistent record interval. If the record interval is not provided the function will attempt to define this. If this fails the function will use a default chunking index of two years.}

\item{cores}{Number of cores to farm processing off to. This assists with increasing the speed of read-write capabilities dealing with chunks.}
}
\description{
If the chunk interval of data needs to be changed. This is the function that handles it. Date are chunked
}
\author{
Paul J Gordijn
}
\keyword{big}
\keyword{chunking;}
\keyword{data}
\keyword{data;}
\keyword{indexing;}
