% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_sub_sf_dta_chunkr_sub_ri.r
\name{sf_dta_chunkr_sub_ri}
\alias{sf_dta_chunkr_sub_ri}
\title{Chunks data by time period}
\usage{
sf_dta_chunkr_sub_ri(
  dta_room = NULL,
  indx = NULL,
  dta_sets = NULL,
  rit = NULL,
  ri = NULL,
  chunk_i = NULL,
  spchr = NULL,
  rechunk = FALSE,
  verbose = TRUE,
  xtra_v = FALSE,
  ...
)
}
\arguments{
\item{dta_room}{File path to the directory where the data will be and/or is chunked and indexed.}

\item{dta_sets}{List of data sets, belonging to the same data series to chunk in the \code{dta_room}.}

\item{rit}{The record interval type: 'continuous', 'mixed', or 'event_based'.}

\item{ri}{The record interval of the data. A string that will be parsed to \code{ipayipi::sts_interval_name()}. This is used to evaluate the integrity of data chunks time-series consistency; missing date-time values will be filled in as NA values. This applies to mixed and continuous data streams, not event-based where there is no consistent record interval. If the record interval is not provided the function will attempt to define this. If this fails the function will use a default chunking index of two years.}

\item{chunk_i}{The chunking interval. Must be a string representing a time period. The string is standardised by \code{ipayipi::sts_interval_name()}. If no chunk interval is provided the function will estimate an appropriate chunking interval based on the record interval.}

\item{verbose}{Logical. Whether or not to report messages and progress.}

\item{i_zeros}{Used to name chunked files; the number of leading zeros to include before a chunk number index.}
}
\description{
Chunks in indexes a list of data sets belonging to the same time-series or table into a specified location. Data must have a 'date_time' column---used for chunking.
}
\details{
This function is an internal function called by others in the pipeline. Its main funciton is to standardise how data is chunked and maintain time-series data integrity. Chunking is done per a 'chunking index' that seperates data chunks by a date floor and ceiling. The index number serves as the identifier for a particular chunk. In addition to the chunk table index the overall min and max date-time of the data series is provided in an index list.
}
\author{
Paul J Gordijn
}
\keyword{big}
\keyword{chunking;}
\keyword{data}
\keyword{data;}
\keyword{indexing;}
