% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_sub_sf_dta_read.r
\name{sf_dta_read}
\alias{sf_dta_read}
\title{Read station data}
\usage{
sf_dta_read(
  sfc = NULL,
  tv = NULL,
  pipe_house = NULL,
  station_file = NULL,
  tmp = TRUE,
  return_dta = FALSE,
  verbose = TRUE,
  start_dttm = NULL,
  end_dttm = NULL,
  xtra_v = FALSE,
  ...
)
}
\arguments{
\item{sfc}{Station file connection gerenated using \code{ipayipi::open_sf_con()}. If \code{sfc} is provided both \code{pipe_house} and \code{station_file} arguments are optional.}

\item{tv}{String vector of the names of values or tables to read from the station file. The strings provided here are used to filter items in the \code{sfc} object. The filtered item is then read into memory.}

\item{pipe_house}{List of pipeline directories. \strong{See} \code{ipayipi::ipip_house()} \strong{for details}.}

\item{station_file}{The file name of the station file (extension included).}

\item{tmp}{Logical. If TRUE then \code{sf_dta_read()} reads from the sessions temporary file location for stations.}

\item{return_dta}{Set to \code{TRUE} to return data files in addition to the data index; applies to chunked data tables.}

\item{verbose}{Logical. Whether or not to report messages and progress.}

\item{start_dttm}{Date-time filter. Applies to data tables with a 'date-time' column.}

\item{end_dttm}{Date-time filter. Applies to data tables with a 'date-time' column.}
}
\value{
Station data tables corresponding to the input table names(\code{tv}), and chunked data indices with data (if \code{return_dta} is \code{TRUE}). If no matching table is found then \code{NULL} is returned.
}
\description{
Reads station data from chunked, or other station data files.
}
\details{
Used to handle reading ipayipi station data from a variety of station file formats.
}
\author{
Paul J Gordijn
}
\keyword{and}
\keyword{chunked}
\keyword{data;}
\keyword{read}
\keyword{station}
\keyword{tables}
