% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_sub_sf_read.r
\name{sf_read}
\alias{sf_read}
\title{Simple data table reading function.}
\usage{
sf_read(
  sfc = NULL,
  tv = NULL,
  station_file = NULL,
  tmp = FALSE,
  pipe_house = NULL,
  verbose = TRUE,
  start_dttm = NULL,
  end_dttm = NULL,
  ...
)
}
\arguments{
\item{sfc}{Station file connection gerenated using \code{ipayipi::open_sf_con()}. If \code{sfc} is provided both \code{pipe_house} and \code{station_file} arguments are optional.}

\item{tv}{String vector of the names of values or tables to read from the station file. The strings provided here are used to filter items in the \code{sfc} object. The filtered item is then read into memory.}

\item{station_file}{The file name of the station file (extension included).}

\item{tmp}{Logical. If TRUE then \code{sf_read()} reads from the sessions temporary file location for stations.}

\item{pipe_house}{List of pipeline directories. \strong{See} \code{ipayipi::ipip_house()} \strong{for details}.}

\item{verbose}{Logical. Whether or not to report messages and progress.}
}
\value{
Station file connection file paths. To be used in the pipeline.
}
\description{
Package subroutine for opening a station file for work. NB! Will not work for chunked data tables.
}
\details{
This function is an internal function called by others in the
pipeline.
}
\author{
Paul J Gordijn
}
\keyword{data}
\keyword{data;}
\keyword{lists}
\keyword{open}
\keyword{read}
\keyword{tables;}
