% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gw_g_sol_to_excel.r
\name{sol_to_csv}
\alias{sol_to_csv}
\title{Exports R solonist data and transformations to csv format}
\usage{
sol_to_csv(
  input_dir = ".",
  output_dir = ".",
  recurr = FALSE,
  prompted = TRUE,
  ...
)
}
\arguments{
\item{input_dir}{Directory in which to search for R solonist objects to
export as csv files.}

\item{output_dir}{Directory to which the csv files will be exported.}

\item{recurr}{Searches recursively in the input directory for R solonist
objects when set to TRUE.}

\item{prompted}{If TRUE (as default), then the user has the option to
interactively select the R solonist objects.}
}
\value{
Exports csv files.
}
\description{
Function that exports selected solonist R objects to csv tables.
}
\author{
Paul J Gordijn
}
