% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_sub_sts_interval_name.r
\name{sts_interval_name}
\alias{sts_interval_name}
\title{Evaluates the consistency of record intervals}
\usage{
sts_interval_name(intv = NULL, ...)
}
\arguments{
\item{intv}{String describing time duration or 'difftime' object.}
}
\value{
Tabulated details on the time duration.
}
\description{
Standardises the string and provides a numerical equivalent
to the time period in seconds.
}
\details{
Function used internally. Note that if the input is in months
the returned time in seconds is based on a 28 day month.
}
\author{
Paul J. Gordijn
}
