% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sts_phen_measure}
\alias{sts_phen_measure}
\title{Standardesed phenomena or variable measures}
\format{
\subsection{'sts_phen_measure'}{

\describe{
\item{phen_prop}{The 'proper' or standardised shortname used to denote
a type of measure}
\item{phen_syn}{Measure synonyms. Multiple arguments are seperated by
a bar character.}
\item{phen_desc}{Full description of measure.}}
}
}
\source{
Internal.
}
\usage{
sts_phen_measure
}
\description{
Table of measure types with synonyms and descriptions.
}
\keyword{datasets}
