% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sts_phen_var_type}
\alias{sts_phen_var_type}
\title{Standardised phenomena or variable types}
\format{
\subsection{'sts_phen_var_type'}{

A data.table with two columns:
\describe{
\item{phen_prop}{The 'proper'/standardised shortname used for a
variable type}
\item{phen_syn}{A synonym of the variable type. These are the keywords
used by \code{ipayipi} to match up and standardise to the proper variable
name}
}
}
}
\source{
Internal.
}
\usage{
sts_phen_var_type
}
\description{
Table of short names describing variable types to aid standardisation
and definition of phenomena/variable types in \code{ipayipi}. Short names
are similar to that used by the \code{tidyverse}.
}
\keyword{datasets}
