% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_transfer_sts_files.R
\name{transfer_sts_files}
\alias{transfer_sts_files}
\title{Transfer standardised ipayipi data files to the 'd3_nomvet_room'}
\usage{
transfer_sts_files(
  pipe_house = NULL,
  limiting_fields = c("location", "station", "stnd_title", "start_dttm", "end_dttm",
    "record_interval_type", "record_interval", "table_name"),
  verbose = FALSE,
  xtra_v = FALSE,
  ...
)
}
\arguments{
\item{pipe_house}{List of pipeline directories. \emph{See} \code{\link[=ipip_house]{ipip_house()}} \emph{for details}.}

\item{verbose}{Logical. Print some details and progress of function progress?}

\item{xtra_v}{Logical. Should some 'x'tra messaging be done? Use to help diagnose problems, and for guidance.}
}
\value{
Saves files in native R format for 'ipayipi' in set folder. Returns a list of file names which were successfully transferred.
}
\description{
Moves files from the 'd2_wait_room' to the 'd3_nomvet_room'---which contains standardised logger data files. This function avoids duplicating files in the 'd3_nomvet_room' and ensures file names are unique. Files will be named using their station and table names, their date coverage, and an integer (that is unique per data file & respective date-time stamps).
}
\details{
Aim: Move standardised data files into the '\strong{nomvet_room}' --- part of the 'ipayipi'  data pipeline. Once moved, the data will be used for further processing. Here are some of the specific steps of this funtion:
\enumerate{
\item Generate inventories of standardised files (with file extension ".ipi"`) that have been processed by \code{\link[=nomenclature_sts]{nomenclature_sts()}} and \code{\link[=phenomena_sts]{phenomena_sts()}} in the 'd3_nomvet_room'.
\item Determine which files need to be transferred (to avoid duplication\code{*}) into the 'd3_nomvet_room'.
\item Transfer files---saving these in RDS format in the 'd3_nomvet_room' with the ".ipi" file extension.
}

\code{*} Duplicate files are determined using the \code{limiting_fields} parameter. The \code{limiting_fields} string vector is used to set the columns or fields of each files \code{data_summary} table used to ascertain duplicate status. Duplicate files will not be transfered to the '\strong{nomvet_room}' --- rather they will be moved to z folder called 'zdups' in the pipe house working directory, i.e., \code{pipe_house$pipe_house_dir} (\strong{see} \code{\link[=ipip_house]{ipip_house()}}).

A \emph{aa_nomvet.log} retains the first (and only) row of each files \code{data_summary}. This log is queried to determine which files are duplicates before moving files from the '\strong{wait_room}' to the '\strong{nomvet_room}'. NB! Do not delete files in the '\strong{nomvet_room}' manually without updating the \emph{aa_nomvet.log}.
}
\author{
Paul J. Gordijn
}
\keyword{Time}
\keyword{batch}
\keyword{data}
\keyword{data;}
\keyword{files;}
\keyword{processing;}
\keyword{saving}
\keyword{series}
\keyword{standardisation;}
