% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_sub_win_rows.r
\name{win_rows}
\alias{win_rows}
\title{Window width determination}
\usage{
win_rows(
  dta = NULL,
  rng_side = 1,
  rng_max = 3,
  c_row = NULL,
  pos = "center",
  include_center = TRUE,
  ...
)
}
\arguments{
\item{dta}{The vector data series of interest.}

\item{rng_side}{The desired size of the window side---this is
adjusted}

\item{rng_max}{If there are no data values in the input vectors
then the window side size increases to a maximum size of
\emph{rng_max}.}

\item{c_row}{Centre position in the vector around which to construct the
window.}

\item{pos}{Optional string describing the position of the window relative
to the centre row. "centre" attempts to put the \emph{c_row} in the centre.
"lead" attempts to put the window before (and including the c_row) and
"trail" attempts to do this after the \emph{c_row}.}

\item{include_center}{If the center point is not included the range is
effectively reduced by one.}
}
\value{
Indicies relating to a window centre and sizes.
}
\description{
returns a customized window for a given vector.
}
\details{
This function is an internal function called by others in the
pipeline.
}
\author{
Paul J Gordijn
}
\keyword{function}
\keyword{internal}
\keyword{size;}
\keyword{window}
