% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_sub_dt_agg_param_eval.r
\name{agg_param_eval}
\alias{agg_param_eval}
\title{'dt' processing pipeline: genearte aggregation criteria.}
\usage{
agg_param_eval(
  full_eval = FALSE,
  f_summary = NULL,
  f_params = NULL,
  agg_offset = c("0 secs", "0 secs"),
  all_phens = FALSE,
  ignore_nas = FALSE,
  eval_seq = NULL,
  ppsij = NULL,
  sf = NULL,
  ...
)
}
\arguments{
\item{full_eval}{Logical indicating whether to perform a full evaluation
or not. A partial evaluation only produces a string single string for
a summary of the function's parameters. Defaults to \code{FALSE}.}

\item{f_summary}{Summary information used in full evaluation. Includes
any tables in the station file with keywords, "phens" or "summary".}

\item{f_params}{For the partial evaluation multiple phenomena can
be described using \code{ipayipi::agg_params()}. If left blank defaults will
be used on all phenomena harvested.}

\item{agg_offset}{Character vector of length two. Strings describe period
of offset from the rounded time interval used to aggregate data. For
example, if aggregating rainfall data from five minute to daily records,
but staggered so that daily rainfall is totalled from 8 am to 8 pm the
next day the \code{agg_offset} must be set to \code{c("8 hours", "8 hours")}.
When \code{full_eval == TRUE} the function overwrites this parameter in favor
of \code{agg_offset} extracted from the partial evaluation.}

\item{all_phens}{Defaults to \code{FALSE}. If \code{TRUE}, all harvested phenomena will
be aggregated, despite the lesser number of phenomena listed/described in
the partial evaluation.}

\item{ignore_nas}{If \code{NA} values are ignores any periods or time
interval of the data with any NA values will result in an \code{NA}
aggregation value. Important if missing values have not been
interpolated.}

\item{eval_seq}{Dynamic list of tables accessed throughout the evaluation
process. The final evaluation tables are appended to the station file.}

\item{sf}{Station file object.}

\item{phens_summary}{A summary of the phens table of the particular
phenomena that will be aggregated by \code{ipayipi::dt_agg()}.}
}
\description{
'dt' processing pipeline: genearte aggregation criteria.
}
\details{
Default aggregation functions based on measure, variable types, and
the unit of measurement. Function info is stored in the package
data, "ftbl" --- \emph{see examples}.
}
\examples{
# load the table from the function data folder to see default aggregation
# functions
data("ftbl")
print(ftbl)

}
\author{
Paul J. Gordijn
}
