% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_sub_dt_agg_gen.r
\name{aggs}
\alias{aggs}
\title{Set custom aggregation parameters}
\usage{
aggs(...)
}
\value{
List custom aggregation parameters for each phenomena listed.
}
\description{
Used to generate a list of cusstom aggregation parameters, that is, \code{agg_params} using ipayipi::agg_params().
}
\details{
\code{agg()} is an \code{ipayipi} alias for \code{list()}, that will only include elements generated using \code{ipayipi::agg_params()}. Aggregation parameters provided here will override default aggregation parameters. The \code{ipayipi::sts_agg_functions} table stores default aggregation functions for different phenomena types.
}
\author{
Paul J. Gordijn
}
