% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_sub_append_phen_data.r
\name{append_phen_data}
\alias{append_phen_data}
\title{Appends data tables by phenomena}
\usage{
append_phen_data(
  old_dta = NULL,
  old_phen_ds = NULL,
  new_dta = NULL,
  new_phen_ds = NULL,
  phen_dt = NULL,
  overwrite_old = FALSE,
  tn = NULL,
  ...
)
}
\arguments{
\item{phen_dt}{Phenomena table of from both the station and new data files
combined. These are combined and phids standarised in
ipayipi::station_append().}

\item{overwrite_old}{If \code{TRUE} then original data is disgarded
in favour of new data. If TRUE both data sets will be evaluated and where
there are NA values, a replacement, if available, will be used to replace
the NA value. Defaults to FALSE.}

\item{station_data_table}{Data from the station file.}

\item{new_data_table}{Data from the new data file.}
}
\value{
A phenomena table that contains the start and end dates of each
phenomena and the respective data source identification numbers of the
data. Also, the appended data.
}
\description{
This function is for internal use. The function is designed to
append phenomena data and associated phenomena standards efficiently, but
retaining a metadata records for each phenomena. Moreover, the function
evaluates missing data and compares new and older data records so avoid
loosing data.
}
\details{
This function is an internal function called by others in the
pipeline.
}
\author{
Paul J Gordijn
}
\keyword{append}
\keyword{data,}
\keyword{join}
\keyword{overwrite}
\keyword{phenomena}
\keyword{tables,}
