% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_sub_append_phen_data.R
\name{append_phen_data}
\alias{append_phen_data}
\title{Appends data table phenomena by phenomena}
\usage{
append_phen_data(
  pipe_house = NULL,
  station_file = NULL,
  sf_phen_ds = NULL,
  ndt = NULL,
  new_phen_ds = NULL,
  phen_id = FALSE,
  phen_dt = NULL,
  overwrite_sf = FALSE,
  tn = NULL,
  ri = NULL,
  rit = NULL,
  verbose = verbose,
  xtra_v = xtra_v,
  ...
)
}
\arguments{
\item{pipe_house}{List of pipeline directories. \strong{See} \code{ipayipi::ipip_house()} \strong{for details}.}

\item{station_file}{Path and filename of the station file or station file object.}

\item{sf_phen_ds}{Optional argument. Phenomena summary from the station file (developed by the data pipeline).}

\item{ndt}{Data from the new-data file.}

\item{new_phen_ds}{Phenomena summary for the new data.}

\item{phen_id}{Set this to TRUE for generating updated station file phenomena data summaries. When set to FALSE, a new phenomena data summary will not be executed. Set to FALSE if no phenomena data summaries are available.}

\item{phen_dt}{Phenomena table of from both the station and new data files combined. These are combined and phids standarised in ipayipi::station_append().}

\item{overwrite_sf}{If \code{TRUE} then original data is disgarded in favour of new data. If TRUE both data sets will be evaluated and where there are NA values, a replacement, if available, will be used to replace the NA value. Defaults to FALSE.}

\item{tn}{The name of the phenomena data tables.}

\item{ri}{The record interval associated with the data sets. String standardised using \code{ipayipi::sts_interval_name()}.}

\item{rit}{Record interval type. One of the following options for time-series data: 'continuous', 'event_based', or 'mixed'.}
}
\value{
A phenomena table that contains the start and end dates of each
phenomena and the respective data source identification numbers of the
data. Also, the appended data.
}
\description{
This function is for internal use. The function is designed to append phenomena data and associated phenomena standards efficiently, but retaining phenomena associated metadata records. Moreover, the function evaluates missing data and compares new and older data records so avoid   loosing data.
}
\details{
This function is an internal function called by others in the
pipeline.
}
\author{
Paul J Gordijn
}
\keyword{append}
\keyword{data,}
\keyword{join}
\keyword{overwrite}
\keyword{phenomena}
\keyword{tables,}
