% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_sub_append_phen_overlap_data.r
\name{append_phen_overlap_data}
\alias{append_phen_overlap_data}
\title{Appends data tables by phenomena}
\usage{
append_phen_overlap_data(
  ndt = NULL,
  nd_min = NULL,
  nd_max = NULL,
  phen_id = FALSE,
  phens = NULL,
  phen_dt = NULL,
  sf_phen_ds = NULL,
  new_phen_ds = NULL,
  ri = NULL,
  rit = NULL,
  tn = NULL,
  overwrite_sf = FALSE,
  sf_eindx = NULL,
  verbose = FALSE,
  xtra_v = xtra_v,
  ...
)
}
\arguments{
\item{ndt}{Data from the new data file (or 'y' or right table).}

\item{nd_min}{Minimum date-time of the 'new data' being added to the station file.}

\item{nd_max}{Maximum date-time of the 'new data' being added to the station file.}

\item{phen_id}{Set this to TRUE for generating updated station file phenomena data summaries. When set to FALSE processing to produce this data summary will not be executed.}

\item{phens}{Ordered names of phenomena to include in the joined table.}

\item{phen_dt}{Phenomena table of both the station and new data files combined. These are combined and phids standarised in \code{ipayipi::station_append()}.}

\item{sf_phen_ds}{Phenomena summary from the station file.}

\item{new_phen_ds}{Phenomena summary for the new data.}

\item{ri}{The record interval associated with the data sets. As a standardised string. Used to generate a time-series sequence using base::seq().}

\item{rit}{Record interval type. One of the following options for time-series data: 'continuous', 'event_based', or 'mixed'.}

\item{tn}{The name of the phenomena data tables. Argument only required for processing phenomena summary data.}

\item{overwrite_sf}{Logical. If TRUE the station file or 'x' table values, if not NA, will be overwritten by values from the 'y' or new data table.}

\item{sfo}{Overlapping phenomena data table from the 'x' (left) table or station file.}

\item{cores}{Integer representing the number of cores available for processing data. Multiple core usage only supported on linux systems.}
}
\value{
List of useful items for internal use. Overap data.
}
\description{
Custom full join of overlapping time series data. The function is designed to append phenomena data and associated phenomena standards efficiently, and can retain temporal metadata records for each phenomena if a phenomena summary data table is provided. Missing data is evaluated to avoid loosing data.
}
\details{
This function is an internal function called by \code{ipayipi::append_phen_data()}.
}
\author{
Paul J Gordijn
}
\keyword{append}
\keyword{data,}
\keyword{join}
\keyword{overwrite}
\keyword{phenomena}
\keyword{tables,}
