% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/met_f_station_append.r, R/z_e_station_append.r
\name{append_station}
\alias{append_station}
\title{Append meteorological station data}
\usage{
append_station(
  pipe_house = NULL,
  station_file = NULL,
  new_data = NULL,
  overwrite_sf = TRUE,
  by_station_table = FALSE,
  phen_id = TRUE,
  cores = getOption("mc.cores", 2L),
  ...
)

append_station(
  pipe_house = NULL,
  station_file = NULL,
  new_data = NULL,
  overwrite_sf = TRUE,
  by_station_table = FALSE,
  phen_id = TRUE,
  cores = getOption("mc.cores", 2L),
  ...
)
}
\arguments{
\item{station_file}{The main data set to which data will be appended.}

\item{new_data}{The data that will be appended to the main 'hobo station'.}

\item{overwrite_sf}{Logical. Defaults to FALSE so that data from the
\code{station_file} is not over written by the \code{new_data}.}

\item{by_station_table}{If TRUE then multiple station tables will not be
kept in the same station file. Defaults to FALSE.}

\item{cores}{Number of CPU's to use for processing in parallel. Only applies when working on Linux.}

\item{overwrite_old}{Logical. Defaults to FALSE so that data from the
\code{station_file} is not over written by the \code{new_data}.}
}
\description{
Appends two standardised 'ipayipi' meteorological' station
files.

Appends two standardised 'ipayipi' meteorological' station
files.
}
\details{
Special attention is given to the \verb{data summary} table of
\code{SAEON_rain_data} objects. The data summary contains information
which is used in the pipeline to determine which files are appended to
another. Specifically, the standardised ptitle in the data summary is the
identifier for stations, and records in the \code{nomvet_name} field are used to
determine which files have been included in any given \code{SAEON_rain_data}
object.
}
\author{
Paul J. Gordijn
}
\keyword{append}
\keyword{data}
\keyword{data;}
\keyword{logger}
\keyword{meteorological}
\keyword{pipeline;}
