% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_e_append_station.r
\name{append_station}
\alias{append_station}
\title{Append meteorological station data using chunck reads---version 3}
\usage{
append_station(
  pipe_house = NULL,
  station_file = NULL,
  new_data = NULL,
  overwrite_sf = TRUE,
  by_station_table = FALSE,
  phen_id = TRUE,
  verbose = FALSE,
  xtra_v = FALSE,
  ...
)
}
\arguments{
\item{station_file}{The main data set to which data will be appended.}

\item{new_data}{The data that will be appended to the main 'hobo station'.}

\item{overwrite_sf}{Logical. Defaults to FALSE so that data from the
\code{station_file} is not over written by the \code{new_data}.}

\item{by_station_table}{If TRUE then multiple station tables will not be
kept in the same station file. Defaults to FALSE.}

\item{cores}{Number of CPU's to use for processing in parallel. Only applies when working on Linux.}
}
\description{
Appends two standardised 'ipayipi' meteorological' station files.
}
\author{
Paul J. Gordijn
}
\keyword{append}
\keyword{data}
\keyword{data;}
\keyword{logger}
\keyword{pipeline;}
