% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_e_station_append_batch.r
\name{append_station_batch}
\alias{append_station_batch}
\title{Append meteorological station data}
\usage{
append_station_batch(
  ipip_room = NULL,
  nomvet_room = NULL,
  overwrite_old = FALSE,
  by_station_table = FALSE,
  station_ext = ".ipip",
  sts_file_ext = ".ipi",
  ...
)
}
\arguments{
\item{overwrite_old}{If \code{TRUE} then original data (station data) is
disgarded in favour of new data. If TRUE both data sets will be evaluated
and where there are NA values, a replacement, if available, will be used
to replace the NA value. Defaults to FALSE.}

\item{station_file}{The main data set to which data will be appended.}

\item{new_data}{The data that will be appended to the main 'hobo station'.}
}
\description{
Appends two standardised 'ipayipi' meteorological' station
files.
}
\details{
This function searches for all standardised files in the \code{nomvet_room}
and existing 'station files' in the \code{ipip_room}. Files from the
\code{nomvet_room} that have not been appended to station files will be
appended to appropriate stations. Appending is done via the
\code{ipayipi::append_station()}, \code{ipayipi::append_phen_data()}, and
\code{ipayipi::append_tables()} functions.
}
\author{
Paul J Gordijn
}
\keyword{append}
\keyword{data}
\keyword{data;}
\keyword{meteorological}
\keyword{pipeline;}
