% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_sub_dt_calc_param_eval.r
\name{calc_param_eval}
\alias{calc_param_eval}
\title{Organise sequencial processing of phenomena aggregations.}
\usage{
calc_param_eval(
  full_eval = FALSE,
  f_params = NULL,
  f_summary = NULL,
  ppsij = NULL,
  sf = sf,
  ...
)
}
\description{
An alias for list in ipayipi::dt_calc. Used for sequence of
\code{ipayipi::chainer()} calculations.
}
\author{
Paul J. Gordijn
}
