% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_dt_calc_chainer.R
\name{chainer}
\alias{chainer}
\title{Develop a 'data.table' chain link for processing data}
\usage{
chainer(
  i = NULL,
  j = NULL,
  measure = NULL,
  units = NULL,
  var_type = NULL,
  notes = NULL,
  fork_table = NULL,
  temp_var = FALSE,
  link = NULL,
  dt_syn_bc = NULL,
  dt_syn_ac = NULL,
  ...
)
}
\arguments{
\item{i}{\code{data.table} before first-comma syntax. Before the first comma syntax operates on a data.tables rows. Can be provided as a multiple length vector that will be pasted with no seperator into one string. Same for \code{j} below.}

\item{j}{\code{data.table} after first-comma syntax. After the first comma syntax operates on columns. \emph{See examples}.}

\item{measure}{If a new vatiable is described it's measure should be provided. Measure's are standardised using the \code{ipayipi::sts_phen_measure} standard table. New variables can be assigned using the ":=" operator.}

\item{units}{The unit of measure of a new variable, e.g., "mm" for rainfall, or "deg_c" for degrees Celcius. \emph{Notice conformity to \strong{snake case} descriptions --- this must be applied}.}

\item{var_type}{The type (or class) of variable, e.g., "num" for numeric. These are standardised using the \code{ipayipi::sts_phen_var_type} table.}

\item{notes}{An additional description of the calculation/operation can be provided.}

\item{fork_table}{If not \code{NULL} (default) a duplicate table of the reverse \code{dt_syn_bc} (before comma logic) will be produced and kept in the data pipeline.}

\item{temp_var}{Indicates whether an introduced variable is temporary. If temporary it will be removed during the processing stage.}

\item{link}{Integer denoting the numeric order of a chain link in a sequence put together in a list.}
}
\description{
Link together multiple data.table operations into a single chain.
}
\details{
\itemize{
\item New phenomena (variables):
In order to describe phenomena details for further processing it is important to provide metadata as they are introduced. Each time the data.table operator ':=' is used this serves as a que that a new vairable has been added to the processing pipeline. For this, respective 'measure', 'units', and 'var_type' properties must be described.
}

NB! This parameter \code{dt_syn_ac} and \code{dt_syn_bc} are being phased out in favor of \code{i} and \code{j}.
}
\examples{
# Set up chain link to filter out zero mm rainfall events
x <- list(remove0 = chainer(i = "rain_cumm != 0"))
x
# Two stage chaining to filter out zero mm rainfall events then calculate
# sequential time lag (assign a variable called 't_lag').
# Note use of data.table's .N (number of rows).
x <- list(
 remove0 = chainer(i = "rain_cumm != 0"),
 t_lag = chainer(
   j = "t_lag := c(0, date_time[2:.N] - date_time[1:(.N - 1)])",
   measure = "smp", units = "time", var_type = "num"
 )
)
print(x)

# Transform chain links in to data.table syntax
calc_param_eval(x)

}
\author{
Paul J. Gordijn
}
