% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_dt_clean_param_eval.R
\name{clean_param_eval}
\alias{clean_param_eval}
\title{Evaluate arguments for pipeline univariate filter.}
\usage{
clean_param_eval(
  phens = NULL,
  w = 21,
  madf = 3,
  align = "left",
  seg = NA_character_,
  cush = TRUE,
  clean_f = "hampel",
  na_t = 0.25,
  tighten = 0.65,
  owrite = TRUE,
  full_eval = FALSE,
  station_file = NULL,
  ppsij = NULL,
  f_params = NULL,
  sfc = NULL,
  verbose = FALSE,
  xtra_v = FALSE,
  ...
)
}
\arguments{
\item{phens}{List of vectors of phenomena names. Each list item must correspond to a filter 'run'. This allows multiple runs to be performed whilst data is in memory with different parameters being passed to the filter algorithm. If NULL the filter will run on all numeric phenomena. To 'deselect' phenomena from the run include a minus directly before the phenomena name.}

\item{w}{Total window size for the filter algorithm. Defaults to ten. Must be supplied as a vector or list with subsequent values corresponding to each filter 'run'.}

\item{madf}{Scalar factor of MAD (median absolute deviation). Higher values relax oulier detection. Defaults to the standard of three.  Must be supplied as a vector or list with subsequent values corresponding to each filter 'run'.}

\item{align}{One of the following character options following 'data.table' lateral syntax framework: 1) \bold{'right'} (the default) evaluates the current value, on the right, based on preceeding values, falling to the left; option \bold{'left'} does the opposite evaluting the values to the right; option \bold{centre} places the evauated value in middle of the total window size provided by \bold{w} (option 'centre' is 'left' biased for odd window sizes). This argument is parsed to 'data.table' \code{frollapply}.}

\item{seg}{Vector of station data table names that contain values that will be used to slice data series into segments for independent outlier detection runs. Values don't have to correspond to segment numbers---segments are orgaised within this function. If \code{NULL} (default) the series is treated as one segment. An option here is the 'logg_interfere' table. Must be supplied as a vector or list with subsequent values corresponding to each filter 'run'.}

\item{cush}{Logical indicating whether to prevent NAs at data start and end dates owing to window/filter alignment. \code{TRUE} will \bold{cush}ion data start and ends where necessary by adjust window alignment properties (within segments). \code{FALSE} does not prevent NA values.}

\item{clean_f}{Algorithm name. Only the non-linear "hampel" filter currently supported.}

\item{na_t}{Threshold (as a proportion of window length) of NA values tolerated. If the number of NAs exceeds the threshold, filtering will not be performed. Must be supplied as a vector or list with subsequent values corresponding to each filter 'run'.}

\item{tighten}{Scaling fraction between zero and one that sensitizes the detection of outliers near the head and tail ends of segments. The fraction is multiplied by the mad_dev factor. Must be supplied as a vector or list with subsequent values corresponding to each filter 'run'.}

\item{full_eval}{Logical that indicates the depth of evaluation. This argument is provided parsed internally. Full evaluation can only be completed within \code{ipayipi::dt_process()} sequence evaluation.
param f_params}

\item{station_file}{Name of the station being processed.}

\item{ppsij}{Data processing \code{pipe_seq} table from which function parameters and data are extracted/evaluated. This is parsed to this function automatically by \code{ipayipi::dt_process()}.}

\item{sfc}{List of file paths to the temporary station file directory. Generated using \code{ipayipi::sf_open_con()}.}

\item{verbose}{Logical. Whether or not to report messages and progress.}

\item{xtra_v}{Logical. Whether or not to report xtra messages, progess, plus print data tables.}
}
\description{
The default filter used is the 'hampel' non-linear, non-parametric filter, i.e., it application is suited to time series data. The 'hampel' is the only filter currently implemented.
}
\details{
\itemize{
\item \code{cush}: When \code{align} is set to \code{left} at the 'end' of a data series values cannot be evaluated and therefore set to NA, and vice versa when \code{align} is \code{right}. NAs at data start and end extremities are avoided when \code{cush} is set to \code{TRUE} by switching \code{align} properties in these cases. When \code{align} is set to \code{center} both 'left' and 'right' alignment are used to fill NA values.
}
}
\author{
Paul J. Gordijn
}
