% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_sub_dt_clean_param_eval.R
\name{clean_param_eval}
\alias{clean_param_eval}
\title{Evaluate and prepare tables for joins}
\usage{
clean_param_eval(
  clean_f = "hampel",
  phen_names = NULL,
  segs = NULL,
  seg_fuzz = NULL,
  seg_na_t = 0.75,
  w_size = 21,
  mad_dev = 3,
  tighten = 0.65,
  full_eval = FALSE,
  station_file = NULL,
  ppsij = NULL,
  f_params = NULL,
  sfc = NULL,
  verbose = FALSE,
  xtra_v = FALSE,
  ...
)
}
\arguments{
\item{clean_f}{Algorithm name. Currently only the non-linear "hampel" filter supported.}

\item{phen_names}{List of vectors of phenomena names. Each list item must correspond to a filter 'run'. This allows multiple runs to be performed whilst data is in memory with different parameters being passed to the filter algorithm. If NULL the filter will run on all numeric phenomena. To 'deselect' phenomena from the run include a minus directly before the phenomena name.}

\item{segs}{Vector of station data table names that contain timestamps used to slice data series into segments for independent outlier detection runs. If left \code{NULL} (default) the series is treated as one segment. A good option here is the 'logg_interfere' table. Must be supplied as a vector or list with subsequent values corresponding to each filter 'run'.}

\item{seg_fuzz}{String representing the threshold time interval between the list of segment date-time values. If two or more date-time values occur within this threshold time there corresponding values will be merged to the earliest date-time stamp. If left \code{NULL} \code{seg_fuzz} defaults to the record interval duration. Must be supplied as a vector or list with subsequent values corresponding to each filter 'run'.}

\item{seg_na_t}{Fractional tolerace of the count of NA values within a window for interpolation of missing values. If the number of NAs exceeds the threshold, filtering will not be performed. Must be supplied as a vector or list with subsequent values corresponding to each filter 'run'.}

\item{w_size}{Window size for the filter algorithm. Defaults to ten. Must be supplied as a vector or list with subsequent values corresponding to each filter 'run'.}

\item{mad_dev}{Scalar factor of MAD (median absolute deviation). Higher values relax oulier detection. Defaults to the standard of three.  Must be supplied as a vector or list with subsequent values corresponding to each filter 'run'.}

\item{tighten}{Scaling fraction between zero and one that sensitizes the detection of outliers near the head and tail ends of segments. The fraction is multiplied by the mad_dev factor. Must be supplied as a vector or list with subsequent values corresponding to each filter 'run'.}

\item{full_eval}{param f_params}

\item{station_file}{Name of the station being processed.
details}

\item{ppsij}{Data processing \code{pipe_seq} table from which function parameters and data are extracted/evaluated. This is parsed to this function automatically by \code{ipayipi::dt_process()}.
param sf}

\item{sfc}{List of file paths to the temporary station file directory. Generated using \code{ipayipi::open_sf_con()}.}

\item{verbose}{Logical. Whether or not to report messages and progress.}

\item{xtra_v}{Logical. Whether or not to report xtra messages, progess, plus print data tables.}
}
\description{
Flexible and fast joining of \code{ipayipi} data.
}
\author{
Paul J. Gordijn
}
