% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_f1_station_dt_agg.r
\name{dt_agg}
\alias{dt_agg}
\title{Aggregate time series data by a time interval}
\usage{
dt_agg(
  station_file = station_file,
  input_dt = NULL,
  output_dt = NULL,
  time_interval = NULL,
  f_params = NULL,
  f_summary = f_summary,
  ...
)
}
\arguments{
\item{input_dt}{Only 'ipayipi' raw data types supported. The function
relies on the data summaries and phenomena tables of 'station files'
for aggregating data.}

\item{output_dt}{Output table name. If \code{NULL} then generic names are
used. Preferable to use generic names for pipeline processing.}

\item{time_interval}{Table parsed by \code{ipayipi::sts_interval_name()}
desecribing the time interval for data aggregation.}

\item{f_params}{The special function parameters \emph{see}
\code{ipayipi::agg_params()} and \code{ipayipi::p_agg_dt()}.}
}
\description{
This function aggregates 'ipayipi' formatted data by
time periods.
}
\details{
If the \code{input_data} parameter is not supplied then the function
determines which raw data tables to extract data from. When this happens
The function first prioritises continuous raw data with record intervals
closest to the specified \code{time_interval}. If continuous data is not
available then the function uses discontinuous data. The function stores
which data has been extracted in the station_file object.

\emph{\verb{agg\\_f}}
The \emph{\verb{agg\\_f}} parameter can be passed as a \code{data.frame} or \code{data.table},
or as a list that will be coerced to a table. The data table should have
all required (*) fields as a minimum. If the empty fields cannot be
extracted from the station file phenomena summary these are set to
defaut values where possible.\tabular{llllllll}{
   *\code{phen_name} \tab \code{output_phen_name} \tab \code{agg_function} \tab \code{units} \tab \code{measure} \tab \code{var_type} \tab \code{na_ignore} \tab \code{table_name} \cr
   i \tab i \tab i \tab i \tab i \tab i \tab i \tab  \cr
   i + 1 \tab i + 1 \tab i + 1 \tab i + 1 \tab i + 1 \tab i + 1 \tab i + 1 \tab i + 1 \cr
}
}
\author{
Paul J. Gordijn
}
