% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_h1_station_dt_agg.r
\name{dt_agg}
\alias{dt_agg}
\title{Aggregate time series data by a time interval}
\usage{
dt_agg(
  station_file = station_file,
  dta_in = NULL,
  f_params = NULL,
  phens_dt = NULL,
  sf_tmp_fn = NULL,
  ppsij = ppsij,
  f_summary = NULL,
  sf = NULL,
  dt_working = NULL,
  agg_offset = c("0 sec", "0 sec"),
  ...
)
}
\arguments{
\item{station_file}{The file path of the station being assessed.}

\item{dta_in}{A list of data tables to be aggregated. The tables
must come from the same station, but may have different record intervals.}

\item{f_params}{Function parameters evaluated by
\code{ipayipi::agg_param_eval()}.}

\item{phens_dt}{The phenomena table describing each of the phenomena in the
\code{dta_in} parameter. This information is used in the aggregation process.}

\item{sf_tmp_fn}{A temporary file name used for the station file being
processed.}

\item{ppsij}{Table derived from \code{ipayipi::pipe_seq()} showing the summarised
function parameters.}

\item{f_summary}{List of tables with data summaries.}

\item{sf}{Station file object.}

\item{dt_working}{List of data tables in the processing pipeline.}

\item{agg_offset}{Character vector of length two. Strings describe period
of offset from the rounded time interval used to aggregate data. For
example, if aggregating rainfall data from five minute to daily records,
but staggered so that daily rainfall is totalled from 8 am to 8 pm the
next day the \code{agg_offset} must be set to \code{c("8 hours", "8 hours")}.}
}
\description{
This function aggregates 'ipayipi' formatted data by
time periods. The requries an evaluated pipeline.
}
\details{
Function has to work within the \code{ipayipi} pipeline as parameters
for aggregation are determined during the pipeline evaluation process.

Processing is done whilst keeping a minimum amount of data in memory. The
station file being worked on is saved in a temporary location so if any
errors occur in the processing the original file will not be overwritten.

If an offset on the aggregation is used then the secon value of the offset
is added to the 'date_time' stamp for the data.tables 'date_time' value.
Aggregation defaults are determined by \code{ipayipi::agg_param_eval()}. Default
aggregation functions are determined by the phenomena measure parameter ---
see the function for more details.
}
\author{
Paul J. Gordijn
}
