% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_h1_station_dt_calc.r
\name{dt_calc}
\alias{dt_calc}
\title{Perform data processing calculations}
\usage{
dt_calc(
  sfc = NULL,
  station_file = NULL,
  f_params = NULL,
  station_file_ext = ".ipip",
  ppsij = NULL,
  ...
)
}
\arguments{
\item{station_file}{Name of the station being processed.}

\item{f_params}{Function parameters evaluated by \code{ipayipi::calc_param_eval()}. These are parsed to \code{dt_calc()} from \code{dt_process()}.}

\item{dta_in}{A list of data to be processed.
param input_dt
param output_dt}

\item{f_summary}{A summary table of function parameters. This
summary table is stored in the station file object. If there
is a summary table for the function, differences between the
\code{f_params} and table are checked and the 'f_summary' table
is updated---triggering a re-calculation of the whole time series instead
of for a slice of the series.}
}
\value{
A list containing the processed data sets 'dts_dt'.
}
\description{
Calculations are parsed to data.table.
}
\author{
Paul J. Gordijn
details
}
