% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_h1_station_dt_clean.R
\name{dt_clean}
\alias{dt_clean}
\title{Dev function: Apply filter windows to time series data}
\usage{
dt_clean(
  phens = NULL,
  w = 21,
  madf = 3,
  align = "left",
  seg = NULL,
  cush = TRUE,
  clean_f = "hampel",
  owrite = TRUE,
  na_t = 0.75,
  tighten = 0.65,
  station_file = NULL,
  station_file_ext = ".ipip",
  ppsij = NULL,
  f_params = NULL,
  sfc = NULL,
  verbose = FALSE,
  xtra_v = FALSE,
  chunk_v = FALSE,
  ...
)
}
\arguments{
\item{phens}{Vector of the phenomena names that will be evaluated by the hampel filter. If NULL the function will not run.}

\item{w}{Window size for the hempel filter. Defaults to ten.}

\item{madf}{Scalar factor of MAD (median absolute deviation). Higher values relax oulier detection. Defaults to the standard of three.}

\item{seg}{Name of column to segment filtering on. Cleaning within each data segment or slice will be run independently.}

\item{clean_f}{Algorithm name. Only "hampel" supported.}

\item{na_t}{Not yet implemented. Fractional tolerace of the amount of NA values in a segment for linear interpolation of missing values.}

\item{tighten}{Not yet implemented. Scaling fraction between zero and one that sensitizes the detection of outliers near the head and tail ends of segments. The fraction is multiplied by the mad factor.}

\item{station_file}{Name of the station being processed.}

\item{ppsij}{Data processing \code{pipe_seq} table from which function parameters and data are extracted/evaluated. This is parsed to this function automatically by \code{ipayipi::dt_process()}.}

\item{sfc}{List of file paths to the temporary station file directory. Generated using \code{ipayipi::open_sf_con()}.}

\item{verbose}{Logical. Whether or not to report messages and progress.}

\item{xtra_v}{Logical. Whether or not to report xtra messages, progess, plus print data tables.}

\item{chunk_v}{Logical. Print data chunking messages. Useful for debugging/digging into chunking methods.}

\item{seg_fuzz}{Not yet implemented. String representing the threshold time interval between the list of segment date-time}
}
\description{
Function under development. Uses the non-parametric hampel filter to detect anomalies and impute values in a univariate series of numeric data.
}
\details{
This function employs a user-customised hampel filter to check data for outliers and impute missing values with the median of the filter window. Function under development.
\itemize{
\item Need to add generated table details to the phen_ds summary table!
}
}
\author{
Paul J. Gordijn
}
\keyword{data,}
\keyword{detection,}
\keyword{imputation,}
\keyword{outlier}
\keyword{univariate}
\keyword{value}
