% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_h1_station_dt_clean.r
\name{dt_clean}
\alias{dt_clean}
\title{Flag outliers and impute missing values.}
\usage{
dt_clean(
  sfc = NULL,
  station_file = NULL,
  clean_f = "hampel",
  phen_names = NULL,
  w_size = 21,
  mad_dev = 3,
  segs = NULL,
  seg_fuzz = NULL,
  seg_na_t = 0.75,
  last_rule = FALSE,
  tighten = 0.65,
  ppsij = NULL,
  f_params = NULL,
  verbose = FALSE,
  xtra_v = FALSE,
  ...
)
}
\arguments{
\item{sfc}{List of file paths to the temporary station file directory. Generated using \code{ipayipi::open_sf_con()}.}

\item{station_file}{Name of the station being processed.}

\item{clean_f}{Algorithm name. Only "hampel" supported.}

\item{phen_names}{Vector of the phenomena names that will be evaluated by the hampel filter. If NULL the function will not run.}

\item{w_size}{Window size for the hempel filter. Defaults to ten.}

\item{mad_dev}{Scalar factor of MAD (median absolute deviation). Higher values relax oulier detection. Defaults to the standard of three.}

\item{segs}{Vector of station data table names which contain timestamps used to slice data series into segments with independent outlier detection runs. If left \code{NULL} (default) the run will be from the start to end of data being evaluated.}

\item{seg_fuzz}{String representing the threshold time interval between the list of segment date-time}

\item{seg_na_t}{Fractional tolerace of the amount of NA values in a segment for linear interpolation of missing values.}

\item{last_rule}{Whether or not to reapply stored rules in the outlier rule column. TRUE will apply old rules.}

\item{tighten}{Scaling fraction between zero and one that sensitizes the detection of outliers near the head and tail ends of segments. The fraction is multiplied by the mad_dev factor.}

\item{ppsij}{Data processing \code{pipe_seq} table from which function parameters and data are extracted/evaluated. This is parsed to this function automatically by \code{ipayipi::dt_process()}.}

\item{verbose}{Logical. Whether or not to report messages and progress.}

\item{xtra_v}{Logical. Whether or not to report xtra messages, progess, plus print data tables.}

\item{cores}{Number of CPU's to use for processing in parallel. Only applies when working on Linux.}
}
\value{
A vector of xle file paths.
}
\description{
Uses the non-parametric hampel filter to detect anomalies and impute values in a univariate series of numeric data.
}
\author{
Paul J. Gordijn
}
\keyword{data,}
\keyword{detection,}
\keyword{interpolation,}
\keyword{outlier}
\keyword{univariate}
\keyword{value}
