% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/h1_dt_drift.R
\name{dt_drift}
\alias{dt_drift}
\title{Dev function: Linear drift corrections based on calibration readings.}
\usage{
dt_drift(
  phen = NULL,
  wside = 3,
  wside_max = 12,
  robust = TRUE,
  edge_adj = NULL,
  pipe_house = NULL,
  station_file = NULL,
  ppsij = NULL,
  f_params = NULL,
  sfc = NULL,
  verbose = FALSE,
  xtra_v = FALSE,
  chunk_v = FALSE,
  ...
)
}
\arguments{
\item{pipe_house}{List of pipeline directories. \emph{See} \code{\link[=ipip_house]{ipip_house()}} \emph{for details}.}

\item{station_file}{Name of the station being processed.}

\item{ppsij}{Data processing \code{pipe_seq} table from which function parameters and data are extracted/evaluated. This is parsed to this function automatically by \code{ipayipi::dt_process()}.}

\item{sfc}{List of file paths to the temporary station file directory. Generated using \code{ipayipi::sf_open_con()}.}

\item{verbose}{Logical. Whether or not to report messages and progress.}

\item{xtra_v}{Logical. Whether or not to report xtra messages, progess, plus print data tables.}

\item{chunk_v}{Logical. Print data chunking messages. Useful for debugging/digging into chunking methods.}
}
\description{
Function under development. Uses the medians of defined window sizes, around calibration points to perform linear dirift correction for a variable .
}
\details{
This function employs a user-customised hampel filter to check data for outliers and impute missing values with the median of the filter window. Function under development.
\itemize{
\item Need to add generated table details to the phen_ds summary table!
}
}
\author{
Paul J. Gordijn
}
\keyword{data,}
\keyword{detection,}
\keyword{imputation,}
\keyword{outlier}
\keyword{univariate}
\keyword{value}
