% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_f1_station_dt_harvest.r
\name{dt_harvest}
\alias{dt_harvest}
\title{Harvest data from another station or table.}
\usage{
dt_harvest(
  station_file = station_file,
  output_dt = NULL,
  ext_dir = NULL,
  ext_station = NULL,
  ext_station_table = NULL,
  output_dt_preffix = "dt_",
  output_dt_suffix = NULL,
  f_params = NULL,
  f_summary = NULL,
  time_interval = NULL,
  recurr = TRUE,
  ...
)
}
\arguments{
\item{station_file}{Name of the station being processed.}

\item{output_dt}{The output table name where harvested data is extracted too.}

\item{ext_dir}{The path of the directory in which to search for the
external data and/or data table.}

\item{ext_station}{The name (or keyword) of the station from which to extract
data.}

\item{f_params}{A vector of phenomena name to be extracted from the
harvested data table. If NULL all column names are extracted.}

\item{time_interval}{The desirsed time_tinterval associated with the
extracted data. If this is \code{NULL} then the table is extracted as is.}
}
\description{
Used to extract data from another table, station, or source
and ...
}
\author{
Paul J. Gordijn
}
