% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_h1_station_dt_harvest.r
\name{dt_harvest}
\alias{dt_harvest}
\title{'dt' processing pipeline: harvest data}
\usage{
dt_harvest(
  station_file = NULL,
  f_params = NULL,
  ppsij = NULL,
  sfc = NULL,
  verbose = FALSE,
  xtra_v = FALSE,
  ...
)
}
\arguments{
\item{station_file}{File path of the station being processed.}

\item{f_params}{Function parameters parsed by \code{ipayipi::dt_process()}.}

\item{ppsij}{The desirsed time_interval associated with the extracted data. If this is \code{NULL} then the table is extracted as is.}

\item{sfc}{Station file connection; see \code{ipayipi::open_sf_con2()}.}

\item{hsf_table}{The path of the directory in which to search for the external data and/or data table.}
}
\description{
Used to extract data from other table/s, stations, or source/s.
}
\author{
Paul J. Gordijn
details
}
