% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_h1_station_dt_harvest.R
\name{dt_harvest}
\alias{dt_harvest}
\title{'dt' processing pipeline: harvest data}
\usage{
dt_harvest(
  station_file = NULL,
  f_params = NULL,
  ppsij = NULL,
  sfc = NULL,
  verbose = FALSE,
  xtra_v = FALSE,
  ...
)
}
\arguments{
\item{station_file}{File path of the station being processed.}

\item{f_params}{Function parameters parsed by \code{dt_process()}.}

\item{ppsij}{The desirsed time_interval associated with the extracted data. If this is \code{NULL} then the table is extracted as is.}

\item{sfc}{Station file connection; see \code{open_sf_con()}.}

\item{verbose}{Print some basic information on progress? Logical.}

\item{xtra_v}{Print some more details on function processing. Logical.}

\item{hsf_table}{The path of the directory in which to search for the external data and/or data table.}
}
\description{
Used to extract data from other table/s, stations, or source/s.
}
\details{
Harvesting data during processing will function differently depending on type of data. If chunked data is being harvested, essentially only the link to the chunked data and associated chunk index is copied. If unchunked data (that is data without a 'date_time' column processed by 'ipayipi') then the table is copied into the stations station file connection directory as a 'data.table'.
Importantly, if the data has an associated 'gap' table; this table (filtered by the harvest table name and phenomena names --- logger gaps always inluded) is harvested and placed with the copied chunked data link. This 'gap' data is transferred down the pipeline as data is processed by \code{ipayipi::dt_process()}.
}
\author{
Paul J. Gordijn
}
