% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_h0_station_dt_process_batch.R
\name{dt_process_batch}
\alias{dt_process_batch}
\title{Batch processing of pipeline data: \emph{dt}}
\usage{
dt_process_batch(
  pipe_house = NULL,
  pipe_seq = NULL,
  stages = NULL,
  overwrite_pipe_memory = FALSE,
  output_dt_preffix = "dt_",
  output_dt_suffix = NULL,
  station_ext = ".ipip",
  wanted = NULL,
  unwanted = NULL,
  prompt = FALSE,
  unwanted_tbls = "_tmp",
  verbose = FALSE,
  xtra_v = FALSE,
  keep_open = TRUE,
  ...
)
}
\arguments{
\item{pipe_house}{List of pipeline directories. \strong{See} \code{ipayipi::ipip_house()} for details__.}

\item{pipe_seq}{Processing pipeline structure in tablular format as generated by \code{ipayipi::pipe_seq()}. Defaults to \code{NULL}.
If this argument is not provided then the \code{pipe_seq} stored in the station file will be used for data processing. When changes are made to a stations processing pipeline \code{overwrite_pipe_memory} should be set to \code{TRUE}, in which case a \code{pipe_seq} object must be provided here.}

\item{stages}{Integer vector denoting the consecutive stages of the \code{pipe_seq} object to process. Can be used to split a processing pipeline for addition processing between stages.}

\item{overwrite_pipe_memory}{Logical. If \code{TRUE} then the stations pipeline steps, that are described by the \code{pipe_seq} object/table, are overwritten by the new \code{pipe_seq} object.}

\item{output_dt_preffix}{The output table preffix which defaults to "dt_".}

\item{output_dt_suffix}{A custom suffix to be appended to the output tables name.}

\item{station_ext}{The extension of the station file. Defaults to 'ipip'.}

\item{wanted}{Vector of strings listing files that should not be included in station file search.}

\item{unwanted}{Vector of strings listing station files that should not be included in the import.}

\item{prompt}{Should the function use an interactive station file selection function otherwise all files are returned. \code{TRUE} or \code{FALSE}.}

\item{unwanted_tbls}{Some tables generated by the processing pipeline don't need to be stored permanently in the station file object. By adding keywords to this argument (separated by the '|' character) these 'unwanted' (or temporary) tables will be removed from the station file. Defaults to \verb{_tmp} --- so any table with this search key in its name will be removed.}

\item{verbose}{Print some details on the files being processed? Logical.}

\item{xtra_v}{Extra verbose? Logical.}

\item{keep_open}{Logical. If TRUE the station file connection is kept in the temporary directory after the function is finished processing.}

\item{recurr}{Should the function search recursively into sub directories for station files? \code{TRUE} or \code{FALSE}.}

\item{cores}{Number of CPU's to use for processing in parallel. Only applies when working on Linux systems.}
}
\description{
Batch process ipayipi station data data in sequential steps.
}
\details{
This batch processing function calls \code{ipayipi::dt_process()} --- for more details read this functions documentation.
}
\author{
Paul J. Gordijn
}
\keyword{aggregation;}
\keyword{calculations;}
\keyword{data}
\keyword{data;}
\keyword{event}
\keyword{join}
\keyword{metadata}
\keyword{processing;}
\keyword{series}
\keyword{time}
\keyword{transformations;}
