% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_sub_extend_long.r
\name{dttm_extend_long}
\alias{dttm_extend_long}
\title{Generate date-time data for sequence gaps}
\usage{
dttm_extend_long(data_sets = NULL, ri = NULL, ...)
}
\arguments{
\item{data_sets}{List of time-series data sets with a 'date_time' column.}

\item{ri}{Standardised record interval. Function will not process event-based/'discontinuous' ('discnt') data; requires a regular time-interval.}
}
\description{
Internal function to generate time-series data tables, with 'date_time' sequences only, to fill missing date-time values in a list of non-overlapping data sets.
}
\author{
Paul J. Gordijn
}
