% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_sub_extend_long.R
\name{dttm_extend_long}
\alias{dttm_extend_long}
\title{Generate date-time data for sequence gaps}
\usage{
dttm_extend_long(data_sets = NULL, intra_check = FALSE, ri = NULL, ...)
}
\arguments{
\item{data_sets}{List of time-series data sets with a 'date_time' column.}

\item{intra_check}{Logical. If \code{TRUE} the function will join each dataset to a continuous date_time sequence. If \code{FALSE}, only the gaps between listed datasets will be made continuous using the standardised record interval \code{ri}.}

\item{ri}{Standardised record interval. Function will not process event-based/'discontinuous' ('discnt') data; requires a regular time-interval character string standardised by \code{ipayipi::record_interval_eval()}.}
}
\description{
Function to generate time-series data tables with 'date_time' sequences in a list of non-overlapping data sets.
}
\author{
Paul J. Gordijn
}
