% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_f0_gap_eval_batch.R
\name{gap_eval_batch}
\alias{gap_eval_batch}
\title{Highlight 'gaps' (missing data) in station data}
\usage{
gap_eval_batch(
  pipe_house = NULL,
  gap_problem_thresh_s = 6 * 60 * 60,
  event_thresh_s = 10 * 60,
  phens = NULL,
  phen_eval = FALSE,
  meta_events = "meta_events",
  station_ext = ".ipip",
  prompt = FALSE,
  wanted = NULL,
  unwanted = NULL,
  verbose = FALSE,
  keep_open = TRUE,
  xtra_v = FALSE,
  ...
)
}
\arguments{
\item{pipe_house}{Required. List of pipeline directories. \strong{See} \code{ipayipi::ipip_house()} \strong{for details}.}

\item{gap_problem_thresh_s}{A duration threshold (in seconds) beyond which a gap, in event-based time series records, is considered problematic --- a 'true' gap. These gaps require infilling or 'patching'. The default for event-based data is six hours (i.e., 6 * 60 * 60 seconds). If the 'record_interval_type' is not 'mixed' or 'event_based', i.e. it is only continuous, then the 'record_interval' parameter is used as the \code{gap_problem_thresh_s}.}

\item{event_thresh_s}{A gap can also be specified in the event metadata by providing only a single date-time stamp and an 'event_threshold_s' (in seconds). By taking this date-time stamp the function approximates the start and end-time of the data gap by adding and subtracting the given \code{event_thresh_s} to the date-time stamp, respectively. If the \code{event_thresh_s} is supplied in the 'meta_events' data (\emph{see} \code{meta_to_station()}) then the value supplied therein is used in preference to the \code{event_thresh_s} argument in this function. The default event threshold here is ten minutes, i.e., 10 * 60 seconds.}

\item{meta_events}{Optional. The name of the data.table in the station file with event metadata. Defaults to NA.}

\item{prompt}{Should the function use an interactive station file selection function otherwise all files are returned. \code{TRUE} or \code{FALSE}.}

\item{wanted}{Vector of strings listing files that should not be included in station file search.}

\item{unwanted}{Vector of strings listing station files that should not be included in the import.}

\item{keep_open}{Logical. Keep \emph{hidden} 'station_file' open for ease of access. Defaults to \code{TRUE}.}

\item{station_file}{Standardised \strong{ipayipi} station file.}

\item{recurr}{Should the function search recursively into sub directories for station files? \code{TRUE} or \code{FALSE}.}
}
\value{
Names of stations with Updated gap period tables.
}
\description{
The period between the time a logger is stopped, removed, or discommissioned, and when a logger beings recording data again, counts as missing data. However, for event-based data, where the frequency of recordings is temporarily erratic, identifying gaps is more tricky. This function sets some rules when identifying gaps in event-based time-series data.
}
\details{
Batch application of \code{ipayipi::gap_eval()}.
}
\author{
Paul J. Gordijn
}
\keyword{data}
\keyword{data;}
\keyword{event}
\keyword{gaps;}
\keyword{missing}
\keyword{pipeline;}
