% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gw_f3_alpine_butter.r
\name{gw_alpine_butter}
\alias{gw_alpine_butter}
\title{Generate dummy dipper readings}
\usage{
gw_alpine_butter(
  input_dir = ".",
  wanted = NULL,
  unwanted = NULL,
  prompt = FALSE,
  recurr = FALSE,
  dt_format = c("Ymd HMS", "Ymd IMSp", "ymd HMS", "ymd IMSp", "mdY HMS", "mdy IMSp",
    "dmY HMS", "dmy IMSp"),
  dt_tz = "Africa/Johannesburg",
  ...
)
}
\arguments{
\item{input_dir}{The working directory where the standardised 'ipayipi'
water level files are located.}

\item{wanted}{Vector of strings listing files that should not be
included in the import.}

\item{unwanted}{Vector of strings listing files that should not be included
in the import.}

\item{prompt}{If TRUE, a command line prompt will be used to
enable selection of which files in the working directory require
drift correction.}

\item{recurr}{Should the function work recurrsively in the parent
directory.}

\item{dt_format}{The function guesses the date-time format from a vector of
format types supplied to this argument. The 'guessing' is done via
\code{lubridate::parse_date_time()}. \code{lubridate::parse_date_time()} prioritizes
the 'guessing' of date-time formats in the order vector of formats
supplied. The default vector of date-time formats supplied should work
well for most logger outputs.}

\item{dt_tz}{Recognized time-zone (character string) of the data locale. The
default for the package is South African, i.e., "Africa/Johannesburg" which
is equivalent to "SAST".}
}
\value{
Saves station files with table of dummy-drift readings.
}
\description{
Generates dummy dipper readings using linear interpolation and
adds these to the 'log_retreive' table. These will be used for in subsequent
drift correction.
}
\author{
Paul J. Gordijn
}
\keyword{correction}
\keyword{correction,}
\keyword{data}
\keyword{drift}
\keyword{dummy}
\keyword{values,}
