% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gw_e_sub_read_retrieve_log.R
\name{gw_read_retrieve_log}
\alias{gw_read_retrieve_log}
\title{Function to read in the events table}
\usage{
gw_read_retrieve_log(
  input_dir = NULL,
  recurr = FALSE,
  dt_format = c("Ymd HMS", "Ymd IMSp", "ymd HMS", "ymd IMSp", "mdY HMS", "mdy IMSp",
    "dmY HMS", "dmy IMSp"),
  dt_tz = "Africa/Johannesburg",
  ...
)
}
\arguments{
\item{input_dir}{Directory in which the R solonist data files and retrieve
log csv files are located.}

\item{recurr}{Should sub-directories be included in the search for the datum
log csv file - TRUE/FALSE.}

\item{dt_format}{The function guesses the date-time format from a vector of
format types supplied to this argument. The 'guessing' is done via
\code{lubridate::parse_date_time()}. \code{lubridate::parse_date_time()} prioritizes
the 'guessing' of date-time formats in the order vector of formats
supplied. The default vector of date-time formats supplied should work
well for most logger outputs.}

\item{dt_tz}{Recognized time-zone (character string) of the data locale. The
default for the package is South African, i.e., "Africa/Johannesburg" which
is equivalent to "SAST".}
}
\description{
This function reads the dipper and casing height readings
(retrieve log) which are used for drift correction of water levels.
This data is appended to the \strong{R} groundwater data files.
}
\author{
Paul J. Gordijn
}
\keyword{dipper}
\keyword{readings}
