% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_c3_header_sts.r
\name{header_sts}
\alias{header_sts}
\title{Standardises ipayipi raw data summary information}
\usage{
header_sts(
  wait_room = NULL,
  prompt = FALSE,
  recurr = TRUE,
  wanted = NULL,
  unwanted = NULL,
  file_ext_in = ".ipr",
  file_ext_out = ".iph",
  ...
)
}
\arguments{
\item{wait_room}{Path to the 'wait_room' directory.}

\item{prompt}{Should the function use an interactive file selection function
otherwise all files are returned. TRUE or FALSE. Defaults to FALSE.}

\item{recurr}{Should the function search recursively into sub directories
for hobo rainfall csv export files? TRUE or FALSE.}

\item{wanted}{A strong containing keywords to use to filter which stations
are selected for processing. Multiple search kewords should be seperated
with a bar ('|'), and spaces avoided unless part of the keyword.}

\item{unwanted}{Similar to wanted, but keywords for filtering out unwanted
stations.}
}
\description{
Uses the nomenclature table, standardises the recording
interval string, start and end date times in the file header, and finally
the file name is copied from the standardised station name, that is, the
'standardised title'.
}
\details{
This function calls \code{ipayipi::met_nomenclature()} which will take
take the user interactively through a process of standardising station
names and titles.
The phenomena of files can only be checked once nomenclature has been
standardised.
If the fundtion requests that a check of the record interval
make sure the in function record interval table is updated.
}
\author{
Paul J. Gordijn
}
\keyword{Cambell}
\keyword{Scientific;}
\keyword{automatic}
\keyword{batch}
\keyword{data;}
\keyword{file}
\keyword{header}
\keyword{information}
\keyword{meteorological}
\keyword{nomenclature;}
\keyword{process;}
\keyword{standardisation;}
\keyword{station;}
\keyword{weather}
