% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hobo_rain}
\alias{hobo_rain}
\title{Header harvest: Exported hobo rainfall files}
\format{
\subsection{'hobo_rain'}{

iPayipi data_setup list for imibing (\code{ipayipi::imbibe_raw_logger_dt()})
Hobo logger data from a Hobo rainfall logger pendant. The setup includes
harvseting of the loggers serial number using the search pattern
'LGN.S.N..'. The 'location' name is not specified. Only generic
information is provided for harvesting of logger file header information.
__See \code{ipayipi::imbibe_raw_logger_dt()} documentation for a describtion of
a \code{data_setup}.
This data_setup contains two varieties of hobo pendant logger outputs, the
first option contains an id column, whilst the second does not.
}
}
\source{
Internal
}
\usage{
hobo_rain
}
\description{
ipayipi data setup for Hobo logger files exported from a 'rainfall pendant'
}
\keyword{datasets}
