% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_b_imbibe_raw_logger_dt_batch.r
\name{imbibe_raw_batch}
\alias{imbibe_raw_batch}
\title{Batch imbibe of 'flat' data files into the 'ipayipi' format}
\usage{
imbibe_raw_batch(
  pipe_house = NULL,
  file_ext_in = NULL,
  file_ext_out = ".ipr",
  col_dlm = ",",
  dt_format = c("Ymd HMS", "Ymd IMSp", "ymd HMS", "ymd IMSp", "mdY HMS", "mdy IMSp",
    "dmY HMS", "dmy IMSp", "Ymd HMOS", "Ymd IMOSp", "ymd HMOS", "ymd IMOSp", "mdY HMOS",
    "mdy IMOSp", "dmY HMOS", "dmy IMOSp"),
  dt_tz = "Africa/Johannesburg",
  record_interval_type = "continuous",
  data_setup = NULL,
  prompt = FALSE,
  wanted = NULL,
  unwanted = NULL,
  recurr = FALSE,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{pipe_house}{List of pipeline directories. \strong{See}
\code{ipayipi::ipip_init()} \strong{for details}.}

\item{file_ext_in}{The file extension defaults of the raw logger data files.}

\item{file_ext_out}{The file extension used when raw logger data which has
been imbibed into the \code{ipayipi} data pipeline.}

\item{col_dlm}{The column delimter which is fed to \code{data.table::fread()}.
Defaults to a comma for 'csv' files.}

\item{dt_format}{The input date-time format of the time series, e.g.,
"\%y-\%m-\%d \%H:\%M:\%S". See ?base::strptime() for details.}

\item{dt_tz}{recognized time-zone of the data locale.}

\item{data_setup}{List of options used to extract data and metadata from
instrument data outputs. For a description of the \code{data_setup} see
\code{ipayipi::read_logger_dt()}.}

\item{prompt}{Should the function use an interactive file selection function
otherwise all files are returned. TRUE or FALSE.}

\item{wanted}{A strong containing keywords to use to filter which stations
are selected for processing. Multiple search kewords should be seperated
with a bar ('|'), and spaces avoided unless part of the keyword.}

\item{unwanted}{Similar to wanted, but keywords for filtering out unwanted
stations.}

\item{recurr}{Should the function search recursively into sub directories
for hobo rainfall csv export files? TRUE or FALSE.}

\item{record_interval}{If there are is no discrete record interval set
in the logger program, i.e., the sampling is event based, then this
parameter must be set to "event_based". Defaults to "continuous".}

\item{silent}{Logical passed to \code{attempt::attempt()} which reads the logger
text file in with either \code{data.table::fread()} or base R.}
}
\description{
Uses \code{ipayipi::imbibe_logger_dt()} to read and transfer
data files to the begining stages of the \code{ipayipi} data pipeline. Option
to archive all 'raw' data files in the pipeline structure, \emph{see details}.
}
\details{
In the pipeline structure this function should be used post
\code{ipayipi::logger_data_import_batch()}. \code{ipayipi::imbibe_raw_batch()} is
a wrapper for \code{ipayipi::imbibe_raw_logger_dt()} --- see this functions
documentation for more details.

\strong{'Archiving' raw data}
Files brought into the 'wait_room' are only housed there temporarily.
In order to archive these raw data files a 'raw_room' directory must be
provided in the 'pipe_house' object (\emph{see} \code{ipayipi::ipip_init()}).
Files will be archived in structured directories in the 'raw_room' named
by the last year and month in their respective date time records. Original
file names are maintained, and have a suffix with a unique integer plus
the date and time of which they were archived.
}
\author{
Paul J. Gordijn
}
\keyword{automatic}
\keyword{batch}
\keyword{data}
\keyword{data;}
\keyword{meteorological}
\keyword{pipeline}
\keyword{process;}
\keyword{raw}
\keyword{standardisation;}
\keyword{station;}
\keyword{weather}
