% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_b_imbibe_logger_dt_batch.r
\name{imbibe_raw_batch}
\alias{imbibe_raw_batch}
\title{Batch imbibe of 'flat' data files into the 'ipayipi' format}
\usage{
imbibe_raw_batch(
  wait_room = NULL,
  file_ext_in = NULL,
  file_ext_out = ".ipr",
  col_dlm = ",",
  dt_format = c("Ymd HMS", "Ymd IMSp", "ymd HMS", "ymd IMSp", "mdY HMS", "mdy IMSp",
    "dmY HMS", "dmy IMSp"),
  dt_tz = "Africa/Johannesburg",
  record_interval_type = "continuous",
  data_setup = NULL,
  prompt = FALSE,
  wanted = NULL,
  unwanted = NULL,
  recurr = FALSE,
  ...
)
}
\arguments{
\item{wait_room}{Path to the 'wait_room' directory.}

\item{file_ext_in}{The file extension defaults of the raw logger data files.}

\item{file_ext_out}{The file extension used when raw logger data which has
been imbibed into the \code{ipayipi} data pipeline.}

\item{col_dlm}{The column delimter which is fed to \code{data.table::fread()}.
Defaults to a comma for 'csv' files.}

\item{dt_format}{The input date-time format of the time series, e.g.,
"\%y-\%m-\%d \%H:\%M:\%S". See ?base::strptime() for details.}

\item{dt_tz}{recognized time-zone of the data locale.}

\item{data_setup}{List of options used to extract data and metadata from
instrument data outputs. For a description of the \code{data_setup} see
\code{ipayipi::read_logger_dt()}.}

\item{prompt}{Should the function use an interactive file selection function
otherwise all files are returned. TRUE or FALSE.}

\item{wanted}{Vector of strings listing files that should not be
included in the import.}

\item{unwanted}{Vector of strings listing files that should not be included
in the import.}

\item{recurr}{Should the function search recursively into sub directories
for hobo rainfall csv export files? TRUE or FALSE.}

\item{record_interval}{If there are is no discrete record interval set
in the logger program, i.e., the sampling is event based, then this
parameter must be set to "event_based". Defaults to "continuous".}
}
\description{
Uses \code{ipayipi::imbibe_logger_dt()} to read and transfer
data files to the begining stages of the \code{ipayipi} data pipeline.
}
\author{
Paul J. Gordijn
}
\keyword{automatic}
\keyword{batch}
\keyword{data}
\keyword{data;}
\keyword{meteorological}
\keyword{pipeline}
\keyword{process;}
\keyword{raw}
\keyword{standardisation;}
\keyword{station;}
\keyword{weather}
