% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_a_ipip_house.R
\name{ipip_house}
\alias{ipip_house}
\title{Build \code{ipayipi} data pipeline housing for a data family}
\usage{
ipip_house(
  pipe_house_dir = ".",
  r = NULL,
  source_room = NULL,
  wait_room = NULL,
  nomvet_room = NULL,
  ipip_room = NULL,
  dta_out = NULL,
  reports = NULL,
  raw_room = NULL,
  work_dir = NULL,
  verbose = FALSE,
  ...
)
}
\arguments{
\item{pipe_house_dir}{The main pipeline directory/'room' (or folder) in which all pipeline rooms are 'housed'. \bold{NB!} The \code{pipe_house_dir} direcotry must be relative to the active console/terminal working directory, or the full path name.
\cr \cr}

\item{r}{Directory for miscellaneous scripts. Within the 'r' room, a sub directory is created called 'pipe_seq' for housing processing pipeline sequence scripts for \code{ipayipi::dt_process()}. Suggestion is to leave \code{NULL} so that this 'room' is nested within the \code{pipe_house_dir}.
The data is moved through an 'ipayipi' pipeline housing system through the following rooms.
\cr \cr
Data moves through an 'ipayipi' pipeline through the following rooms 1--4. Along the way raw data is organised and archived, and data is standardised before being processed. \cr}

\item{source_room}{\strong{1}. 'Room' from which raw-data files are imported. If left \code{NULL} a 'source_room' will be created in the pipe_house working directory.}

\item{wait_room}{\strong{2}. Directory into which raw-data moves to from the \code{source_room}. This is also the room where standardisation of raw-data files begins.}

\item{nomvet_room}{\strong{3}. 'Room' into which imbibed, standardised raw-data is stored.}

\item{ipip_room}{\strong{4}. 'Room' where 'station files' are housed. Station files consist of appended, standardised raw-data files---pulled from the \code{nomvet_room}---that have been, or are ready for further processing.}

\item{dta_out}{Used for housing select pipeline data products typically in 'csv' file format.}

\item{reports}{Typically used to house 'rmarkdown' reports used for inspection and analysis of data.}

\item{raw_room}{'Room' wherein raw-data files from the \code{source_room} are systematically archived. If this is set to \code{NULL} then raw-data files will not be archived by 'ipayipi'.}

\item{verbose}{Logical. If \code{TRUE} some terminal/console messages may provide useful insight.}
}
\value{
List of pipeline housing 'rooms' (filepaths).
}
\description{
Creates a list of 'rooms', that is, folders/directories, required for an 'ipayipi' data-processing pipeline within an \strong{existing} \code{pipe_house_dir} folder. Suggestion is to use default options: only provide the \code{pipe_house_dir} that refers to an existing directory for your data pipeline. By providing other parameters the directories can be customised for special-use cases.
The recommentation is to build seperate 'pipe_houses' for different data streams or families, for example, separate out automatic weather stations, stand alone rain gauges, and water-level transducers. Whilst the families are housed in different 'pipe houses', iPayipi enables efficient quering and processing of data across different data families.
}
\details{
This function automates the creation of four/five folders/directories that are requried for bulk processing of files in the ipayipi pipeline structure. The flow of data through an 'ipayipi' pipeline housing system is described below. The preffix to the folders created represents this flow of data.
\enumerate{
\item \emph{pipe_house_dir}*: The main directory within which other directories are housed for preparing, standardising and processing data.
\item \emph{source_room}: The source directory where raw data are harvested from.
\item \emph{wait_room}: A staging directory where 'raw data' are standardised before being transferred/archived in the,
\item \emph{nomvet_room}: Folder that standardised input data for multiple stations. By appending these standarised files station records (databases) are compiled.
\item \emph{ipip_room}: The directory that compiles standardised data from the \code{nomvet_room} by stations. Station files can be further processed and exported into other formats from here.
\item \emph{raw_room}:  If defined \code{ipayipi::imbibe_raw_batch()} can harvest (copy or cut---see funtion documentation) and archive 'raw data' files from the \code{source_room} and will archive them in the \code{raw_room} in monthly folders.
\cr
{*} In older versions of 'ipayipi' the 'pipe_house_dir' argument was named 'work_dir'; this alteration prevents confusion with the R terminal/console working directory.
}
}
\examples{
# Inititate pipeline
pd <- "." # define the working directory
pipe_house <- ipip_house(pipe_house_dir = pd)
print(pipe_house)
}
\author{
Paul J. Gordijn
}
\keyword{creation,}
\keyword{directory}
\keyword{folder}
\keyword{initiate}
\keyword{pipeline,}
\keyword{structure}
