% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_a_ipip_house.r
\name{ipip_house}
\alias{ipip_house}
\title{Build \code{ipayipi} data pipeline housing}
\usage{
ipip_house(
  work_dir = ".",
  r = NULL,
  source_room = NULL,
  wait_room = NULL,
  nomvet_room = NULL,
  ipip_room = NULL,
  dta_out = NULL,
  reports = NULL,
  raw_room = NULL,
  sf_con = TRUE,
  ...
)
}
\arguments{
\item{work_dir}{The working directory/'room' (or folder) in which all pipeline rooms are 'housed'.

The data is moved through an 'ipayipi' pipeline housing system through the following rooms.}

\item{source_room}{'Room' from which raw-data files are imported. If left \code{NULL} a 'source_room' will be created in the pipe_house working directory.}

\item{wait_room}{Directory into which raw-data moves to from the \code{source_room}. This is also the room where standardisation of raw-data files begins.}

\item{nomvet_room}{'Room' into which imbibed, standardised raw-data is stored.}

\item{ipip_room}{'Room' where 'station files' are housed. Station files consist of appended, standardised raw-data files---pulled from the \code{nomvet_room}---that have been, or are ready for further processing.}

\item{raw_room}{'Room' wherein raw-data files from the \code{source_room} are systematically archived. If this is set to \code{NULL} then raw-data files will not be archived by 'ipayipi'.}

\item{full}{If \code{TRUE} additonal folders in the working directory will be generated for 1)}
}
\value{
List of pipeline housing 'rooms' (filepaths).
}
\description{
Compiles/creates a list of 'rooms' (folders/directories) required for an 'ipayipi' data-processing pipeline.
}
\details{
This function automates the creation of four/five folders/directories that are requried for bulk processing of files in the ipayipi pipeline structure. The flow of data through an 'ipayipi' pipeline housing system is illustrated below.

1-|--work_dir -------------------------------------------------------------|
2-|--source_room-->|                                                       |
3-|                |--wait_room-->|                                        |
4-|                |              |--nomvet_room-->|                       |
5-|                |                               |--ipip_room            |
|                |                                                       |
6-|                |--> raw_room                                           |

\emph{Note the above illustration shows data flow not the structure of an ipayipi directory.}
\enumerate{
\item \emph{work_dir}: The working directory within which other pipeline directories are housed.
\item The source directory where raw data are harvested from.
\item A staging directory where 'raw data' are standardised before being transferred/archived in the,
\item \code{nomvet_room}.
\item The directory that compiles standardised data from the \code{nomvet_room} by stations. Station files can be further processed and exported into other formats from here.
\item If the \code{raw_room} is defined \code{ipayipi::imbibe_raw_batch()} can harvest (copy or cut---see funtion documentation) and archive 'raw data' files from the \code{source_room} and will archive them in the \code{raw_room} in monthly folders.
}
}
\examples{
# Inititate pipeline
wd <- "." # define the working directory
pipe_house <- ipip_house(work_dir = wd)
print(pipe_house)
}
\author{
Paul J. Gordijn
}
\keyword{creation,}
\keyword{directory}
\keyword{folder}
\keyword{initiate}
\keyword{pipeline,}
\keyword{structure}
