% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/a_logger_data_import_batch.R
\name{logger_data_import_batch}
\alias{logger_data_import_batch}
\title{Import logger data}
\usage{
logger_data_import_batch(
  pipe_house = NULL,
  wanted = NULL,
  unwanted = NULL,
  recurr = TRUE,
  file_ext = NULL,
  verbose = FALSE,
  xtra_v = FALSE,
  prompt = FALSE,
  ...
)
}
\arguments{
\item{pipe_house}{List of pipeline directories. \emph{See} \code{\link[=ipip_house]{ipip_house()}} \emph{for details}.}

\item{wanted}{Regex string of files to select for listing. Seperate search tags by using the bar character '|'.}

\item{unwanted}{Regex string of files to filter out the listing. Seperate search tags by using the bar character '|'.
in the import.}

\item{recurr}{Should the function search recursively i.e., thorugh sub-folders as well - \code{TRUE}/\code{FALSE.}}

\item{file_ext}{Character string of the file extension of the input data files. E.g., ".csv" for hobo rainfall file exports. This input character string should contain the period as in the previous sentence.}

\item{verbose}{Logical. Print some details and progress of function progress?}

\item{xtra_v}{Logical. Should some 'x'tra messaging be done? Use to help diagnose problems, and for guidance.}

\item{prompt}{Set to TRUE for interactive mode. Note this will not work if embedded in a parallel processing instance.}
}
\description{
Copies and pastes data matching search criteria from \code{d1_source_room} into 'd2_wait_room'.
}
\details{
\code{logger_data_import_batch()} copies logger data files from multiple stations in 'd1_source_room' into 'd2_wait_room' where data standardisation will take place using: 1) the \code{imbibe_raw_batch()}, 2) \code{header_sts()}, and 3) \code{phenomena_sts()}, in that order. Once standardised in native R format files the data is transferred into the \code{d3_nomvet_room} with \code{transfer_sts_files()}. The standardised files in the 'd3_nomvet_room' are kept there and station records are developed in the 'd4_ipip_room' using \code{append_station_batch()}. Data can be furhter processed with \code{dt_process_batch()}. \cr
}
\section{Notes:}{
\itemize{
\item This function will process data from multiple stations where their raw-flat files are stored in the 'd1_source_room'.
\item Duplicate file names will have unique consequtive integers added as a suffix---so no files are ignored if named similarly.
}
}

\author{
Paul J. Gordijn
}
\keyword{automatic}
\keyword{batch}
\keyword{data}
\keyword{data;}
\keyword{files;}
\keyword{hydrological}
\keyword{import}
\keyword{logger}
\keyword{meteorological}
\keyword{process;}
\keyword{series}
\keyword{station;}
\keyword{time}
\keyword{weather}
