% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sub_dt_process_p_step.R
\name{p_step}
\alias{p_step}
\title{Unpack processing step}
\usage{
p_step(
  dt_n = 1,
  dtp_n = 1,
  f = NULL,
  f_params = NULL,
  input_dt = NA,
  output_dt = NA,
  output_dt_preffix = "dt_",
  output_dt_suffix = NULL,
  time_interval = NA,
  ...
)
}
\arguments{
\item{dt_n}{The numeric (integer) order of the processing stage.}

\item{dtp_n}{A processing stage ('dt_n') contains multiple steps. A stage generally only outputs one processed data table with the specified \code{time_interval} given at the harvesting stage.}

\item{f}{One of the following functions given as a string. Pre-evaluation for each of these functions is done through the linked functions shown in brackets below:
\itemize{
\item "dt_harvest" (\code{\link[=hrv_param_eval]{hrv_param_eval()}}): function used to harvest data from a station or other source.
\item "dt_calc_chain" (\code{\link[=calc_param_eval]{calc_param_eval()}}): uses \code{data.table} chaining to perform calculations, filtering/subsetting, and generate new variables.
\item "dt_agg" (\code{\link[=agg_param_eval]{agg_param_eval()}}): used to aggregate) data by a specified/default function by a desired \code{time_interval}.
\item "dt_join" (\code{\link[=join_param_eval]{join_param_eval()}}): merges two datasets via overlap, left, right, inner and outer joins.
\item *"dt_clean" (\code{\link[=clean_param_eval]{clean_param_eval()}}): runs window filters to detect univariate outliers. Imputes with median if requested (Hampel filter).
\item *"dt_drift" (\code{\link[=drift_param_eval]{drift_param_eval()}}): performs drift corrections based on calibration readings. Under development.
}}

\item{f_params}{Paritially evaluated function parameters generated by respective functions' pre-evaluation criteria (\emph{see below}).}

\item{input_dt}{The name/k)eyword of the input data table (character string).
If not specified at the harvesting stage, "raw" data imported into a station file will be harvested.}

\item{output_dt}{The name of the output data table (character string) --- the final table is appended to a station file object.}

\item{time_interval}{Only has to be specified in the harvesting step. Only one time interval per stage. If the time interval is left as the default \code{NA} then \code{time_interval} becomes discontinuous ("discnt").}
}
\description{
Prepares formatting of a single data processing step.
}
\details{
Note that the \code{f_params} argument is parsed to text in this function. It is only evaluated as text again in \code{\link[=dt_process]{dt_process()}}.
}
\author{
Paul J. Gordijn
}
