% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_plotdy_cleanr.R
\name{plotdy_cleanr}
\alias{plotdy_cleanr}
\title{plot hampel filter run results}
\usage{
plotdy_cleanr(
  station_file = NULL,
  pipe_house = NULL,
  tbl_name = NULL,
  phen_key = NULL,
  mn = NULL,
  mx = NULL,
  slide = 1,
  dflt_nobs = NULL,
  verbose = FALSE,
  xtra_v = FALSE,
  chunk_v = FALSE,
  ...
)
}
\description{
plot hampel filter run results
}
