% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/z_c2_read_nomtab_csv.r
\name{read_nomtab_csv}
\alias{read_nomtab_csv}
\title{Imbibe edited nomenclature csv}
\usage{
read_nomtab_csv(pipe_house = NULL, file = NULL, ...)
}
\arguments{
\item{pipe_house}{List of pipeline directories. \strong{See}
\code{ipayipi::ipip_house()} \strong{for details}.}

\item{file}{name of the csv file to be read as the nomenclature table.}
}
\value{
Saves the read nomenclature table in 'rds' format. Prints the
nomenclature table.
}
\description{
Reads a user edited nomenclature file and saves as a usable
nomtab (nomenclature table) file in the 'waiting room' folder.
}
\details{
\itemize{
\item Searches for the most recently edited nomenclature table csv
('nomtab.csv') file in the \code{wait_room} directory.
\item Checks whether there are any \code{NA} values corresposing to unstandardised
synonyms of the following fields: 'location', 'station', and a station's
standard title, that is, 'stnd_title'.
\item The 'table_name' also requires standardisation. By default this field is
given as concatenation of a preffix, that is, 'raw_', and the standardised
record interval (standardised by \code{ipayipi::record_interval_eval()} during
logger data 'imbibing' (\code{ipayipi::imbibe_raw_logger_dt()})). This
convention should be maintained for the pipelines structure. Note that for
event based or discontinuous data the table name will default to
'raw_discnt'.
}

\strong{NB!} Only edit these fields (above) of the nomenclature table during
standardisation.

\strong{NB!} The 'ipayipi' data pipeline strongly suggests conforming with
tidyverse data standards. Therefore, use lower case characters, no special
characters, except for the underscore character which should be used for
spacing.
}
\author{
Paul J. Gordijn
}
\keyword{Naming}
\keyword{convention,}
\keyword{nomenclature}
\keyword{pipeline}
\keyword{standardisation.}
\keyword{synonyms,}
